% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_year_column.R
\name{add_best_year_column}
\alias{add_best_year_column}
\title{Select the best option for the year column and add it to the data}
\usage{
add_best_year_column(dat, possible_cols, single_year)
}
\arguments{
\item{dat}{dataframe containing the columns that could be used for 'year'.}

\item{possible_cols}{named vector in which the name is the type of year
column and the value is the name of the column. Types of year are
'financial', 'calendar', 'mixed', and 'single'.}

\item{single_year}{list of length 2: 'year' is a character string containing
a valid year; 'warn' is boolean. TRUE if a warning needs to be raised.
single_year is only used if use_single_year is TRUE. IF so, the 'year'
column is populated with this year value. Must be specified if
use_single_year is TRUE. In pub sec this variable is taken either from the
file name (which gets its year from the user config selections) or from the
information above the data. 'warn' is TRUE when the file name year and the
year from above the data both exist but do not match.}
}
\value{
dataframe. dat with year column added/overwritten with the best
year.
}
\description{
From possible_cols select the column name to use as the 'year'
column and add this to the data.
}
\details{
If a column called 'year' already exists in
the data it will be overwritten.

If there is more than one potential year column the order of preference is
financial year, then calendar year, then mixed. Only use single year if
single_overrides_all is TRUE, or if no other possible_cols exist.

Even if single_year is provided, if it is not in possible_cols it will not
be used.
}
\examples{
\dontrun{
dat <- data.frame(
    financial_year = c("2021-22 to 2022-23", "2022_2023"),
    calendar_year = c(2021, 2022)
    )
possible_cols <- c(
    calendar = "calendar_year",
    financial = "financial_year",
    mixed = NA,
    single = NA
    )
add_best_year_column(dat, possible_cols, "2021-22")
}
}
