% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot_data.R
\name{add_columns_to_right_of_data_back_in}
\alias{add_columns_to_right_of_data_back_in}
\title{Join data to the right of the table to the rest of the data by row.}
\usage{
add_columns_to_right_of_data_back_in(
  whole,
  left_and_right,
  col_location,
  first_header_row,
  header_row_count
)
}
\arguments{
\item{whole}{dataframe of all the xlsx_cells data.}

\item{left_and_right}{dataframe of the beheaded data.}

\item{col_location}{integer giving the column number of the outer column.}

\item{first_header_row}{integer giving the row number of the first header.}

\item{header_row_count}{integer giving the number of header rows.}
}
\value{
If there is a column to the right of the data, the left_and_right
dataframe with this information added to it. If not, left_and_right is
returned unchanged.
}
\description{
When there is non-numeric information to the right of the
numeric data it has to be beheaded separately (because it is not numeric).
It is referred to here as the 'outer column'.

In order to associate the outer column with the correct rows in the
main data, this function takes the specified column and uses a full_join
to join it to the rest of the data as a column (full join used to ensure we
don't lose any rows, though a left_join should also work because of how data
have been created).

NOTE: This function currently only works for where there is just one column
in the right_block. Because of how col_location is created, it will only be
of length 1. If more col_locations are provided, an error is raised.
}
\examples{
\dontrun{
whole <- data.frame(
  sheet = "one header test data",
  row = c(15, 15, 15, 16, 16, 16),
  col = c(1, 2, 3, 1, 2, 3),
  character = c("date", "education", "Notes", "2020 Q1", NA, "note 1"),
  numeric = c(NA, NA, NA, NA, 100, NA)
)
left_and_right <- data.frame(
  sheet = "one header test data",
  row = 16,
  col = c(1, 3),
  date = c("2020 Q1", NA),
  description = c("education", NA),
  numeric = c(100, NA),
  character = c(NA, "note 1")
)

add_columns_to_right_of_data_back_in(whole, left_and_right, 3, 15, 1)
}
}
