% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_info.R
\name{add_metadata_columns}
\alias{add_metadata_columns}
\title{Add single value (metadata) variables as columns}
\usage{
add_metadata_columns(dat, ...)
}
\arguments{
\item{dat}{dataframe to which the ... params should be added as columns.}

\item{...}{variable names, each of which contains a single value.}
}
\value{
dat with additional columns
}
\description{
Include any variables given in dots (...) as a column in dat.
Each variable must be of length 1. The name of the column will be taken
as the name of the variable.
}
\details{
This function has a similar effect to add_single_value_columns, in that
it add columns to the data where the value of each column is the same on
every row. However, add_single_value_columns takes the column name and value
as separate vectors. add_single_value_columns can be controlled by user
settings. add_metadata_columns is not, so will add a column and fill it with
NA if there is no value for the variable in question.
}
\examples{
\dontrun{
dat <- tibble(
    Year = c(2023, 2023),
    Value = c(1, 2),
    vintage = c("provisional", "final")
)
units <- c("count", "thousands")
title <- c("Counts by year")
vintage <- "final"

# units will not be added as a column because it is of length >1,
# vintage will not be added because it already exists as a column
# title will be added:
add_metadata_columns(dat, title, units, vintage)
}
}
