% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_period_columns.R
\name{add_quarter_column}
\alias{add_quarter_column}
\title{Get quarter from a column containing more than just quarter.}
\usage{
add_quarter_column(dat, quarter_from_pattern, quarter_col_name)
}
\arguments{
\item{dat}{dataframe containing one column name that matches the
quarter_from_pattern}

\item{quarter_from_pattern}{character string. A regular expression that
matches only the column containing quarter information. In pub sec this
variable is supplied by quarter_from_col_pattern.}

\item{quarter_col_name}{character string. The name of the column that quarter
will be put in.}
}
\value{
dat with quarter column added. Warnings are raised about rows where
no single quarter or multiple single quarters have been found.
}
\description{
Extract single quarters (i.e. not quarter ranges) from a given
column e.g. quarter may be given alongside year.
}
\examples{
\dontrun{
dat <- data.frame(
  x = c("2023Q4", "2024Q1", "2023Q1 to Q3", "2023Q1 Q3", "Q4", "2021", "no")
)
add_quarter_column(dat, "x", "quarter")
}
}
