% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_info.R
\name{add_row_headers_column}
\alias{add_row_headers_column}
\title{Put row headers that others are nested under into their own column.}
\usage{
add_row_headers_column(
  dat,
  xlsx_cells_names,
  headers_col,
  nested_col,
  from_pattern,
  fill_dir = "down",
  group_row_na_identifier = "all"
)
}
\arguments{
\item{dat}{dataframe or tibble. Post being imported by xlsx_cells and
unpivotted. Must include a column called 'row'.}

\item{xlsx_cells_names}{character string vector. The names of columns in
dataframes imported using xlsx_sells. These columns are not included when
searching for column names that match regular expressions.}

\item{headers_col}{character string. The name of the column to put the row
header information in. In pub sec, this variable is specified with
name_for_group_row_column}

\item{nested_col}{character string. The name of the column in which to put
the information that falls under row header. In pub sec, this variable is
specified with name_for_nested_row_column.}

\item{from_pattern}{character string. Regular expression matching the name of
the column currently containing both the grouping variable and the nested
variable.}

\item{fill_dir}{string. Default is 'down'. The direction that headers will
be fillled in. If header rows are above the information nested under them,
the direction should be 'down'. If header information is below the info
nested under it, then dirction should be set to 'up' Options: "down", "up".
"downup" (i.e. first down and then up) or "updown" (first up and then down)
can also be used.}

\item{group_row_na_identifier}{string. Either the name of a column in dat
that contains NAs for rows containing the grouping variable, or 'all' (the
default), in which case  grouping variable rows are identified by being
associated with an otherwise  blank row.}
}
\value{
tibble. dat but with a new column for group_row
}
\description{
Create a column for grouping variables that are in the same
column as their nested items.

This function is for datasets where there are grouping variables as row names
that don't have a value associated with them because they are just the group
under which other items sit. Put these grouping items in a column in
their own right, and the nested items in another column.

Turn data that look like this:\tabular{ll}{
   from \tab numeric \cr
   Education \tab  \cr
   primary \tab 1 \cr
   secondary \tab 3 \cr
   Transport \tab  \cr
   roads \tab 5 \cr
}


Into this:\tabular{lll}{
   from \tab row_header \tab numeric \cr
   primary \tab Education \tab 1 \cr
   secondary \tab Education \tab 3 \cr
   roads \tab Transport \tab 5 \cr
}
}
\examples{
\dontrun{
sample_data <- data.frame(
  address = c('B3', 'B4', 'B5', 'B6', 'B7'),
  row = c(3:7),
  col = 2,
  is_blank = c(TRUE, FALSE, FALSE, TRUE, FALSE),
  data_type = c('blank', rep('numeric', 2), 'blank', 'numeric'),
  numeric = c(NA, 1, 3, NA, 5),
  subservice = c('education', 'primary',  'secondary', 'Transport', 'roads')
)

xlsx_cells_names <- c(
    "address", "row", "col", "is_blank", "data_type", "numeric"
    )
add_row_headers_column(
    dat = sample_data, xlsx_cells_names = xlsx_cells_names,
    headers_col = 'description_1', nested_col = 'description_2',
    from_pattern = '(?i)service', fill_dir = 'down', NA)
}
}
