% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_info.R
\name{add_single_value_columns}
\alias{add_single_value_columns}
\title{Add columns with unique values}
\usage{
add_single_value_columns(
  dat,
  column_names = NA,
  column_fillers = NA,
  description = NA
)
}
\arguments{
\item{dat}{dataframe to which single value columns will be added.}

\item{column_names}{vector of column names. Must
be the same length as single_value_values. In pub sec this variable
is specified by single_value_names.}

\item{column_fillers}{vector of values. The order and number of
values must correspond to the order of single_value_names. In pub sec
this variable is specified by single_value_values.}

\item{description}{character string. Default is NA. Used in messaging.}
}
\value{
the dataframe with new columns added and filled. Returns a warning
if a column of the given name already exists. If only one of column_names or
column_fillers is missing an error is raised. If both column_names and
column_fillers are NA, dat is returned with no changes.
}
\description{
Add columns to the data that each have a single specified value.
If a column of the given name already exists the action will be skipped and
a warning given.
}
\examples{
single_value_names <- "description"
single_value_values <- "net current expenditure"
dat <- data.frame(a = 1:3, b = 4:6)
add_single_value_columns(
dat, single_value_names, single_value_values
)
}
