% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_year_column.R
\name{add_single_year_to_possibilities}
\alias{add_single_year_to_possibilities}
\title{Add single_year as an option in possible cols}
\usage{
add_single_year_to_possibilities(
  possible_cols,
  use_single_year,
  year,
  single_overrides_all
)
}
\arguments{
\item{possible_cols}{named vector in which the name is the type of year
column and the value is the name of the column. This function assigns 'year'
to 'single_column'.}

\item{use_single_year}{boolean. If FALSE do nothing.}

\item{year}{character string or integer containing a valid year. Only
used if use_single_year is TRUE. The year with which to populate the 'year'
column. Must be of length 1 and must be specified if use_single_year is TRUE.}

\item{single_overrides_all}{boolean. If TRUE, and use_single_year is
TRUE, single_year overrides all other possible_cols. If FALSE or NA,
single_year is just appended to possible_cols.}
}
\value{
named vector possible_cols with a new name appended to it
(single_cols) if single_overrides_all is FALSE, or with only single_cols if
single_overrides_all is TRUE.
}
\description{
single_year is just a year and doesn't yet have a column. Check
that it only contains one year (either financial or calendar) and if so give
it the future column name 'year' and add it to the vector of possible columns
from which to make the 'year' column.
}
\examples{
\dontrun{
possible_cols <- c("financial" = NA, "calendar" = "Year", "mixed" = NA)

# single_overrides_all is FALSE
add_single_year_to_possibilities(possible_cols, TRUE, "2021/22", FALSE)

# single_overrides_all is TRUE
add_single_year_to_possibilities(possible_cols, TRUE, "2021/22", TRUE)

# Empty possible cols and single year exists
possible_cols <- c("financial" = NA, "calendar" = NA, "mixed" = NA)
add_single_year_to_possibilities(possible_cols, TRUE, "2021/22", FALSE)
}


}
