% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_info.R
\name{add_subtable_names}
\alias{add_subtable_names}
\title{Add subtable titles to the relevant rows of the data}
\usage{
add_subtable_names(dat, subtable_names)
}
\arguments{
\item{dat}{dataframe with 'row' and 'col' columns.}

\item{subtable_names}{dataframe with an integer column called 'row' and
a character column giving the subtitle found on that row. The name of this
character column is the name the column will have in the returned data.}
}
\value{
dataframe dat with subtable names added.
}
\description{
add subtable titles to all rows they apply to. It is assumed
that all rows require a subtable title entry and that the subtable title is
at the top of each subtable.
}
\examples{
\dontrun{

subtable_names <- data.frame(
    row = c(1, 3),
    subtitle = c("table A", "table B")
    )
dat <- data.frame(row = 1:4, col = 2, value = 10:13)
add_subtable_names(dat, subtable_names)
}
}
