% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_info.R
\name{add_totals_as_column}
\alias{add_totals_as_column}
\title{Give the name of the group that has totals in a new column}
\usage{
add_totals_as_column(
  dat,
  xlsx_cells_names,
  totals_col,
  from_pattern,
  fill_dir = "up"
)
}
\arguments{
\item{dat}{dataframe}

\item{xlsx_cells_names}{character string vector. The names of columns in
dataframes imported using xlsx_cells. These columns are not included when
searching for column names that match regular expressions.}

\item{totals_col}{character string. The name of the column in which to put
the names of the groups that have a total. In pub sec, this variable is
specified with name_for_total_column.}

\item{from_pattern}{character string. Regular expression matching the name of
the column currently containing both the totals. In pub sec this variable is
specified by col_with_totals_pattern}

\item{fill_dir}{string. Default is 'up'. The direction that names of the
totals will be filled in. If totals are below the information they sum,
the direction should be 'up'. If totals are above the items they sum,
then fill_dir should be set to 'down' Options: "down", "up".
"downup" (i.e. first down and then up) or "updown" (first up and then down)
can theoretically also be used.In pub_sec this variabble is specified with
total_column_fill_dir.}
}
\value{
tibble. dat but with a new column for the group that are totalled.
If the row does not have a total associated with it, the new column contains
the same entry as that in the 'from' column.
}
\description{
If a descriptor column has items followed by the totals of those
items, this function can be used to move the name of the totalled group into
a separate column and associate it with the items above it.
}
\details{
For example:\tabular{ll}{
   \strong{description} \tab \strong{value} \cr
   red \tab 10 \cr
   blue \tab 15 \cr
   Total squares \tab 25 \cr
   red \tab 1 \cr
   green \tab 15 \cr
   total circles \tab 16 \cr
}


becomes:\tabular{lll}{
   \strong{species} \tab \strong{group} \tab \strong{value} \cr
   red \tab squares \tab 10 \cr
   blue \tab squares \tab 15 \cr
   Total \tab squares \tab 25 \cr
   red \tab circles \tab 1 \cr
   green \tab circles \tab 15 \cr
   total \tab circles \tab 16 \cr
}
}
\examples{
\dontrun{
dat <- data.frame(
    item = c("A", "B", "total letts", "1.2.3", "total code", "total other"),
    numeric = c(1, 2, 3, 1, 1, 7)
    )

add_totals_as_column(
    dat, xlsx_cells_names = "numeric",
    totals_col = "col1", from_pattern = "item",
    fill_dir = NA
    )
}
}
