% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_variables.R
\name{assign_variables}
\alias{assign_variables}
\title{assign_variables}
\usage{
assign_variables(my_args)
}
\arguments{
\item{my_args}{A character vector containing the arguments. The first element should be "--args",
and the subsequent elements should be the actual arguments.}
}
\value{
An environment containing the assigned variables.
}
\description{
Puts command line arguments into an environment and passes it back to the main code
}
\details{
Puts command line arguments into an environment and passes it back to the main code
Assign Variables to Environment : this can be broken out to a library when we refactor 
to use our own R package.
This function creates a new environment and assigns variables from the provided arguments.
}
\examples{
my_args <- c("--args", "input/file/path", "tab_regex", "output/file/path", "current/location",
             "sheet_structure_json", "layout_json", "file_part")
my_env <- assign_variables(my_args)
}
\author{
Mark London \email{mark.london@ext.ons.gov.uk}
}
