% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot_data.R
\name{behead_left_block}
\alias{behead_left_block}
\title{Behead xlsx_cells data to the left the required number of times.}
\usage{
behead_left_block(dat, cols_to_name, first_header_row, header_row_count)
}
\arguments{
\item{dat}{dataframe imported using tidyxl::xlsx_cells.}

\item{cols_to_name}{character string vector. Columns to the left of the data
are given these as column names. The length of cols_to_name must equal the
number of times you want to behead left. i.e. if there are three character
columns to the left of the numeric data in Excel, cols_to_name will have 3
elements.}

\item{first_header_row}{integer. Row numebr the first header is on.}

\item{header_row_count}{integer. The number of header rows.}
}
\value{
dataframe that looks like the row descriptor columns in the Excel
data.
}
\description{
xlsx_cells imports data so that each row describes a single cell
in an xlsx table. This function re-associates cells to the left of data with
the corresponding values to the right (see example). The cols to the left of
the data must contain character data.
}
\examples{
\dontrun{
dat <- data.frame(
  address = c("A1", "B1", "C1", "A2", "B2", "C2"),
  row = c(1, 1, 1, 2, 2, 2),
  col = c(1, 2, 3, 1, 2, 3),
  data_type = c(rep("character", 4), rep("numeric", 2)),
  numeric = c(rep(NA, 4), 100, 22),
  character = c("Year", "primary", "secondary", "2021", NA, NA))

behead_left_block(dat, 'Year', 1,  1)
}
}
