% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_handlers.r
\name{build_regex}
\alias{build_regex}
\title{Convert instructions into a regex pattern}
\usage{
build_regex(instructions)
}
\arguments{
\item{instructions}{vector of strings containing words given in regex_lookup.
Each string in the vector can have multiple instruction words in it.
Each instruction word is converted to a regular expression and these are
pasted together to create a single pattern from each string. Must not
contain punctuation.}
}
\value{
a vector of regular expressions; one for each instruction string
Raises an error if there is an instruction that does not appear in
regex_lookup.
}
\description{
It can be problematic to pass regular expressions from Python
through subprocess. It is hard to isolate which patterns cause issues, but
the error raised in Python is '\if{html}{\out{<x>}} is not recognized as an internal or
external command'.
This function uses standard strings that don't cause problems in subprocess
to build the regular expression inside R.
}
\examples{
\dontrun{

instructions <- c("(?i)a[0-2]b", "^year$")
build_regex(instructions)

instructions <- c("ALT_colon_whitespace", "ALT_newline_poundsign_newline")
build_regex(instructions)
}
}
