% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all_nchar.R
\name{check_all_nchar}
\alias{check_all_nchar}
\title{check_all_nchar}
\usage{
check_all_nchar(dat, col_name, num_char)
}
\arguments{
\item{dat}{dataframe containing col_name}

\item{col_name}{string. Name of a column in dat. DColumn does not have to be character}

\item{num_char}{int. Number of characters you want to check for}
}
\value{
bool. TRUE if all entries (not including NAs) have the number of 
characters specified by num_char
}
\description{
Get the column names for the left block
}
\details{
Check that all entries in a column have a certain number of characters
}
\examples{
# example code
   
all4 <- data.frame(year = c("2020", "2021")) 
mixed <- data.frame(year = c("2020-21", "2021")) 

check_all_nchar(all4, "year", 4) # TRUE
check_all_nchar(mixed, "year", 4) # FALSE
}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
