% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_dropdown}
\alias{check_dropdown}
\title{Check if given string is present in the character column}
\usage{
check_dropdown(pattern, dat)
}
\arguments{
\item{pattern}{string. A regular expression that only matches the
dropdown cell. In pub sec this variable is specified by dropdown_pattern.}

\item{dat}{dataframe imported using tidyxl::xlsx_cells}
}
\value{
message or warning
}
\description{
Check if given string is present in the character column of
a dataframe improted using tidyxl::xlsx_cells(). Created for checking if
a given string is in a dropdown type cell (e.g. in MHCLG data 2024), but
could be used to check if a string exists anywhere in a dataset.
}
\examples{
\dontrun{#'
dat <- data.frame(
  row = c(1:3),
  col = rep(1,3),
  character = c("Title", "some other info", "England")
)

check_dropdown("Eng", dat) # found
check_dropdown("Wales", dat) # not found
}
}
