% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_for_duplicate_names}
\alias{check_for_duplicate_names}
\title{Check for duplicated column names}
\usage{
check_for_duplicate_names(dat)
}
\arguments{
\item{dat}{dataframe to check for duplicated names.}
}
\value{
Raises an error if duplicated names are found.
}
\description{
Check for duplicated column names and raise an informative error
if any are found.
}
\examples{
\dontrun{

dat <- data.frame(source = "scottish_gov", source = c("A", "B"), value = 1:2)
check_for_duplicate_names(dat)
}
}
