% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_handlers.R
\name{check_for_multiple_years}
\alias{check_for_multiple_years}
\title{Check if there are multiple different years in any one entry.}
\usage{
check_for_multiple_years(dat, from, prefer = "financial", type = "both")
}
\arguments{
\item{dat}{dataframe that contains a column with the name from \code{from}.}

\item{from}{character string. The name of the column containing year
information.}

\item{prefer}{character string. Either 'financial' (default), or 'calendar'.
If type is 'both' and both a calendar year and a financial year are found in
a single entry, this determines which is checked If type is 'single', prefer
determines whether calendar years or financial years are checked for multiple
different values.}

\item{type}{character string. Either 'both' (default), or 'single'. If
'single', a value with both a calendar year and a financial year will not
be classed as having multiple different values.}
}
\value{
boolean. TRUE if any entry contains more than one unique year.
}
\description{
Check if there are multiple different years in any  single entry
of specified column. Used internally by get_year.

If \code{type} is not specified, both financial and calendar years will be
checked - so if an entry contains both a financial year and a calendar year,
these will be classed as different years and TRUE will be returned.

If \code{type} = 'single' only multiple values of the preferred year (calendar or
financial) will be looked for, unless no values of that type are found, in
which case multiple values of the non-preferred year will be checked for.
}
\examples{
\dontrun{
check_for_multiple_years(data.frame(char = c("2024", "2025 2026")), "char")
check_for_multiple_years(data.frame(char = c("2024", "2025")), "char")
check_for_multiple_years(
    data.frame(char = c("2021 2022-23 2023-24")), "char",
    prefer = "financial",
    type = "single"
    )
check_for_multiple_years(
    data.frame(char = c("2021 2022-23 2023-24")), "char",
    prefer = "calendar",
    type = "single"
    )
check_for_multiple_years(
    data.frame(char = c("2021 2022 2023-24")), "char",
    prefer = "calendar",
    type = "single"
    )
}
}
