% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_na_error.R
\name{check_na_error}
\alias{check_na_error}
\title{check_na_error}
\usage{
check_na_error(...)
}
\arguments{
\item{...}{Variables to be checked for NA.}
}
\description{
Check if variables are NA and stop if any are
}
\details{
Check if variables are NA and stop if any are.

This function checks if the given variables are NA. If any of them are,
the function stops execution with an error message containing the variable
name and the message "is NA". Ellipsis (...) allows the function to accept
any number of arguments.
}
\examples{
header_identifier <- "header"
check_na_error(header_identifier)  # No error, as header_identifier is not NA

group_col <- NA
check_na_error(group_col)  # Error: "group_col is NA"

Note for multiple arguments

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
