% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_reserved_words}
\alias{check_reserved_words}
\title{Check for reserved words in settings}
\usage{
check_reserved_words(dict, exclude)
}
\arguments{
\item{dict}{dataframe. A dictionary-style dataframe with setting names in the
key column and the settings themselves in the value column.}

\item{exclude}{Dictionary keys that should not be checked because they are
expected to contain reserved words.}
}
\value{
boolean. TRUE if no settings use reserved words, otherwise an error
is raised
}
\description{
Check if any reserved words are used in the settings.
The list of reserved words is taken from the column names created by
xlsx_cells. Using them for new column names can cause issues that
are hard to identify.
The variables that must not be assigned a reserved word at time of writing
are group_col, nested_col_1 , nested_col_2, group_row, nested_row_1,
nested_row_2, and left_headers. However a conservative approach
is taken so that we don't need to specify new variables here
if and when new settings are added that also must not use reserved words.
}
\examples{
\dontrun{
dict_with_error <- tibble(
  "key" = c("header_identifier", "left_headers", "table_split_dirs"),
  "item" = list("safe", c("character", "numeric"), "row")
  )

dict_no_error <- tibble(
  "key" = c("header_identifier", "table_split_dirs"),
  "item" = list("safe", "row")
  )

 check_reserved_words(dict_with_error, c("table_split_dirs", NA))
 check_reserved_words(dict_no_error, c("table_split_dirs", NA))
}
}
