% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sheet.R
\name{clean_xlsx_cells_data}
\alias{clean_xlsx_cells_data}
\title{Remove unwanted cells from xlsx_cells data}
\usage{
clean_xlsx_cells_data(
  dat,
  cells_to_remove = NA,
  input_filepath,
  hidden_strings = NA
)
}
\arguments{
\item{dat}{data that has been imported using xlsx_cells.}

\item{hidden_strings}{bool. TRUE, FALSE or NA. Default is NA. FALSE leads to
the same behaviour as NA. In pub sec this variable is specified by
remove_hidden_strings_bool.}
}
\description{
Remove unwanted cells from xlsx_cells data
}
\details{
The cells removed by this function are usually specified as needing removal
because not doing so results in errors during beheading (unpivotting).
}
