% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_xlsx_cells_table_data.R
\name{clean_xlsx_cells_table_data}
\alias{clean_xlsx_cells_table_data}
\title{Prepare table for un-pivotting}
\usage{
clean_xlsx_cells_table_data(
  dat,
  extend_row_pattern = NA,
  extend_row_order = NA,
  extend_row_with = NA,
  cell_removal_patterns = NA,
  rows_to_check_for_removal = NA,
  combine_start_row_identifier = NA,
  combine_end_row_identifier = NA,
  columns_to_remove_patterns = NA,
  columns_to_remove_offset = NA,
  columns_to_create = NA,
  first_header_row = NA,
  target_all_rows = FALSE
)
}
\arguments{
\item{dat}{Dataframe created using tidyxl::xlsx_cells(), in which metadata
above the table has been removed.}

\item{extend_row_pattern}{character string. Regular expressions used to
identify the rows that need added information from the row above or below.}

\item{extend_row_order}{character string. Either 'forward' or 'reverse'. Used
to determine if the added information is placed before or after original row
value.}

\item{extend_row_with}{character string. Either 'above' or 'below'. Used to
determine if the added information is from above or below the row that
matches pattern.}

\item{cell_removal_patterns}{character string vector. Each must be a regular
expression used to identify unwanted cells. More than one pattern may be
given, but there should only be one pattern for each cell to be removed:
Only the first match to each pattern will be removed. If two patterns are
given and the first match for each is the same cell, only that one cell will
be removed. In pub sec this variable is specified by
metadata_cells_to_remove_patterns.}

\item{rows_to_check_for_removal}{integer. An integer specifying the
number of rows to check for matches to the metadata_cells_to_remove_patterns.
This is based on the rows of dat, not original row number (stated in the
column 'row') i.e. if this function is called after blank rows have been
removed, it will only check that many rows. If cell_removal_patterns are
specified, but rows_to_check_for_removal is NA, 5 will be used.}

\item{combine_start_row_identifier}{character string. A regular expression
that first appears in the first of the header rows that needs to be combined.}

\item{combine_end_row_identifier}{character string. A word or pattern that
first appears in the last of the header rows that needs to be combined.}

\item{columns_to_remove_patterns}{character string vector. Each is a regular
expression that matches the name of a column in the raw data to be removed.}

\item{columns_to_remove_offset}{integer vector. Must be either NA or the same
length as columns_to_remove_patterns If the column to remove is not named but
it's location is known in relation to a named column, specify the pattern for
the named column and use offset to move x columns to the left (negative) or
right (positive).}

\item{columns_to_create}{character vector. The names of columns that will
be created during behead. Used to calculate the number of header rows.}

\item{first_header_row}{integer. the number of the first row in which there
are headers.}

\item{target_all_rows}{bool. Default is false. If false the
columns_to_remove_patterns are only found if they have a match in the header
rows. If false the pattern is valid in any row.}
}
\value{
dataframe that is ready for un-pivotting.
}
\description{
Prepare data for use in unpivot_data. This includes removing
unwanted cells, rows, and columns, and completing partial information with
information from other cells.
}
\details{
See documentation for convert_date_to_char, extend_row_value,
remove_empty_lines, remove_metadata_cells, combine_rows_by_column and
remove_columns for more information.
}
