% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanr.R
\docType{package}
\name{cleanr}
\alias{tidysheet}
\alias{tidysheet-package}
\alias{cleanr}
\title{cleanr}
\description{
cleanr provides helper functions to preformat and clean the data files
prior to input to the python modules
}
\details{
cleanr: preformatting of PSST Pipeline files
}
\author{
Mark London \email{mark.london@ext.ons.gov.uk}
}
