% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_year_column.R
\name{clear_invalid_year_and_type}
\alias{clear_invalid_year_and_type}
\title{Set values to NA when year has more than 7 characters}
\usage{
clear_invalid_year_and_type(dat, year_to_na)
}
\arguments{
\item{dat}{dataframe. Must contain columns called 'year' and 'year_type'.}

\item{year_to_na}{boolean. If TRUE, year and year type will
both be NA when more than one year, or no valid years are found. If FALSE,
only year_type, and entries where NO years are found will be changed to NA.
Defaults to FALSE. In pub sec this variable is specified by
multi_year_entries_as_na.}
}
\description{
If year column contains an invalid entry such as multi-years
(e.g. '2020-21 to 2022-23'), year_type is set to NA. 'year' will also be set
to NA for these cases if year_to_na is TRUE.
}
\examples{
\dontrun{
dat <- tibble(
   year = c(NA, "note", "2021-22", "2021-22 to 2022-23"),
   year_type = c(NA, NA, "financial", "financial")
   )
clear_invalid_year_and_type(dat, TRUE)
clear_invalid_year_and_type(dat, FALSE)
}
}
