% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_headings.R
\name{combine_headings}
\alias{combine_headings}
\title{combine_headings}
\usage{
combine_headings(dat, headers_to_combine)
}
\arguments{
\item{dat}{dataframe}

\item{headers_to_combine}{vector of strings taken from the data dict. Each 
string is the identifier of a column - this can either be its position e.g.
'group_col' or the name of group_col, e.g. 'item'}
}
\value{
dat with columns concatenated
}
\description{
concatenate columns that hold raw data header information
}
\details{
concatenate columns that hold raw data header information
}
\examples{
dat <- data.frame("group" = c("Fox", "Fox", "Bear"),
                  "common_name" = c("Red", "Arctic", "Black"),
                  "item_detail" = c("Vulpes vulpes", "Vulpes lagopus", "Ursus americanus"),
                  "value" = 1:3)
newdat <- combine_headings(dat, 
                           headers_to_combine = c("item_detail", "group", "common_name"))
newdat  
}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
