% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_information.R
\name{combine_rows_by_column}
\alias{combine_rows_by_column}
\title{Combine consecutive rows within the same column.}
\usage{
combine_rows_by_column(dat, start_pattern = NA, end_pattern = NA)
}
\arguments{
\item{dat}{dataframe imported with tidyxl::xlsx_cells}

\item{start_pattern}{character string. A regular expression that first
appears in the first of the header rows that needs to be combined. In pub sec
this variable is specified by combine_start_row_identifier.}

\item{end_pattern}{character string. A word or pattern that first
appears in the last of the header rows that needs to be combined. In pub sec
this variable is specified by combine_end_row_identifier.}
}
\value{
dat dataframe with rows up to the end_row_identifier row
concatenated onto the start_row_identifier row for each column. Of the
concatenated rows, only the first is kept, the others are removed from the
data.
}
\description{
Use regular expression patterns to find the first
and last of the rows that need to be combined into a single value, and
concatenate those values. Cells will only be combined with others that sit in
the same column. Combined character strings are spearated by a space.

This was written to address an issue in the pub sec DLUHC (MHCLG) NNDR and
council tax data, where headings had been split over multiple Excel rows,
rather than using alt enter to get a new line in the cell (see example for a
simplified situation.
}
\examples{
\dontrun{

|      || this    || another |
|      ||---------||---------|
|      || is one  || heading |
|      ||---------||---------|
| year || heading || here    |
|------||---------||---------|
| 2020 ||   1162  ||  1034   |

The data shown above is input by xlsx_cells to look like this:

dat <- data.frame(
    "row" = c(1, 1, 2, 2, 3, 3, 3, 4, 4, 4),
    "col" = c(2, 3, 2, 3, 1, 2, 3, 1, 2, 3),
    "character" = c( "this", "another",
                     "is one", "heading",
                     "year", "heading", "here",
                     NA, NA, NA),
    "numeric" = c(rep(NA, 7),
                  2020, 1162, 1034),
    "data_type" = c(rep("character", 7), rep("numeric", 3))
)

result <- combine_rows_by_column(dat, "this", "here")
}
}
