% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_columns.R
\name{concatenate_columns}
\alias{concatenate_columns}
\title{concatenate_columns}
\usage{
concatenate_columns(dat, col_1, col_2)
}
\arguments{
\item{dat}{A dataframe containing columns used for col_1 and col_2.}

\item{col_1}{string. The column whose information will come first.}

\item{col_2}{string. The column whose information will come second. 
This column will then be deleted.}
}
\value{
tibble. The same as dat but containing the new column, and missing col_2.
}
\description{
Concatenate two columns
}
\details{
Concatenate two columns

Concatenate the strings from two columns. The new column overwrites col_1.
col_2 is deleted. Information from the two columns is separated by a bar (|)
}
\examples{
sample_data <- data.frame(
  notes = c('Note 1', NA, 'Note 2', NA),
  info = c(NA, "Info A", "Info B", NA)
)


concatenate_columns(sample_data, "notes", "info")

simple_table_behead(sample_data, 'subservice', 'transaction')

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
