% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_handlers.R
\name{consecutive_years_to_financial}
\alias{consecutive_years_to_financial}
\title{Convert consecutive years to a financial year}
\usage{
consecutive_years_to_financial(year)
}
\arguments{
\item{year}{character string or numeric vector of length 2.}
}
\value{
character string of length 1 in the format YYYY-YY. If the length
of the year vector is not 2 an error is raised. If the years found are not
consecutive or not in the format YYYY a warning is given and the input
year is returned.
}
\description{
Given two years that are consecutive, combine them to form a
single financial year in the format YYYY-YY
}
\examples{
\dontrun{
year <- consecutive_years_to_financial(c("2021", "2022"))
year <- consecutive_years_to_financial(c("2020", "2022"))
year <- consecutive_years_to_financial(c("2020-21", "2021-22"))
}
}
