% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_handlers.R
\name{convert_date_to_char}
\alias{convert_date_to_char}
\title{change dates to character strings}
\usage{
convert_date_to_char(dat)
}
\arguments{
\item{dat}{dataframe imported using tidyxl::xlsx_cells()}
}
\value{
dataframe dat with date strings set to be treated as character
strings.
}
\description{
Treat date strings as character string in xlsx_cells data.
If the data_type is 'date' and the date column is not NA, update
the data_type column to character and cut information from the date
column into the character column.
}
\examples{
\dontrun{
convert_date_to_char(source_df)
}
}
