% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_year_column.R
\name{convert_long_form_fy_to_short}
\alias{convert_long_form_fy_to_short}
\title{Convert long form calendar years to short form}
\usage{
convert_long_form_fy_to_short(year)
}
\arguments{
\item{year}{character string. A single year or a vector of single years.}
}
\value{
character string. year in short form. If year is a calendar year or
a financial year already in short from no action is taken.
}
\description{
Where financial year end is four characters e.g. YYYY-YYYY,
remove the first two digits of the end year to get YYYY-YY. This should
only be called on columns that are known to contain valid years as it does
not year format specific - it just removes the preantepenultimate and
antepenultimate characters.
}
\examples{
\dontrun{
convert_long_form_fy_to_short(c("2021-2022", "2022 to 2023"))
}
}
