% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_year_end_to_fy.R
\name{convert_year_end_to_fy}
\alias{convert_year_end_to_fy}
\title{convert_year_end_to_fy}
\usage{
convert_year_end_to_fy(dat, end_year)
}
\arguments{
\item{dat}{dataframe containing end_year}

\item{end_year}{string. Name of a column in dat containing years as yyyy}
}
\value{
tibble. dataframe with a new column  called year, with strings in 
the format yyyy-yy
}
\description{
Convert financial year end column (yyyy) to yyyy-yy
}
\details{
Convert financial year end column (yyyy) to yyyy-yy

Add a new column to a datframe. Use the financial year end (given as yyyy) to
generate a column that gives financial year in the format yyyy-yy.
}
\examples{
 
dat <- data.frame(fin_year_end = c("2020", "2021")) 
convert_year_end_to_fy(dat, "fin_year_end")

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
