% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_year_format.R
\name{convert_year_format}
\alias{convert_year_format}
\title{convert_year_format}
\usage{
convert_year_format(year_string)
}
\arguments{
\item{year_string}{character string e.g. 2021/22}
}
\value{
clean_year as the output replaced with hyphen in the year
}
\description{
replace non alphanumeric characters in the year strings with hyphens
}
\details{
replace non alphanumeric characters in the year strings with hyphens
}
\examples{
year <- c("2021/22", "2022/23")
convert_year_format(year)

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
