% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tables.R
\name{create_table_list}
\alias{create_table_list}
\title{Add two new tables to the list by splitting an existing table.}
\usage{
create_table_list(dat, table_id, split_loc, direction)
}
\arguments{
\item{dat}{list of dataframes originally imported using tidyxl::xlsx_cells().}

\item{table_id}{integer. The list index of the table that is to be split.}

\item{split_loc}{iteger giving the column or row number at which the second
new table starts.}

\item{direction}{character string. Either 'col' if the table is to be split
by column, or 'row' if it is to be split by row.}
}
\value{
list of dataframes. dat with two new tables appended to the end.
}
\description{
Split one of the tables in a list of tables in two, based on a
provided split location and whether it should be split by row or column. Add
the two new tables to the end of the list of tables.
}
\examples{
\dontrun{
table <- data.frame(
  "row" = c(1, 1, 1, 2, 2, 2),
  "col" = c(1, 2, 3, 1, 2, 3),
  "address" = c("A1", "B1", "C1", "A2", "B2", "C2"),
  "character"= c(NA,NA,NA,"table 1","NA","table 2")
)
dat <- list(table)

create_table_list(dat, 1, 3, 'col')
}
}
