% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_from_output.R
\name{deduplicate_data}
\alias{deduplicate_data}
\title{Remove repeated data rows}
\usage{
deduplicate_data(dat, ignore_cols = NA, arrange_by = NA)
}
\arguments{
\item{dat}{dataframe.}

\item{ignore_cols}{character vector of column names to be ignored.}

\item{arrange_by}{character vector of column names by which to arrange the
output dataframe.}
}
\value{
dataframe where only the first occurrence of each duplicate row is
retained.
}
\description{
Remove repeated rows from a dataset based on the content of
columns that were not originally in the xlsx_cells data. i.e. these rows
look like duplicates in the Excel data but not the xlsx_cells data (which
has one row per cell).

Columns specified in ignore_cols are excluded from comparison with other
rows.

A hash is generated for each row based on the content columns, and duplicates
are identified by comparing these hashes.
}
\examples{
\dontrun{
dat <- tibble(
    row = rep(c(1:3), times = 2),
    col = rep(c(1, 2), each = 3),
    description = c(NA, NA, NA, "A", "A", "B"),
    value = c(NA, NA, NA, 1, 1, 2)
    )
deduplicate_data(dat)
}
}
