% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_to_multiple_columns.R
\name{do_the_splits}
\alias{do_the_splits}
\title{split strings in the order provided}
\usage{
do_the_splits(dat, split_patterns, from)
}
\arguments{
\item{dat}{Dataframe output containing columns prefixed with
'tmp_use_split_point_', 'tmp_split_', 'tmp_matches_next_pattern_' and
'tmp_whitespace_'.}

\item{split_patterns}{character array of regex patterns.}

\item{from}{character string giving the name of the column holding strings to
be split.}
}
\value{
dat with columns that hold each element of the from strings after
being split. These column names are prefixed with 'tmp_string_'
}
\description{
Split strings in the 'from' column and put the resulting strings
in new columns with the prefix 'tmp_string_'. No 'to' column will be empty
even if not all 'from' strings contain matches to all the split_patterns
patterns.
}
\examples{
\dontrun{
split_patterns <- c( "(\n)\\\\s*1\\\\.", "(\n)\\\\s*2\\\\.", "(\n)\\\\s*3\\\\.")
split_point_descriptions <- c(
   "newline_whitespace_=1", "newline_whitespace_=2", "newline_whitespace_=3"
   )

dat <- data.frame(
  col_to_split = c("split\n 1.this \n 2.one \n 3.thousand",
                     "split\n 2.one"),
  tmp_use_split_point_0 = c(TRUE, TRUE),
  tmp_use_split_point_1 = c(TRUE, FALSE),
  tmp_use_split_point_2 = c(TRUE, TRUE),
  tmp_use_split_point_3 = c(TRUE, FALSE),
  tmp_split_1 = c(1, 2),
  tmp_split_2 = c(2, 2),
  tmp_split_3 = c(3, NA),
  tmp_matches_next_pattern_1 = c(FALSE, TRUE),
  tmp_matches_next_pattern_2 = c(FALSE, FALSE),
  tmp_matches_next_pattern_3 = c(FALSE, FALSE),
  tmp_whitespace_1 = c(TRUE, TRUE),
  tmp_whitespace_2 = c(TRUE, TRUE),
  tmp_whitespace_3 = c(TRUE, TRUE),
  value = c(100, 600)
)

do_the_splits(dat, split_patterns, 'col_to_split')
}
}
