% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_information.R
\name{drop_rows_with_NA}
\alias{drop_rows_with_NA}
\title{Remove rows with missing values in specified columns}
\usage{
drop_rows_with_NA(dat, patterns = NA)
}
\arguments{
\item{dat}{dataframe.}

\item{patterns}{vector of character strings. Each is a regular expression
that matches a column name that you want to check for blanks. In pub sec this
variable is specified by col_patterns_to_drop_NA_rows.}
}
\value{
dataframe with rows containing missing values in the matched
columns removed. Warnings are given stating which rows have been removed,
or that there was not a single column name that matched a pattern.
}
\description{
Remove rows with missing values in columns that match any
of the provided regular expressions. Rows are removed if any of the
specified columns contain blanks. In other words the only rows
to be kept are those that have a non-NA entry in ALL specified columns.
Note that this function is written to be performed on unpivotted (beheaded)
data.
}
\examples{
\dontrun{
data <- data.frame(
  row = 4:6,
  col = 1,
  year = c(2020, NA, 2021),
  value = c(100, NA, NA),
  description = c("A", "B", NA),
  address = c("A4", "A5", "A6")
)
patterns <- c("year", "value")
clean_data <- drop_rows_with_NA(data, patterns)
print(clean_data)
}
}
