% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_information.R
\name{drop_rows_with_values}
\alias{drop_rows_with_values}
\title{Drop rows with a specified value in a specified column}
\usage{
drop_rows_with_values(dat, col_patterns, value_patterns)
}
\arguments{
\item{dat}{dataframe. Must contain only one column name that matches
col_pattern.}

\item{col_patterns}{character vector. Regular expressions to match the
names of the columns containing the values that match value_patterns. In pub
sec this variable is specified with col_patterns_with_values_to_drop.}

\item{value_patterns}{character vector. Regular expressions to match the
values whose rows are to be dropped. In pub sec this variable is specified
with value_patterns_to_drop.}
}
\value{
datafame with rows removed.
}
\description{
Look for specified values in specified columns and drop rows
where a match is found.

Multiple values can be specified but there must be a col_pattern value for
every value_pattern, even if they are all to be dropped from the same column.
}
\examples{
\dontrun{
dat <- data.frame(
    a = c("ant", "apple", NA, "anger"),
    b = c("badger", "banana", "boredom", NA)
    )
drop_rows_with_values(dat, c("b", "a"), c("bad", "ang"))
}

}
