% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_read_data.R
\name{e_read_data}
\alias{e_read_data}
\title{e_read_data}
\usage{
e_read_data()
}
\description{
e_read_data
}
\details{
reads in files from pub_sec and builds a corpus
}
\author{
Mark London \email{mark.london@ext.ons.gov.uk}
}
