% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_function.R
\name{example_function}
\alias{example_function}
\title{example_function}
\usage{
example_function(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
value
An empty function to copy as a template for new functions.
Delete when the package is finished
}
\details{
value
An Example Function
}
\examples{
example_function()
}
\author{
Mark London \email{mark.london@ext.ons.gov.uk}
}
\keyword{cats}
