% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_handlers.R
\name{extract_all_years}
\alias{extract_all_years}
\title{Identify all years from a vector of character strings}
\usage{
extract_all_years(dat, type = "both")
}
\arguments{
\item{dat}{numeric or character string. Accepts vector, matrices and
dataframe data.}

\item{type}{character string. Either 'both' (default), 'financial', or
'calendar'.}
}
\value{
character string vector, or NA. Returns a vector of all strings
that matched the pattern for the year_type. If no years are identified in
any of the input strings, NA is returned.
}
\description{
Get all years (from the 20th and 21st centuries).
If no years are found NA is returned. If one or more years are found, unique
years are returned.
}
\examples{
\dontrun{
test_dat <- c("a date: 1998-1999", "1995 6", "1993 end", "a1995", "1995a", "a
date:1995, another date: 2012", "not a date: 475", "also not a date: 1234",
"abc 2021-22",  "34200943", "190932")
extract_all_years(test_dat)

test_dat2 <- data.frame(year_unedited = c("2021q1", "2022"))
test_dat2 \%>\% mutate(year = extract_all_years(year_unedited, 'calendar'))
}
}
\seealso{
\code{\link[=get_year]{get_year()}} for getting the first year from each entry of a
dataframe column.
}
