% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_information.R
\name{extract_matches}
\alias{extract_matches}
\title{Extract patterns from a vector of strings}
\usage{
extract_matches(dat, pattern)
}
\arguments{
\item{dat}{vector of numeric or string elements.}

\item{pattern}{character string. A single regular expression.}
}
\value{
List of character strings. Each item of the input is returned in
it's own vector, with each match given as an item in that vector. The output
is therefore a list of vectors. NA is returned for elements of the input that
do not contain a match to the pattern.
}
\description{
Look for a pattern and return the parts of the strings that
matched as a list. Converts numeric input to strings.
}
\examples{
numbers <- extract_matches(c( "a2", "2 and 3", "none"), '[0-9]')

}
