% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_information.R
\name{extract_units}
\alias{extract_units}
\title{Extract information about units}
\usage{
extract_units(dat, pattern)
}
\arguments{
\item{dat}{dataframe imported using tidyxl::xlsx_cells() containing the cells
that are likely to include information about units.}

\item{pattern}{character string. Regular expression used for matching units.}
}
\value{
A string containing extracted unique unit labels separated by
commas, or an empty string if no units are found.
}
\description{
Extract unit labels such as 'thousand', 'million', '000s', or
'£ Millions' from the character column of data imported using
tidyxl::xlsx_cells().
}
\examples{
\dontrun{
info_at_top_of_sheet <- data.frame(
    character = c(
        "Revenue in £ millions (millions)",
        "All figures are in thousands of pounds (£ thousand)."
        )
    )
pattern <- "((?i)thousand|(?i)000s)|(?i)million"
units <- extract_units(info_at_top_of_sheet, pattern)
}
}
