% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_info.R
\name{fill_blanks}
\alias{fill_blanks}
\title{fill blanks in a descriptor column}
\usage{
fill_blanks(dat, column_names, fill_dirs = "down")
}
\arguments{
\item{dat}{dataframe.}

\item{column_names}{charater string vector. The names of the columns to
apply the function to. In pub sec this variable is specified using
fill_columns_column_names.}

\item{fill_dirs}{character string vector. Must be the same length as
column_names. Values allowed are "down", "up", "downup" (i.e. first down and
then up) or "updown" (first up and then down). Defaults to 'down'. In pub sec
this variable is specified using fill_columns_fill_dirs.}
}
\value{
dat with blanks filled.
}
\description{
If data have row headers that are nested, with some being in
e.g. merged cells, this function can be used to fill the blanks with the
descriptors above or below.
}
\examples{
\dontrun{
dat <- data.frame(
    var1 = c("colour", NA, NA, "wavelength"),
    var2 = c("red", "green", "blue", "short")
    )
fill_blanks(dat, "var1", "down")
}

}
