% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_period_columns.R
\name{fill_missing_fy_start}
\alias{fill_missing_fy_start}
\title{Fill missing financial year start values.}
\usage{
fill_missing_fy_start(dat)
}
\arguments{
\item{dat}{dataframe containing columns called 'year' and 'year_type'.}
}
\value{
dataframe. dat with fy_start filled in for financial years.
}
\description{
Fill in missing fy_start values for rows where year_type is
'financial'.
}
\details{
If a row has 'financial' as year_type and fy_start is missing (or
the column does not exist), extract the first four digits from the year
column and use this for fy_start. Note that this requires the year column to
ONLY contain year.

Rows with other year_type values are left unchanged. This function is
intended for use with datasets that contain both calendar and financial years
but could be used in other scenarios.
}
\examples{
\dontrun{
dat <- data.frame(
    year = c("2020", "2020-21", "2021-22"),
    year_type = c("calendar", "financial", "financial"),
    fy_start = c(NA, NA, "2021")
    )

fill_missing_fy_start(dat)

}
}
