% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_info.R
\name{fill_missing_info}
\alias{fill_missing_info}
\title{Fill missing info in un-pivotted data}
\usage{
fill_missing_info(
  dat,
  xlsx_cells_names,
  fill_columns_column_names,
  fill_columns_fill_dirs,
  single_value_names,
  single_value_values,
  main_dropdown_name,
  main_dropdown_value,
  table_dropdown_name,
  table_dropdown_value,
  name_for_total_column,
  col_with_totals_pattern,
  total_column_fill_dir,
  subtable_names,
  name_for_group_row_column,
  name_for_nested_row_column,
  col_with_row_headers_pattern,
  row_header_fill_dir,
  group_row_na_identifier,
  POSIX_column,
  vintage_with_year_col_pattern,
  title,
  table_title,
  units,
  vintage,
  supplier,
  source_group,
  dataset
)
}
\arguments{
\item{dat}{dataframe.}

\item{xlsx_cells_names}{character string vector. Column names of the source
data which will have been imported using xlsx_cells.}

\item{fill_columns_column_names}{charater string vector. The names of the
columns in which information above or below NA rows will be used to replace
the NAs.}

\item{fill_columns_fill_dirs}{character string vector. Must be the same
length as fill_columns_column_names Values allowed are "down", "up", "downup"
(i.e. first down and then up) or "updown" (first up and then down). If NA,
and fill_columns_column_names is not NA, it defaults to 'down'.}

\item{single_value_names}{vector of column names. Must
be the same length as single_value_values.}

\item{single_value_values}{vector of values. The order and number of
values must correspond to the order of single_value_names.}

\item{main_dropdown_name}{character_string. The name of the column you want to
populate with main_dropdown_value. In pub sec this variable is specified
using dropdown_name.}

\item{main_dropdown_value}{character_string. The value found in the dropdown
cell at the top of the sheet.}

\item{table_dropdown_name}{character_string. The name of the column you want
to populate with table_dropdown_value.}

\item{table_dropdown_value}{character_string. The value found in the dropdown
cell above the table.}

\item{name_for_total_column}{character string. The name of the column in
which to put the names of the groups that have a total.}

\item{col_with_totals_pattern}{character string. Regular expression matching
the name of the column currently containing the totals.}

\item{total_column_fill_dir}{character string. Default is 'up'. The direction
that names of the totals will be filled in. If totals are below the
information they sum, the direction should be 'up'. If totals are above the
items they sum, then fill_dir should be set to 'down' Options: "down", "up".
"downup" (i.e. first down and then up) or "updown" (first up and then down)
can theoretically also be used.}

\item{subtable_names}{dataframe with an integer column called 'row' and
a character column giving the subtitle found on that row. The name of this
character column is the name the column will have in the returned data. In
pub sec this is created by get_subtitles() using the settings
subtable_title_column, subtable_title_patterns, subtitle_offset, and
subtitle_horizontal_index}

\item{name_for_group_row_column}{character string. The name of the column to
put row header information in when putting row headers that other
descriptions are nested under into their own column.}

\item{name_for_nested_row_column}{character string. The name of the column
in which to put information that falls under row headers. Used when putting
row headers that other descriptions are nested under into their own column.}

\item{col_with_row_headers_pattern}{character string. Regular expression
matching the name of the column currently containing both the grouping
variable and the nested variable. Used when putting row headers that other
descriptions are nested under into their own column.}

\item{row_header_fill_dir}{string. The direction that headers will
be filled in. If header rows are above the information nested under them,
the direction should be 'down'. If header information is below the info
nested under it, then direction should be set to 'up' Options: "down", "up".
"downup" (i.e. first down and then up) or "updown" (first up and then down)
can also be used. If other row_headers settings are not NA, the default is
'down'.}

\item{group_row_na_identifier}{string. Either the name of a column in dat
that contains NAs for rows containing the grouping variable, or 'all' (the
default), in which case grouping variable rows are identified by being
associated with an otherwise blank row.}

\item{POSIX_column}{character string. The exact name of the column containing
the date.}

\item{vintage_with_year_col_pattern}{character string. A single regular
expression to match a column containing both the year and the vintage. Year
can be financial or calendar. Accepted vintages are budget, provisional,
forecast, final, and actual (not case sensitive), though these could be
expanded upon.}

\item{title}{character string. The title of the sheet.}

\item{table_title}{character string. The title of the table.}

\item{units}{character string. The units of the table.}

\item{vintage}{character string. The vintage of the table (budget,
provisional, or final).}

\item{supplier}{character string. The data supplier, e.g. scottish_gov,
MHCLG etc.}

\item{source_group}{character string. The name of the source (the name of
the collection of data from a webpage) e.g. revenue_expediture_budget.}

\item{dataset}{character string. The name of the dataset.}
}
\value{
dataframe with additions made.
}
\description{
Fill missing information in both rows and columns of
un-pivotted data (i.e. a standard data frame rather than xlsx_cells data).
}
\details{
This function includes some functions that deal with time periods
(split_date_to_columns, and split_year_and_vintage), but most of that is
done in add_time_period_columns.

Whilst vintage is added as a column if vintage_with_year_col_pattern is
specified, it will not at this point be refined e.g 'forecast' will not be
converted to 'budget' by this function. To do so replace_string variables
need to be specified, and are called y functions later in tidy_sheet.

For more details see documentation for rename_reserved_colnames, fill_blanks,
add_single_value_columns, add_totals_as_column, add_subtable_names,
add_row_headers_column, split_date_to_columns, split_year_and_vintage, and
add_metadata_columns.
}
