% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_information.R
\name{find_all_instances}
\alias{find_all_instances}
\title{Find all rows or columns where a pattern occurs}
\usage{
find_all_instances(dat, pattern, direction = "row")
}
\arguments{
\item{dat}{An xlsx_cells dataframe}

\item{pattern}{character vector of length 1. Regular expression to be matched.}

\item{direction}{"row" or "col". Default is row. If col, the column numbers in
which the pattern occurs is returned.}
}
\value{
numeric vector. A list of unique rows/cols from the xlsx where the
pattern occurs.
Returns an empty vector if no matches are found.
}
\description{
Take an xlsx_cells dataframe and a pattern known to be
in the character cells of the rows or columns you wish to identify. Note that
numeric cells are ignored even if pattern is a number.
}
\examples{
\dontrun{
dat <- data.frame("character" = c("title", "million", NA,
                                 "Education services", "Highways services",
                                 NA),
                  "row" = c(1:3, 4, 4, 5),
                  "numeric" = c(rep(NA, times = 5), 50))
find_all_instances(dat, "services")
}
}
