% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_to_multiple_columns.R
\name{flag_consecutive_matching_splits}
\alias{flag_consecutive_matching_splits}
\title{Identify consecutive matching split patterns.}
\usage{
flag_consecutive_matching_splits(dat, split_patterns, split_point_descriptions)
}
\arguments{
\item{dat}{Dataframe output containing columns prefixed with 'split_'.}

\item{split_patterns}{character array of regex patterns.}

\item{split_point_descriptions}{character array of regex pattern
descriptions passed from Python. N.B. white space type patterns are currently
covered by 'whtespace' and 'newline'. If any more are added, they will also
need adding to the code in this function}
}
\value{
dat with 2 new columns for each split_patterns. One prefixed with
'matches_next_pattern_' the other prefixed with 'whitespace_'.
}
\description{
Adds boolean columns prefixed with 'matches_next_pattern_' and
'whitespace_' for every split_point pattern. 'matches_next_pattern_' states
whether a split_patterns pattern is the same as the previous pattern.
'whitespace_' states whether the split_point_description starts with either
'whitespace' or 'newline'.

Eventual use (see remove_repeated_pattern_from_remaining):
The characters that match the split_point pattern are not themselves removed
from the 'tmp_remaining' column in do_the_splits().For example if we take the
value 'split \emph{newline} this \emph{newline} here' and the split_patterns are
c('newline', 'newline') after the first split string_1 would be 'split' and '
tmp_remaining' would be
'\emph{newline} this \emph{newline} here'.
If didn't have the output of this function) the second split
would result in string_2 being '' and tmp_remaining being
'\emph{newline} this \emph{newline} here',
because it would have split at the first newline.

The 'whitespace_' columns are useful later on because if there are
consecutive matching split patterns that satrt with whitespace or newline we
can just remove that element of the pattern. Otherwise we need to remove the
whole pattern which may contain useful info that we would rather keep.

IMPORTANT: 'whitespace' and 'newline' are referenced in build_regex. Changes
to the whitespace type patterns there should also be reflected in this
funciton.
}
\examples{
\dontrun{
patterns <- c( "(\n)\\\\s*1\\\\.", "(\n)\\\\s*2\\\\.", "(\n)\\\\s*3\\\\.")
split_point_descriptions <- c(
   "newline_whitespace_=1", "newline_whitespace_=2", "newline_whitespace_=3"
   )

dat <- data.frame(
  col_to_split = c("split\n 1.this \n 2.one \n 3.thousand", "split\n 2.one"),
  tmp_split_1 = c(1, 2),
  tmp_split_2 = c(2, 2),
  tmp_split_3 = c(3, NA),
  value = c(100, 600)
)

flag_consecutive_matching_splits(dat, patterns, split_point_descriptions)
}
}
