% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_to_multiple_columns.R
\name{flag_existing_split_points}
\alias{flag_existing_split_points}
\title{Add a boolean column stating whether pattern was found}
\usage{
flag_existing_split_points(dat, split_patterns, from)
}
\arguments{
\item{dat}{dataframe containing a column with the name given in 'from'}

\item{split_patterns}{array of regex strings. One for each place the 'from'
column should be split. Each pattern is only used once.}

\item{from}{string. The name of the column containing the strings to be split.}
}
\value{
dataframe with a tmp_use_split_point boolean column for each
split_pattern (column name appended with a number). An extra final
tmp_use_split_point  column is added that is not related to a split pattern,
but is instead a  mechanism to avoid NAs in columns that will be added later.
}
\description{
For each split point add a column e.g. 'use_split_point_1' that
states whether that split point (in this case the first) was found in the
'from' column.

split_patterns must be given in the order in which they are found - items
where the first split pattern occurs after the second split pattern will
only be split at the second pattern.
}
\examples{
\dontrun{
split_patterns <- c( "(\n)\\\\s*1\\\\.", "(\n)\\\\s*2\\\\.", "(\n)\\\\s*3\\\\.")

dat <- data.frame(
  "col_to_split" = c("split\n 1.this \n 2.one \n 3.thousand", "split\n 2.one"),
  "value" = c(100, 600)
)
flag_existing_split_points(dat, split_patterns, "col_to_split")
}
}
