% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_handlers.R
\name{get_bookend_year}
\alias{get_bookend_year}
\title{Get the years before or after the current year}
\usage{
get_bookend_year(year, x, type)
}
\arguments{
\item{year}{character string or integer. Only a single year must be given,
but this can be a financial year (YYYY-YY or YYYY-YYYY) or calendar (YYYY).}

\item{x}{integer. The number of years before (negative) ot after (positive)
year that you want to be returned.}

\item{type}{character string. Must be "financial" or "calendar"}
}
\value{
calendar string. Theyear that is x years before or after the
provided year.
}
\description{
Get the string for the year x years before or after the given
year. Only a single year (calendar or financial must be given). If multiple
valid years (or no valid years) are provided, NA is returned with a warning.
}
\examples{
\dontrun{
get_bookend_year("2021-22", -1, "financial")
get_bookend_year("2021-22", 1, "financial")
get_bookend_year("2021", -1, "calendar")
}
}
