% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_information.R
\name{get_col_letters_as_string}
\alias{get_col_letters_as_string}
\title{Get unique column letters from col number in an xlsx_cells dataset.}
\usage{
get_col_letters_as_string(dat, col_numbers)
}
\arguments{
\item{dat}{Dataframe. Imported using xlsx_cells}

\item{col_numbers}{vector of integers}
}
\value{
character string of comma separated column letters.
}
\description{
Use cell addresses in an xlsx_cells dataframe to get the Excel
column letters for a string of comma-separated column  numbers.
Required for messages to users - users are used to
looking up a  column in an Excel file based on letters not numbers.
}
\details{
For alternative code that gets excel column letters from column number
without an xlsx_cells dataframe for lookup, see
https://stackoverflow.com/questions/9905533/convert-excel-column-alphabet-e-g-aa-to-number-e-g-25
}
\examples{
\dontrun{
dat <- data.frame(
  "row" = c(1, 1, 1, 2, 2, 2),
  "col" = c(1, 2, 27, 1, 2, 27),
  "address" = c("A1", "B1", "AA1", "A2", "B2", "AA2")
)

get_col_letters_as_string(dat, c(1, 2))
}
}
