% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_info.R
\name{get_col_with_row_headers_name}
\alias{get_col_with_row_headers_name}
\title{Get the name of the column containing row headers}
\usage{
get_col_with_row_headers_name(possibilities, pattern)
}
\arguments{
\item{possibilities}{character string vector. The names of possible columns.
In pub sec this will be the names of the left headers.}

\item{pattern}{character string. A regular expression that matches the name of the
column with row headers. In pub sec this variable is specified with
col_with_row_headers_pattern.}
}
\value{
The name of the column matching the pattern. If more than one match
is found, the first is returned with a warning. If none are found the first
of the possibilities is found with a warning. If pattern is NA, NA is
returned.
}
\description{
If a column contains row headers, use the provided pattern
to get the name of that column.
}
\examples{
\dontrun{
get_col_with_row_headers_name(c("column 1", "column 2"), "2")
}
}
