% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_information.R
\name{get_colnames_from_pattern}
\alias{get_colnames_from_pattern}
\title{Get column names that match regular expressions}
\usage{
get_colnames_from_pattern(dat, identities, patterns)
}
\arguments{
\item{dat}{dataframe containing the column names to find.}

\item{identities}{vector of character strings. Each is a name that will be
given to the column whose name matches the relevant pattern.}

\item{patterns}{vector of character strings. Each is a regular expression
that is used to identify the column to rename. Must be given in the same
order as columns.}
}
\value{
named vector - names are from identities, values are the names of
the columns that matched the pattern. If the number of matches is not 1 a
warning is given and NA is returned for that item.
}
\description{
For each pattern supplied column name that matches is returned.
}
\examples{
dat <- data.frame(
  Year = "2010",
  Value = 1,
  `Item name` = "foo"
)

colnames <- get_colnames_from_pattern(
    dat, c("year", "item"), c("(?i)year", "(?i)item")
    )
}
