% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_year_column.R
\name{get_column_with_most_valid_years}
\alias{get_column_with_most_valid_years}
\title{Get the column name that has the highest count value}
\usage{
get_column_with_most_valid_years(columns, counts)
}
\arguments{
\item{columns}{character string vector of the same length as counts.}

\item{counts}{integer vector of the same length as columns.}
}
\value{
character string. The column name associated with the maximum count.
If more than one column is has the maximum number of valid years, a warning
is raised stating that more than one column was found of the same type, and
which has been chosen. The chosen column will be 'year' if that exists, or
the first of the years with the maximum number of valid years.
}
\description{
Get the 'columns' string that is in the same position as the
maximum count value.
}
\details{
Designed to select the name of the column that holds the
most valid years, where 'columns' gives the names of the columns holding
valid years, and counts gives the number of valid years found in each of those
columns, though it could be used in other contexts.
}
\examples{
\dontrun{
get_column_with_most_valid_years(c("year", "year_notes"), c(25, 2))
get_column_with_most_valid_years(
    c("year", "year2", "year_notes"), c(25, 25, 2)
    )
}
}
