% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Read in data using xlsx_cells}
\usage{
get_data(filepath, pattern)
}
\arguments{
\item{filepath}{character string. Full filepath of an xlsx file.}

\item{pattern}{character string. Regular expression to match the required
sheet.}
}
\value{
dataframe in an xlsx_cells format with one row for every cell of
the excel data. If pattern is not provided NULL is returned. If filepath
is not fir an xlsx file an error is raised.
}
\description{
Determine which sheet to read based on a regular expression
and read in the data using tidyxl::xlsx_cells().
}
