% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_handlers.r
\name{get_dict_item}
\alias{get_dict_item}
\title{Get value for a key in a string_to_dict style dataframe}
\usage{
get_dict_item(key, dict)
}
\arguments{
\item{key}{character string to look up the value for.}

\item{dict}{dataframe with key and item columns.}
}
\value{
The item from dict for the key in question. If the key does not
exist NA is returned.
}
\description{
When a dict is passed from python and converted to a dataframe
by string_to_dict() it has a key column and a value column. Get the value
on the row of the given key.
}
\examples{
\dontrun{
layout_dict <- data.frame("key" = c("header", "group"),
                           "item" = c("NA", "mice"))
group <- get_dict_item("header", layout_dict)
print(group)
}
}
