% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_information.R
\name{get_first_data_row_number}
\alias{get_first_data_row_number}
\title{Get the row number for the first line of data}
\usage{
get_first_data_row_number(
  dat,
  first_header_row,
  columns_to_create,
  combine_start_row_identifier,
  combine_end_row_identifier
)
}
\arguments{
\item{dat}{dataframe imported using xlsx_cells and not yet pivotted i.e. all
character strings are still in the character column.}

\item{first_header_row}{integer. The row number on which the first row of
headers can be found.}

\item{columns_to_create}{integer. Vector of names that will be used for
column names after unpivotting the data. If combine_end_row_identifier is NA,
this will be the same length as the number of header rows.}

\item{combine_start_row_identifier}{character string. Regular expression whose
first match is in the last row of the headers prior to header rows being
combined.}

\item{combine_end_row_identifier}{character string. Regular expression whose
first match is in the last row of the headers prior to header rows being
combined.}
}
\value{
integer. The number of the first populated row after the last header
row
}
\description{
In all cases except where combine_rows_by_column is required, the number
of the first row of data can be determined using first_header_row and the
number of names in columns_to_create. However, where column names have
been split over multiple rows, and need to be recombined, this method does
not work. In such cases the combine_start_row_identifier and
combine_end_row_identifier patterns are also required.
}
\examples{
\dontrun{

dat <- data.frame(
    row = rep(1:5, each = 3),
    col = rep(1:3, times = 5),
    character = c(NA, "A", "B",
                  NA, "messed", NA,
                  NA, "up", NA,
                  NA, "header", "simple",
                  "count", NA, NA),
    numeric = c(rep(NA, 13), 4, 1),
    is_blank = c(TRUE, FALSE, FALSE,
                TRUE, FALSE, TRUE,
                TRUE, FALSE, TRUE,
                TRUE, FALSE, FALSE,
                FALSE, FALSE, FALSE)
    )

get_first_data_row_number(dat, 1, c("type", "name"), "messed", "simple")

dat <- data.frame(
    row = rep(1:5, each = 3),
    col = rep(1:3, times = 5),
    character = c(NA, "messed", NA,
                  NA, "up", NA,
                  NA, "header", "simple",
                  NA, "A", "B",
                  "count", NA, NA),
    numeric = c(rep(NA, 13), 4, 1),
    is_blank = c(TRUE, FALSE, TRUE,
                TRUE, FALSE, TRUE,
                TRUE, FALSE, FALSE,
                TRUE, FALSE, FALSE,
                FALSE, FALSE, FALSE)
    )
# With a blank row in headers
dat <- data.frame(
    row = rep(1:6, each = 3),
    col = rep(1:3, times = 6),
    character = c(NA, "messed", NA,
                  NA, NA, NA,
                  NA, "up", NA,
                  NA, "header", "simple",
                  NA, "A", "B",
                  "count", NA, NA),
    numeric = c(rep(NA, 16), 4, 1),
    is_blank = c(TRUE, FALSE, TRUE,
                TRUE, TRUE, TRUE,
                TRUE, FALSE, TRUE,
                TRUE, FALSE, FALSE,
                TRUE, FALSE, FALSE,
                FALSE, FALSE, FALSE)
    )

}

}
