% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_header_row.R
\name{get_first_header_row}
\alias{get_first_header_row}
\title{get_first_header_row}
\usage{
get_first_header_row(
  date_as_character,
  header_identifiers,
  second_identifiers,
  file_part
)
}
\arguments{
\item{date_as_character}{A data table of the excel file.}

\item{header_identifiers}{A list of header identifiers to search for in the file part.}

\item{second_identifiers}{A list of second identifiers to search for in the file part.}

\item{file_part}{A file part to search for the identifiers, usually 1, or sometimes 2.}
}
\value{
The index of the first header row.
}
\description{
Get the first header row index based on specified identifiers

This function searches for the specified header identifiers and second identifiers (if provided) in the file part.
It then identifies the first row of the header based on the found identifiers and returns its index.
}
\details{
This function takes a date as a character, a list of header identifiers, a list of second identifiers, and a file part as input.
It identifies the first row of the header based on the specified identifiers and returns its index.
}
\author{
Mark London \email{mark.london@ext.ons.gov.uk}
}
