% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot_data.R
\name{get_first_of_consecutives}
\alias{get_first_of_consecutives}
\title{Get the first integer where the following number is consecutive}
\usage{
get_first_of_consecutives(x)
}
\arguments{
\item{x}{vector of integers}
}
\value{
integer. If there are consecutive values, this will be the first of
the consecutive values. If there are no consecutive values a warning will
be given and the first value returned.
}
\description{
Get the first number in a vector where the next number is
the consecutive number.

For example, if you pass it a vector of numbers that are column positions
where the datatype is numeric, it will return the position of the first
column where both it and the following column are numeric.
}
\examples{
\dontrun{
x <- c(1, 1, 3, 5, 6, 8)
get_first_of_consecutives(x)
}
}
