% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flanked_sequences.R
\name{get_flanked_sequences}
\alias{get_flanked_sequences}
\title{get_first_data_row}
\usage{
get_flanked_sequences(flank_nums, target_nums)
}
\arguments{
\item{flank_nums}{integer vector.}

\item{target_nums}{integer vector. Numbers that}
}
\description{
Get the number of the row that contains the first line of data
}
\details{
Identify consecutive runs of numbers that are flanked on both sides by 
numbers in another sequence.
}
\examples{
flank_nums <- c( 5, 6, 8, 9, 12, 13, 15)
target_nums <- c(10, 14:15)
# missing numbers 7 and 11 from both series
# so result will be 13, 14, 15 
get_flanked_sequences(flank_nums, target_nums)
}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
