% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fy_start.R
\name{get_fy_from_fy_end}
\alias{get_fy_from_fy_end}
\title{Convert financial year end column (yyyy) to yyyy-yy}
\usage{
get_fy_from_fy_end(dat, end_year)
}
\arguments{
\item{dat}{dataframe containing end_year.}

\item{end_year}{character string. Name of a column in dat containing years as
yyyy.}
}
\value{
dataframe dat with a new column  called year, with strings in
the format yyyy-yy, using the previous year to the one passed in.
}
\description{
Add a new column to a daatframe. Use the financial year end
(given as yyyy) to generate a column that gives financial year in the format
yyyy-yy.

If the end_year column contains invalid year entries (no numbers, or a number
that is not four characters long), year will be given as NA, and the user
is given a warning.
}
\examples{
\dontrun{
dat <- data.frame(fin_year_end = c("2020", "2021", "999", "non"))
get_fy_from_fy_end(dat, "fin_year_end")
}
}
