% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fy_start.R
\name{get_fy_start}
\alias{get_fy_start}
\title{Get fy_start from the available information on year.}
\usage{
get_fy_start(
  dat,
  year_from_pattern = NA,
  fy_from_fy_end = FALSE,
  fy_start_from_fy_end = FALSE,
  fy_end_pattern = NA,
  calendar_to_fy_start = FALSE,
  q1_is_jan_to_mar = FALSE,
  quarter_col_name = NA,
  month_pattern = NA
)
}
\arguments{
\item{dat}{dataframe}

\item{year_from_pattern}{character string. Regular expression that matches
the name of the column containing year. This column may contain extra
information alongside year, such as quarter. Required unless fy_from_fy_end is provided.}

\item{fy_from_fy_end}{boolean. Default is FALSE. TRUE if the financial
year is to be calculated from the financial year end.fy_end_pattern must be
supplied if TRUE.}

\item{fy_start_from_fy_end}{boolean. Default is FALSE. TRUE if the financial
year start is to be calculated from the financial year end.}

\item{fy_end_pattern}{characters string. Regular expression that matches the
name of the column containing the financial year end. Default is NA}

\item{calendar_to_fy_start}{boolean. TRUE if calendar year is to be converted
to financial year start. Default is FALSE.}

\item{q1_is_jan_to_mar}{boolean. TRUE if Jan to Mar is Q1. FALSE if Apr to
Jun is Q1. Default is FALSE.}

\item{quarter_col_name}{character string. Exact name of the column that
contains only the quarter number ('1', '2', '3', '4', or NA).}

\item{month_pattern}{character string. Regular expression used to
identify the column that contains only the month.}
}
\value{
dataframe. Dat with fy_start added as a column. Depending on how
fy_start is ascertained, other columns may also be created ('year' and
'quarter').
}
\description{
If financial year start year is not already available, extract
it from the available information e.g. from financial year, fy end year, or
calendar year and quarter/month (where Q1 is either Jan to March or Apr to
Jun).
}
\examples{
\dontrun{
# 1. Only financial year end is available.
dat <- data.frame(
    fy_end = c(2021, 2022),
    value = c(1, 2)
    )
get_fy_start(
    dat, fy_from_fy_end = TRUE, fy_start_from_fy_end = TRUE,
    fy_end_pattern = "fy"
    )

# 2. Only calendar year and month are available in separate columns.
dat <- data.frame(
    year = c("2021", "2021"),
    month = c("jan-mar", "apr-jun"),
    value = c(1, 2)
    )
get_fy_start(
    dat, year_from_pattern = "year", month_pattern = "month",
    calendar_to_fy_start = TRUE,  q1_is_jan_to_mar = FALSE,
    )
}
}
