% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fy_start.R
\name{get_fy_start_from_calendar_year}
\alias{get_fy_start_from_calendar_year}
\title{Get financial year start for quarters in datasets with Q1 as Apr-Jun}
\usage{
get_fy_start_from_calendar_year(
  dat,
  calendar_year_to_fy_start = FALSE,
  q1_is_jan_to_mar = TRUE,
  year_from_pattern = "(?i)year",
  month_col_pattern = NA,
  quarter_col_name = NA,
  fy_start_preference = NA
)
}
\arguments{
\item{dat}{dataframe containing a year column and either a month or quarter
column}

\item{calendar_year_to_fy_start}{boolean. Default is FALSE. TRUE if the
action is to be performed, NA or FALSE to return the original data.}

\item{q1_is_jan_to_mar}{boolean. Default is TRUE and means that
Q1 is for Jan to Mar. Use FALSE when Q1 is for Apr to Jun i.e. the first
quarter of the financial year.}

\item{year_from_pattern}{character string. Defaults to "(?i)year". Regular
expression matched by the column name for the year column. Must only match
one column name. This column must be in the format YYYY NOT YYYY-YY or YY.
Non-numeric information can be present as it will be removed (i.e. '2024 -r'
will be seen as '2024'). There must be no numbers present other than the
year. Column can be character or integer.}

\item{month_col_pattern}{character string. Same as year_from_pattern but for
month. Only required if quarter_col_name is not given.}

\item{quarter_col_name}{character string. Name of the coloumn containing the
quarter number. Only required if month_col_pattern is not given.}

\item{fy_start_preference}{string, optional. Must be "quarter" or "month". If
the year is given in both the month and the quarter column, and they do not
match, this variable is used to decide between them. If they do not match and
this is not supplied, NA will be given for fy_start.}
}
\value{
dat with fy_start (character string) column if
calendar_year_to_fy_start is TRUE, or the original dat if not.
}
\description{
Get financial year (FY) start for quarters. This is calculated
differently when year is given as calendar year but Q1 is Apr-Jun and when
Q1 is Jan-May. In the former, financial year start will equal the calendar
year for Q1-Q3 but Q4/Jan-Mar financial year start will be year minus one.
}
\details{
Where the output of tidysheet is used for calculations on a financial year
basis, but the source data was published on a calendar year basis (where Q1
is the first quarter of the calendar year), financial year may not be given.
This function calculates financial year start from the calendar year and
either the quarter or the month.

See make_calendar_q1_month_and_quarter_patterns() for the accepted
presentation of months and quarters (e.g. 'jan'/'Jan-Mar'/'Q1'/'Quarter 1').

For 2021, jan-mar, Q1, Financial year will be 2020-21 <- fy start == year-1.
For 2021, apr-jun, Q1, Financial year will be 2021-22 <- fy start == year.
For 2021, apr-dec, Q1-Q3, Financial year will be 2021-22 <- fy start == year.
}
\examples{
\dontrun{
# q1_is_jan_to_mar is false:
dat <- tibble(
  Year = c("2021","2021", "2021", "2021", "2022"),
  Quarter = c("Q4", "Q1", "Q2", "Q3", "Q4"),
  Months = c("jan - mar", "apr - jun", "jul - sep", "oct - dec", "jan - mar"),
  Value = 1:5
)
get_fy_start_from_calendar_year(
    dat, calendar_year_to_fy_start = TRUE,
    q1_is_jan_to_mar = FALSE,
    year_from_pattern = "(?i)year",
    quarter_col_name = "Quarter",
    month_col_pattern = NA
    )

# q1_is_jan_to_mar is true:
dat <- tibble(
  Year = c("2021","2021", "2021", "2021", "2022"),
  Quarter = c("Q1", "Q2", "Q3", "Q4", "Q1"),
  Month = c("jan - mar", "apr - jun", "jul - sep", "oct - dec", "jan - mar"),
  Value = 1:5
)
get_fy_start_from_calendar_year(
    dat, calendar_year_to_fy_start = TRUE,
    q1_is_jan_to_mar = FALSE,
    year_from_pattern = "(?i)year",
    quarter_col_name = "Quarter"
    )

# month is given for some rows, quarter is given for others, and some have both:
dat <- tibble(
  Year = c("2021", "2022", "2022", "2022", "2022"),
  Quarter = c("Q3", "Q4", NA, "Q1", NA),
  Months = c("oct - dec", "jan", "jan - mar", NA, "Jul"),
  Value = 1:5
)
get_fy_start_from_calendar_year(
    dat, calendar_year_to_fy_start = TRUE,
    q1_is_jan_to_mar = FALSE,
    year_from_pattern = "(?i)year",
    quarter_col_name = "Quarter",
    month_col_pattern = "(?i)month"
    )
}
}
