% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fy_start.R
\name{get_fy_start_from_fy_end}
\alias{get_fy_start_from_fy_end}
\title{get financial year start year from financial year end year}
\usage{
get_fy_start_from_fy_end(dat, end_year)
}
\arguments{
\item{dat}{dataframe}

\item{end_year}{character string. The name of the column containing fy_end
information in the format YYYY.}
}
\value{
dataframe dat with additional column 'fy_start' and with invalid
year information removed from the end_year column. If there are any entries
in the end_year column that do not contain a valid year, fy_start is given
as NA for those rows, and a warning is raised.
}
\description{
Get the financial year start year from a column containing
the financial year end year as YYYY.  (fy_start = fy_end - 1).
}
\examples{
\dontrun{
dat <- data.frame(fy_end = c("2000", "this 2020", "blue"))
get_fy_start_from_fy_end(dat, "fy_end")
}
}
