% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_header_identifier.R
\name{get_header_identifier}
\alias{get_header_identifier}
\title{get_header_identifier}
\usage{
get_header_identifier(header_identifier, file_part)
}
\arguments{
\item{file_part}{(string): An integer/number cannot be a float or 0}

\item{header_identifiers}{: A string vector}
}
\value{
header_identifier : A string selected from the header_identifier string vector
}
\description{
Get header_identifier from header_identifiers string vector
}
\details{
Get header_identifier from header_identifiers string vector

Function to identify the correct header_identifier for preprocessing by using the file_part variable
to subset the string vector (header_identifiers) and return the correct header_identifier respective to file_part.
}
\examples{
header_identifiers <- c("total", "HRA Reserves")
file_part <- 1
result <- get_header_identifier(header_identifiers, file_part)
print(result)

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}
}
