% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_information.R
\name{get_header_row}
\alias{get_header_row}
\title{Get the row index of a header based on specified identifiers}
\usage{
get_header_row(dat, pattern, instance = 1, offset_by = NA)
}
\arguments{
\item{dat}{A data table of the Excel file, in xlsx_cells format.}

\item{pattern}{A string pattern to search for in the data to identify the
header row. In pub sec this variable is specified with header_identifier.}

\item{instance}{An integer specifying which instance of the header identifier
pattern to use. Defaults to 1 (the first instance). In pub sec this variable
is specified with header_identifier_instance.}

\item{offset_by}{An integer specifying the number of rows to offset
the header row. Defaults to 0. In pub sec this variable is specified with
header_row_offset. If the row the pattern is found on is the row after the
first header row, offset_by will be 1. If the pattern is on the row before
the first header row, offset_by will be -1.}
}
\value{
The row index of the header based on the specified pattern and
instance. Returns \code{NA} if no match is found. Returns the number of the first
populated row if no pattern is provided.
}
\description{
Identify the row number of the first header in the dataset based on
a specified regular expression, desired instance of the match (1st, 2nd etc),
and the header row offset. If the header row offset is not specified, the
header row is assumed to be the row on which the specified instance of the
match occurs. However if the matching cell is known to be found on e.g. the
row after the first header row, offset_by can be used.

If the specified instance or row is invalid, warnings are issued, and the
function defaults to the first valid instance.
}
\examples{
\dontrun{
|row | data title         |                   |
|----|--------------------|-------------------|
|2   | million            |                   |
|3   |                    |                   |
|4   | education services | highways services |
|5   | primary services   | roads             |
|6   |      50            |      10           |

first header row in the above table is row 4,
first row of data is row 6

dat <- data.frame("character" = c("title", "million", NA,
                                 "Education services", "Highways services",
                                 "primary services", "roads",
                                 NA, NA),
                  "row" = c(1:3, 4, 4, 5, 5, 6, 6),
                  "numeric" = c(rep(NA, times = 7), 50, 10))

get_header_row(dat, "services", 1, 0, 1)
}
}
