% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_info_above_table}
\alias{get_info_above_table}
\title{Get the character string information from above the data}
\usage{
get_info_above_table(dat, row_num)
}
\arguments{
\item{dat}{dataframe imported using tidyxl::xlsx_cells().}

\item{row_num}{The row number of the first row that is to be excluded.}
}
\value{
dataframe in an xlsx_cells layout containing everything from dat that
is in rows prior to row_num.
}
\description{
Get character information from rows of the data that are above
the provided row number. If data_type is anything other than character the
row is removed.
}
\examples{
\dontrun{
dat <- data.frame(
    row = as.numeric(1:4),
    col = 1,
    character = c("title", NA, "first_header", NA),
    numeric = c(NA, 1995, NA, 1000)
    )
get_info_above_table(dat, 3)
}
}
