% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_release_number.R
\name{get_latest_release_number}
\alias{get_latest_release_number}
\title{get_latest_release_number}
\usage{
get_latest_release_number(release_numbers)
}
\arguments{
\item{release_numbers}{(character vector). These can be either 1st, 2nd etc, or first, second etc}
}
\value{
latest_release_number (string)
}
\description{
get_latest_release_number - Get the vintage from the release number
}
\details{
get_latest_release_number - Get the vintage from the release number
 
Use the highest release number to determine vintage. This works regardless of
whether e.g. "first", "1st" or a mix of these is used.

If more than one vintage was found, the user is given a warning with the other 
releases found

If the highest release_number is 1st, or first, vintage is "provisional"
If the highest release_number is 2nd, or second, or higher, vintage is "final"
}
\examples{
release_numbers <- c("2nd release", "first release")
get_latest_release_number(release_numbers)

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
