% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layout_type.R
\name{get_layout_type}
\alias{get_layout_type}
\title{get_layout_type}
\usage{
get_layout_type(year, layout_dict, element = 1)
}
\arguments{
\item{year}{character string e.g. 2021}

\item{layout_dict}{dataframe with key and item columns. annual/Financial year
is given in the key column}
}
\value{
The item from layout_dict for the year in question
}
\description{
Get the layout type from the data dictionary
}
\details{
Get the layout type from the data dictionary

Find which layout type the file is (given the year).
}
\examples{
layout_dict <- data.frame("key" = c("from 2019-20", "from 2020-21", "from 2020-21"),
                           item = c("single_header", "multi-header", "single_header"))
layout <- get_layout_type(2021, layout_dict)
print(layout)

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
