% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot_data.R
\name{get_left_col_names}
\alias{get_left_col_names}
\title{Get the column names for the left block.}
\usage{
get_left_col_names(
  dat,
  first_right_header_col,
  left_columns_to_remove,
  left_headers,
  first_data_row
)
}
\arguments{
\item{dat}{dataframe. Must be xlsx_cells format.}

\item{first_right_header_col}{int. The number of the first column of data.}

\item{left_columns_to_remove}{character vector. One regular expression for
each column inthe left header block that should be removed before assigning
left header names. If the pattern matches a single column, it will be
removed.}

\item{left_headers}{character vector. Can be NA (see rules above).}

\item{first_data_row}{integer. The row number of the first line of data
(below the headers).}
}
\value{
character vector with names of columns in the left block of data.
}
\description{
Get the names of the columns to the left of the numeric columns
(referred to here as the left block).

Sometimes columns in the left block are all named in the raw
data, sometimes none of them are named, and sometimes only some are, with
others left blank.

We have the option to provide names for columns in the left header block in
the data dictionary (left_headers), however this is risky as suppliers could
change the number or order of those columns.

A third option is to use placeholder names (e.g. 'column_1', 'column_2' etc)

Rules:
\itemize{
\item If there is the correct number of names in the data (existing names) use
these, even if names have been given in the data dictionary (left_headers).
\item If the number of names given in the dict is too many do not use any of them.
\item If there are names missing from the existing names, fill them with
placeholder names, but use the ones that are given.
\item Use all of left_headers if no names already exist in the data AND the
correct number of names has been supplied.
}
}
\examples{
\dontrun{

dat <- data.frame(
    row = c(1, 1, 1, 2, 2, 2),
    col = c(1, 2, 3, 1, 2, 3),
    data_type = c(rep("blank", 2), rep("character", 3), "numeric"),
    character = c(NA, NA, "right", "a", "b", NA)
    )

get_left_col_names(dat, 3, c("alt 1", "alt 2"), 2)
}
}
