% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_information.R
\name{get_loc_from_instance}
\alias{get_loc_from_instance}
\title{Get row or column number from a vector of possibilities}
\usage{
get_loc_from_instance(possibilities, instance, offset_by)
}
\arguments{
\item{possibilities}{integer vector. The row or column numbers on which a
pattern match was found. These must be in numeric order.}

\item{instance}{integer. The index of the intended pattern match.}

\item{offset_by}{integer. The number to subtract from the intended number in
possibilities.}
}
\value{
integer.
}
\description{
Given a vector of row or column numbers in which a match to a
given pattern was found, use a provided index (instance) and/ or an offset
value to find the desired row or column number.
}
\examples{
\dontrun{
get_loc_from_instance(c(3, 6, 9), 2, -1)
}
}
