% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_information.R
\name{get_matching_colnames}
\alias{get_matching_colnames}
\title{Get column names that contain a pattern}
\usage{
get_matching_colnames(dat, pattern)
}
\arguments{
\item{dat}{dataframe}

\item{pattern}{string. Regular expression you want to match}
}
\value{
vector of character strings. Names of the columns that match the
given pattern. If no names are found a character vector of length 0 is
returned.
}
\description{
Return all column names from a dataset that match a regular
expression.
}
\details{
See get_colnames_from_pattern for a similar function.
}
\examples{
\dontrun{
dat <- data.frame(fin_year_end = "2020", "Year" = ("2021"), "other" = NA)
get_matching_colnames(dat, "(?i)year")
}
}
