% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_information.R
\name{get_metadata}
\alias{get_metadata}
\title{Extract metadata from a dataframe and from other supplied information.}
\usage{
get_metadata(
  sheet_dat,
  dropdown_pattern = NA,
  single_vintage = NA,
  release_number = NA,
  table_dat = NA,
  sheet_title = NA,
  sheet_units = NA
)
}
\arguments{
\item{sheet_dat}{dataframe in xlsx_cells format, containing metadata from
the top of the sheet.}

\item{dropdown_pattern}{character string. regular expression to match the
expected contents of the dropdown cell. In pub sec this variable is specified
by dropdown_pattern.}

\item{single_vintage}{character string. Vintage as specified in the settings.
In pub sec this variable is specified by single_vintage.}

\item{release_number}{integer between 1 and 9 or NA.}

\item{table_dat}{dataframe in xlsx_cells format, containing metadata from
above the table (i.e. after the first header row, but before the table first
header row). Only relevant for datasets with multiple tables in a sheet.}

\item{sheet_title}{character string. The title at the top of the sheet.}

\item{sheet_units}{character string. Units taken from the top of the sheet.}
}
\value{
named list giving character string values for 'units', 'dropdown'
'title', 'year', and 'vintage'. 'year' can be a vector with more than one
element. All others have a single element.
}
\description{
Get units, title, dropdown cell contents, year, and vintage
values from metadata held above the data.
}
\details{
This is used in public sector to look for the relevant information at the top
of the sheet and above the subtables. Where there is conflicting title,
units, or vintage information between the sheet and subtable metadata, rules
are implemented by the called functions to select the correct one.

Conflicts between the year found in sheet and subtable metadata are handled
outside this function (by get_year_for_use_in_data) as it is more complex.

Title and units are simply extracted from the information above the data.
Vintage and dropdown values also use information supplied from the user
settings and other places (i.e. release number).

Note on the dropdown cell value: In some DLUHC (MHCLG) datasets we have to
use the LA_dropdown tab. In these sheets there is a dropdown at the top that
controls what data are shown on the sheet, and therefore what data are
imported. If the wrong value is selected in the dropdown, users need to be
made aware of this, hence including it in the metadata.
}
\examples{
\dontrun{
sheet_dat <- data.frame(
    character = c(
        "All data for 2022-23 to 2026-27",
        "£ million unless otherwise stated",
        "England"
        ),
    address = c("A1", "A2", "A4"),
    row = c(1, 2, 4),
    col = 1
    )
table_dat <- data.frame(
    character = c("Data for 2024-25", "£ thousand"),
    address = "A40",
    row = 40,
    col = 1
    )
dropdown_pattern <- "(?i)Eng"
single_vintage <- "final"

sheet_metadata <- get_metadata(
    sheet_dat, dropdown_pattern, single_vintage
    )
table_metadata <- get_metadata(
    sheet_dat, dropdown_pattern, single_vintage, NA, table_dat,
    sheet_metadata[['title']], sheet_metadata[['units']]
    )
}
}
