% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_to_multiple_columns.R
\name{get_new_string}
\alias{get_new_string}
\title{Create a column containing the newly split off string.}
\usage{
get_new_string(
  dat,
  split_patterns,
  current_string_col,
  prev_string_col,
  use_split_point,
  current_split
)
}
\arguments{
\item{dat}{Dataframe output containing columns prefixed with
'tmp_use_split_point_', 'tmp_split_', 'tmp_matches_next_pattern_' and
'tmp_whitespace_'.}

\item{split_patterns}{character array of regex patterns.}

\item{current_string_col}{character. Name of column with prefix 'tmp_string_'
This column will be filled with the new (split) string.}

\item{prev_string_col}{character. Name of the column containing the previous
string split. If no splits have yet been performed this will be
'tmp_remaining'}

\item{use_split_point}{character. Name of the boolean column with prefix
'tmp_use_split_point_'}

\item{current_split}{character. Name of the integer column with prefix
'tmp_split_' containing the index of the split_patterns pattern to be used.}
}
\value{
dat with new columns containing the split string, with the prefix
'tmp_string_'.
}
\description{
Each newly split off string is put in a column prefixed with
'tmp_string_'. Use current_split to know which split_patterns pattern to use
(it is an index).
\itemize{
\item If there are no split_point pattern matches in an item, all 'tmp_string_'
columns will equal the entry in the original column (which has been copied to
'tmp_remaining' by this point).
\item If current_split is NA, there are no splits left to do so the last string
is carried forward.
\item If use_split_point is FALSE, the regex pattern in split_patterns does not
exist, so the last string is also carried forward in this situation.
}
}
\examples{
\dontrun{
split_patterns <- c( "(\n)\\\\s*1\\\\.", "(\n)\\\\s*2\\\\.", "(\n)\\\\s*3\\\\.")
current_string_col <- "tmp_string_1"
use_split_point <- "tmp_use_split_point_0"
current_split <- "tmp_split_1"
matches_next_pattern <- "tmp_matches_next_pattern_1"
whitespace <- "tmp_whitespace_1"

dat <- data.frame(
  tmp_remaining = "split\n 1.this \n 2.one \n 3.thousand",
  tmp_use_split_point_0 = TRUE,
  tmp_split_1 = 1,
  tmp_matches_next_pattern_1 = FALSE,
  tmp_whitespace_1 = TRUE,
  value = 100
)

get_new_string(
   dat, split_patterns, current_string_col, "tmp_remaining",
   use_split_point, current_split
   )
}
}
