% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot_data.R
\name{get_outer_descriptor_col_loc}
\alias{get_outer_descriptor_col_loc}
\title{Get the column number of an outer right descriptor column.}
\usage{
get_outer_descriptor_col_loc(
  dat,
  header_row_count,
  first_right_header_col,
  first_header_row,
  name_regex
)
}
\arguments{
\item{dat}{xlsx_cells data.}

\item{header_row_count}{integer. the number of header rows.}

\item{first_right_header_col}{the column number of the first column of the
numeric data block.}

\item{first_header_row}{integer. The row number of the first header.}

\item{name_regex}{Character string. Regular expression. Only one can be
given. If more are needed the function will need to be updated. In pub sec
this variable is specified using column_to_right_of_data_name_pattern.}
}
\value{
integer. The column number or NA (see description for detail).
}
\description{
Some datasets have the following structure:
Descriptor columns e.g. Year > Numeric columns > Descriptor e.g. notes

In such cases where the column to the right of the data block is a character
column, if it is not dealt with separately the information in it will be lost
when the right block is beheaded (unpivotted). This is because the
unpivotting only keeps numeric data. We can choose to either delete such
columns using settings in sheet_structure, or (in pub sec) use this function
to keep them by specifying column_to_right_of_data_name_pattern
(name_regex) in settings.

This function returns the column number of a descriptor column to the right
of the data. However, it will only return the column number if it fulfils
the following conditions:
\itemize{
\item name_regex is not NA
\item There is wording within the header rows that matches the name_regex.
\item The string that matches name_regex is found within the LAST 3 columns
that have any information in the header rows.
}

If any of these conditions are not met, NA will be returned.

Note: Currently we only have an issue with a single non-numeric descriptor
column being to the right of the data (e.g. DLUHC CPR3 year and quarter
series 2023-24). Future improvements are required to allow multiple columns.
}
\examples{
\dontrun{
dat <- data.frame(
address = c("A1", "B1", "C1", "A2", "B2", "C2", "A3", "B3", "C3"),
row = rep(1:3, each = 3),
col = rep(1:3, times = 3),
character = c(NA, "education", "Notes",
              "Year", "primary", "(see frontsheet)",
              "2020", NA, "note 1"),
numeric = c(NA, NA, NA, NA, NA, NA, NA, 100, NA)
)
get_outer_descriptor_col_loc(
                   dat,
                   header_row_count = 2,
                   first_right_header_col = 2,
                   first_header_row = 1,
                   name_regex = "(?i)note"
                   )
}
}
