% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sheet.R
\name{get_pattern_names}
\alias{get_pattern_names}
\title{Get the names of variables that can be regular expressions}
\usage{
get_pattern_names()
}
\value{
character string vector of variable names.
}
\description{
So that we can convert pattern strings to regular expressions
for the correct variables (see build_regex).
Note that split_points is intentionally not included because the pattern
descriptions used to build the regular expression are themselves required
in the split_to_multiple_columns functions.
}
