% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_information.R
\name{get_release_number}
\alias{get_release_number}
\title{Get the release number from an xlsx_cells dataframe}
\usage{
get_release_number(dat)
}
\arguments{
\item{dat}{character string vector.}
}
\value{
integer. Version number between 1 and 9. If multiple possibilities
are found the highest number is returned with a warning. If none are found,
or id dat is NULL NA is returned with a warning.
}
\description{
Get the release number (first, 1st, second, 2nd etc) from the
character column of a dataframe imported using tidyxl::xlsx_cells. Written
for use with DLUHC (MHCLG) Revenue expenditure provisional and final data.
}
\examples{
\dontrun{
dat <- data.frame(
    character = c(
       "title - First release",
       "This is release 1 not release 2",
       "next will be release 25"
     ))
 get_release_number(dat)
}
}
