% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_release_number_vintage.R
\name{get_release_number_vintage}
\alias{get_release_number_vintage}
\title{get_release_number_vintage}
\usage{
get_release_number_vintage(front_page_sheet)
}
\arguments{
\item{character_matches:}{A dataframe containing a column called 'character' that contains the release information.}
}
\value{
string: the vintage or NA
}
\description{
get_release_number_vintage_version - Get the version from a release number in a tibble/dataframe
}
\details{
get_release_number_vintage_version - Get the version from a release number in a tibble/dataframe

Get the release number (first, 1st, second, 2nd etc) from the column of a 
dataframe or tibble called 'character'. Convert this into  vintage (either 
'provisional' when 1st release, or 'final' when anything later than first).
}
\examples{
character_matches <- c(NA, "character(0)", 
                     "Data from this workbook have been used to compile 'LA 2022-23 Outturn, England - Second release' .",
                     "notes on first release")
front_page_sheet <- data.frame(character = character_matches)
get_release_number_vintage(character_matches)

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
