% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot_data.R
\name{get_right_block}
\alias{get_right_block}
\title{Only retain rows that relate to numeric columns in the Excel data}
\usage{
get_right_block(dat, first_right_header_col, excluded)
}
\arguments{
\item{dat}{dataframe imported using tidyxl::xlsx_cells.}

\item{first_right_header_col}{integer. The first column number of the numeric
data block (aka the right block).}

\item{excluded}{integer. The column number of the column of descriptors to
the right of the data, or NA.}
}
\value{
dat with rows that relate to non-numeric columns in the excel data
removed.
}
\description{
Remove the left block of row descriptors and, if it exists, the
column of descriptors to the right of the data.
}
\examples{
\dontrun{
dat <- data.frame(
  address = c("A1", "B1", "C1", "D1", "A2", "B2", "C2", "D2"),
  row = c(1, 1, 1, 1, 2, 2, 2, 2),
  col = c(1, 2, 3, 4, 1, 2, 3, 4),
  data_type = c(rep("character", 5), rep("numeric", 2), "character"),
  numeric = c(rep(NA, 5), 100, 22, NA),
  character = c("Year", "primary", "secondary", "Notes",
                 "2021", NA, NA, "note a"))

get_right_block(dat, 2, 4)
}
}
