% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_info.R
\name{get_row_headers_colname}
\alias{get_row_headers_colname}
\title{Get the column name of the column containing row headers/totals}
\usage{
get_row_headers_colname(possibilities, pattern)
}
\arguments{
\item{possibilities}{character string vector. The names of possible columns.
This can be all column names, or a subset in which the match is expected.}

\item{pattern}{character string. A regular expression to match a column name.}
}
\value{
The name of the column matching the pattern. If more than one match
is found, the first is returned with a warning. If none are found NA is
returned with a warning. If pattern is NA, NA is returned.
}
\description{
Get the column name that matches the provided pattern for
the column containing row headers. If there is no match, the first possible
column name is returned.
}
\details{
The wording used in different releases of the same data may
change slightly. As exact column names are therefore unlikely to be known for
descriptor columns in the source data, a pattern is provided instead. This
function is specifcally for use in add_row_headers_column and
add_totals_as_column: if there is no match, the first possible column name
is returned. The messages are also specific to add_row_headers_pattern

See get_matching_colnames and get_colnames_from_pattern for more general
functions with a similar purpose.
}
\examples{
\dontrun{
get_row_headers_colname(c("column 1", "column 2"), ".*2")
}
}
