% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_source_data.R
\name{get_source_data}
\alias{get_source_data}
\title{get_source_data}
\usage{
get_source_data(sheets, tab_regex_front_page)
}
\description{
get_source_data
}
\examples{
get_source_data()
}
\author{
Mark London \email{mark.london@ext.ons.gov.uk}
}
