% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tables.R
\name{get_split_loc}
\alias{get_split_loc}
\title{Get the column number for the start of each table}
\usage{
get_split_loc(dat, pattern, instance, offset_by, direction)
}
\arguments{
\item{dat}{dataframe imported using tidyxl::xlsx_cells().}

\item{pattern}{character string. Regular expression that matches a string
in the first row or column after where the split needs to happen.}

\item{instance}{integer. The instance of the pattern match (1st, 2nd etc).}

\item{offset_by}{integer. The number of rows or columns to subtract from the
row/column on which the pattern match was found.}

\item{direction}{character string. Either 'col' if the table is to be split
by column, or 'row' if it is to be split by row.}
}
\value{
integer indicating the column or row at which the second table
starts.
}
\description{
Get the number of the row or column that will come immediately
after the split (i.e. the first row or col of the 2nd table). If tables are
stacked one above the other, the split will be done by row and the split loc
will be a row number. If the tables are side-by-side, the split will be done
by column (col) and the split loc will be a column number.
}
\examples{
\dontrun{
dat <- data.frame(
  "row" = c(1, 1, 1, 2, 2, 2),
  "col" = c(1, 2, 3, 1, 2, 3),
  "address" = c("A1", "B1", "C1", "A2", "B2", "C2"),
  "character"= c(NA,NA,NA,"Education Services","NA","NA")
)
get_split_loc(dat, '(?i)education', 'col')
}
}
