% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_to_multiple_columns.R
\name{get_split_order}
\alias{get_split_order}
\title{Give the order in which split_point patterns should be used per item}
\usage{
get_split_order(dat)
}
\arguments{
\item{dat}{Dataframe containing boolean columns prefixed by
'use_split_point_'}
}
\value{
dat with new columns prefixed by "split_"
}
\description{
Not all items will contain all split_patterns patterns so the
first split may for example require different patterns to be used for
different items. Some (or all) split_point patterns can be skipped but they
cannot be done out of order. For example the first split for an item can be
the 2nd element of split_patterns (rather than the first), and the third
split can be the 4th element of split_patterns. However, if the first split
is the 2nd element of split_patterns the second split CANNOT be the 1st
because this doesn't exist once the 2nd element has been used).

This function uses the boolean columns created by flag_existing_split_points
to return an integer column for each split. See split_to_multiple_columns for
the desired outcome when an item does not contain all split_point patterns.
}
\examples{
\dontrun{
dat <- data.frame(
  col_to_split = c("split\n 1.this \n 2.one \n 3.thousand", "split\n 2.one"),
  'tmp_use_split_point_1' = c(TRUE, FALSE),
  'tmp_use_split_point_2' = c(TRUE, TRUE),
  'tmp_use_split_point_3' = c(TRUE, FALSE),
  value = c(100, 600)
)

get_split_order(dat)
}
}
