% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_information.R
\name{get_subtitles}
\alias{get_subtitles}
\title{Record subtable titles and their location}
\usage{
get_subtitles(dat, column, patterns, row_offset, horizontal_index)
}
\arguments{
\item{dat}{dataframe imported using xlsx_cells and not yet pivotted i.e. all
character strings are still in the character column.}

\item{column}{character string. The name of the new column to store the
subtable title. In pub sec this variable is specified with
subtable_title_column.}

\item{patterns}{character string vector. Any number of regular expressions.
The row number that is identified will be the one(s) that have a match to
EVERY pattern provided. In pub sec this variable is specified with
subtable_title_patterns.}

\item{row_offset}{integer. The number to add to the row that matches
patterns to get the row number on which the subtitle is found, e.g if the
row below the found pattern contains the subtitle, use 1. if the
row above the found pattern contains the subtitle, use -1.}

\item{horizontal_index}{character string. An index number where 1 is the first
populated cell, and -1 is the last.}
}
\value{
dataframe with a column for row and a a column named using the
'column' arg containing the subtitles.
}
\description{
Use regular expressions to find specific rows. The row that you
want to locate must contain all the provided regular expressions. Once the
row is identified, the row containing the subtitles can be identified through
its relative position to the found row. If it is known to be two rows above,
use offset = -2. Ideally, the patterns provided will match the subtable title
rows, in which case offset is 0 and does not need to be specified. Once the
target row is identified, if there is more than one character cell in that
row, the index of the required cell can be specified using horizontal_index.
}
\examples{
\dontrun{
dat <- data.frame(
    row = rep(c(1:5), each = 3),
    col = rep(c(1:3), times = 5),
    character =
        c("million", NA, "table A",
          NA, "col A", "col B",
          "c", NA, NA,
          NA, NA, "table B",
          NA, "col A", "col B"
          ),
    numeric = c(rep(NA, 7), 1, 2, rep(NA, 6)),
    is_blank = c(FALSE, TRUE, FALSE, TRUE, rep(FALSE, 5),
                 rep(c(TRUE, TRUE, FALSE), 2))
    )
get_subtitles(dat, "subtitle", c("(?i)a", "(?i)col.*b"), -1, -1)
}
}
