% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_table_data}
\alias{get_table_data}
\title{Get the table data (excluding information above the table)}
\usage{
get_table_data(dat, row_num)
}
\arguments{
\item{dat}{dataframe imported using tidyxl::xlsx_cells().}

\item{row_num}{The row number of the first row that is to be included. Any
rows with a smaller row number will be removed.}
}
\value{
dataframe. dat with rows prior to row_num removed.
}
\description{
Retain only data that is part of the table itself.
}
\examples{
\dontrun{
dat <- data.frame(
    row = as.numeric(1:4),
    col = 1,
    character = c("title", NA, "first_header", NA),
    numeric = c(NA, 1995, NA, 1000)
    )
get_table_data(dat, 3)
}

}
