% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tables.R
\name{get_tables_as_list}
\alias{get_tables_as_list}
\title{Split xlsx_cells data into a list of tables}
\usage{
get_tables_as_list(
  dat,
  directions,
  table_ids,
  patterns,
  pattern_instances,
  pattern_offsets,
  keep = NA
)
}
\arguments{
\item{dat}{dataframe imported using tidyxl::xlsx_tables()}

\item{directions}{character string vector the same length as the number of
splits required. Must only contain 'col' and 'row'. In pub sec this variable
is specified using table_split_dirs.}

\item{table_ids}{integer vector. The full table is id 1. Table 1 will be
split into table 2 and 3. If split direction is col, table 2 is to the left,
table 3 to the right. If split direction is row, table 2 is at the top, table
3 below. Length of vector is equal to the number of splits required. In pub
sec this variable is specified using tables_to_split.}

\item{patterns}{character string vector. Regular expression patterns that
match an entry in the first column or row of where the table in question will
be split.  In pub sec this variable is specified using split_patterns.}

\item{pattern_instances}{character or integer vector. If characters they must
be convertible to integers. If the split location is at the second occurrence
of the pattern, the pattern_instance would be 2. Optional, but if specified
for any split, it must be specified for all.}

\item{pattern_offsets}{character or integer vector. If characters they must
be convertible to integers. If the split location is the row or column after
the occurrence of the pattern, the pattern_offset would be -1. If it is the
row/column before it would be 1. Optional, but if specified for any split,
it must be specified for all.}

\item{keep}{integer vector. The IDs of the tables to keep. Default is NA, in
which case all tables are kept. In pub sec this variable is specified by
tables_to_process.}
}
\value{
dat split into a list of tables.
}
\description{
Split a dataframe imported using tidyxl::xlsx_cells into a
list of tables.
}
\details{
The dataframe can be split horizontally or vertically or
any combination of those. Splits are done iteratively, with each split
resulting in two new tables. The number of times the table is split is
equal to the number of elements passed in the variables.
}
\examples{
\dontrun{
# 3 tables - one above the other on the left, and an overlapping one on the
# right.
# The first split has to be vertical (as table c overlaps the two tables
# stacked to the left of it), so the first complete table is table c.
# The second split is horizontal, and tables are returned top to bottom so
# the second table is table a and the third is table b.

dat <- data.frame(
    row = c(1, 2, 2, 3, 3, 4, 4, 4, 5, 5, 6, 6, 6, 7, 7, 8, 8),
    col = c(1, 1, 2, 1, 2, 1, 2, 3, 3, 4, 1, 3, 4, 1, 2, 1, 2),
    character = c(
        "table a",
        "x", "value",
        "a1", NA,
        "a2", NA, "table c",
        "y", "value",
        "table b", "c1", NA,
        "z", "value",
        "b1", NA
        ),
    numeric = c(
        rep(NA, 4), 1, NA, 1, rep(NA, 5), 3, rep(NA, 3), 2
        )
    )
get_tables_as_list(dat, c("col", "row"), c(1, 2), c("table c", "table b"))
}
}
