% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sheet.R
\name{get_variable_names}
\alias{get_variable_names}
\title{Get the names of all possible arguments}
\usage{
get_variable_names()
}
\value{
character string vector of all arg names.
}
\description{
So that we can assign NA to any unspecified arguments, this
function provides a full list of all arguments.
}
