% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vintage_identifiers.R
\name{get_vintage}
\alias{get_vintage}
\title{get vintage (budget, provisional, or final)}
\usage{
get_vintage(
  single_vintage,
  release_number,
  sheet_title,
  sheet_info,
  table_title = NA,
  table_info = NA
)
}
\arguments{
\item{single_vintage}{character string. Vintage as specified in the settings.
In pub sec this variable is specified by single_vintage.}

\item{release_number}{integer between 1 and 9 or NA.}

\item{sheet_title}{character string. The title at the top of the sheet.}

\item{sheet_info}{dataframe imported using tidyxl::xlsx_cells()- this should
be a subset of the dataset that just includes information given above the
data in the sheet.}

\item{table_title}{character string. The title above the table.}

\item{table_info}{dataframe imported using tidyxl::xlsx_cells() Same as
sheet_info but for an individual subtable.}
}
\value{
character string. The highest priority vintage. Possible values are
"final", "budget", "provisional", or NA if no vintage can be determined. If
multiple different vintages are found a warning is given but the highest
priority vintage is still returned.
}
\description{
Get a single vintage (i.e. budget, provisional, or final) for
the dataset. Vintage can be found in different places depending on the
supplier and dataset. These options have an order of preference. The source
of vintage considered to be the most reliable is returned.

Note that 'budget' is used to describe both 'forecast' and 'estimated return'
data.

In order of priority:
\itemize{
\item as specified in the settings.
\item release number.
\item subtable title.
\item table title.
\item other information found in information above the data (table_info takes
priority over sheet_info).
}

We would usually expect all found vintages to be the same, so give a warning
if a mismatch is found (subtables may be different in which case the warning
could just be ignored).
}
\examples{
\dontrun{
# the following shows what happens when the sources of vintage disagree
# with each other
get_vintage(
  single_vintage = "final",
  release_number = 2,
  sheet_title = "provisional and final data",
  table_title = "final data",
  sheet_info = data.frame(
      character = "This sheet contains final and provisional data", row = 1
      ),
  table_info = data.frame(character = "provisional data", row = 1),
)
}
}
