% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vintage_from.R
\name{get_vintage_from}
\alias{get_vintage_from}
\title{get_vintage_from}
\usage{
get_vintage_from(vintage_check_item, sheet_vintage = NA)
}
\arguments{
\item{vintage_check_item}{string that could contain the vintage information}

\item{sheet_vintage}{string. Default is NA.}
}
\value{
vector of strings. The matched result or 'NA' if none found.
}
\description{
Gets a vintage as a string match
}
\details{
If sheet_vintage is NA, get_vintage_from checks items against a regex string
to find a match for the vintage string possibilities (budget, forecast,
provisional, final and estimated return).

Forecast and estimated return are other words for budget, so for example where forecast is picked up, return
'budget' for consistency.
}
\examples{
example code sets the sheet vintage to 'forecast', finding it in the patterns
  sheet_vintage <- "NA"
  vintage_check_item <- "forecast final item"
  expected_result <- "forecast"
  sheet_vintage <- get_vintage_from(vintage_check_item, sheet_vintage)

}
\author{
Mark London \email{mark.london@ext.ons.gov.uk}
}
