% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vintage_identifiers.R
\name{get_vintage_from_number}
\alias{get_vintage_from_number}
\title{Get the vintage from the release number}
\usage{
get_vintage_from_number(number)
}
\arguments{
\item{number}{integer between 1 and 9.}
}
\value{
character string. If number is 1, "provisional" is returned,
if higher than 1 "final" is returned.
}
\description{
Get the vintage from release numbers. If number is 1, vintage
is set to "provisional", if higher than 1 it is set to "final".
This is based on requirements for MHCLG (DLUHC) revenue expenditure
provisional and final data where the 1st release is always 1, and any number
higher than that is final until another release is made at which point those
lower number releases are relegated to a provisional status.
}
\examples{
\dontrun{
get_latest_release_number(1)
}
}
