% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vintage_identifiers.R
\name{get_vintage_from_string}
\alias{get_vintage_from_string}
\title{Get all valid vintages from a string}
\usage{
get_vintage_from_string(item)
}
\arguments{
\item{item}{string that could contain the vintage information}
}
\value{
vector of strings. The matched result or 'NA' if none or multiple
different vintages are found. If multiple vintages are found a warning is
given stating what they are.
}
\description{
Get vintages (i.e. budget, provisional, or final) from a string.
Vintage can be found in different places depending on the supplier and
dataset. This function can be used to get vintage from any string variables.

'budget' is used to describe both 'forecast' and 'estimated return' data.
}
\examples{
\dontrun{
sheet_vintage <- get_vintage_from_string("some forecast and final data")
}
}
