% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vintage_identifiers.R
\name{get_vintages_from_table}
\alias{get_vintages_from_table}
\title{Get vintage information from a small table imported using xlsx_cells}
\usage{
get_vintages_from_table(dat)
}
\arguments{
\item{dat}{dataframe imported using tidyxl::xlsx_cells. Must not contain
more than 20 unique row ids.}
}
\value{
character string. Possible values are "final", "budget",
"provisional", or NA if no vintage can be determined.
}
\description{
Get vintage information from a small table imported using
tidyxl::xlsx_cells(). Designed to handle the part of a sheet that sits
outside the main data i.e. the information above or below the data. As
such, if there are more than 2 unique Excel rows in the data NA and a
warning are returned.
}
\examples{
\dontrun{
dat <- data.frame(
    row = rep(1:2, each = 2),
    col = rep(1:2, times = 2),
    character = c("Title of the data", NA, "budget data", NA)
    )
get_vintages_from_table(dat)
}
}
