% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_handlers.R
\name{get_year}
\alias{get_year}
\title{Get first financial or calendar year from each entry in a given column}
\usage{
get_year(dat, from = NA, new_col = "year", prefer = "financial", type = "both")
}
\arguments{
\item{dat}{dataframe or vector. If a dataframe, it must contain a column with
the name from \code{from}.}

\item{from}{character string or NA. The name of the column containing year
information if dat is a dataframe, otherwise NA.}

\item{new_col}{character string. The name of the column the new information
will be stored in. Default is 'year'}

\item{prefer}{character string. Either 'financial' (default), or 'calendar'.
If type is 'both' and both a calendar year and a financial year are found in
a single entry, this determines which is chosen. If type is 'single', prefer
determines whether year represents calendar year or financial year.}

\item{type}{character string. Either 'both' (default), or 'single'. If
'single', the new year column will contain only calendar years or only
financial years (depending on prefer). If 'both', the year column can
contain both types.}
}
\value{
dat with a new column called yearif dat is a dataframe, or a vector
of years if input is a vector.
}
\description{
Get the first year from a specified column or vector and put it
in a new column called 'year' (can be changed using new_col arg).

If a string contains more than one year (of the same type), only the first is
put in 'year'.

If \code{prefer} and \code{type} are not specified, financial years will be returned
where available, but where no financial year is found, calendar year will be
returned.

If \code{type} = 'single' year will be restricted to either calendar or financial
years (whichever is set by prefer).
}
\examples{
\dontrun{
dat_fy <- data.frame(
  "year_and_vintage" = c("2019-20, 2020-21 budget", "this is 2020",
                         "2020-21 provisional", "2020 2021")
  )
dat_calendar_only <- data.frame(
  "year_and_vintage" = c("2019 budget", "this is 2020",
                         "2020 provisional", "2020 2021")
  )
dat_mixed <- data.frame(
  "year_and_vintage" = c("2019 and 2020-21 ", "this is 2020-21",
                         "2020 provisional", "2020 2021")
  )
get_year(dat_fy, "year_and_vintage")
get_year(dat_calendar_only, "year_and_vintage")
get_year(dat_mixed, "year_and_vintage")
get_year(c("a 2020", "c 1985"))
}
}
