% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fy_start.R
\name{get_year_col_to_drop_from_quarter}
\alias{get_year_col_to_drop_from_quarter}
\title{Get the year column name to be used by get_quarter}
\usage{
get_year_col_to_drop_from_quarter(dat, pattern)
}
\arguments{
\item{dat}{dataframe with a column name that matches pattern.}

\item{pattern}{character string. Regular expression that matches the column
containing year}
}
\value{
character string. The name of the year column
}
\description{
Get the name of the column that matches the year pattern and
contains a valid years.
}
\details{
If more than one year type is found, this function preferentially
returns calendar year, then the column that contains at least some calendar
years, then financial. This order is selected because the quarter column is
more likely to state a calendar year than financial in sources seen so far.
However, if the financial column is preferred over the calendar year, this
can be done by making the year_pattern more specific to that column name.
}
\examples{
\dontrun{
dat <- data.frame(
  Year = c("2023", "2024"),
  financial_year = c("2023-24", "2023-24"),
  quarter = c("2023Q4", "2024Q1"),
  Value = c(1, 2)
)
get_year_col_to_drop_from_quarter(dat, "(?i)year")
}
}
