% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_year_column.R
\name{get_year_column_names}
\alias{get_year_column_names}
\title{Get the column names of columns containing valid years.}
\usage{
get_year_column_names(dat, pattern)
}
\arguments{
\item{dat}{dataframe}

\item{pattern}{character string or NA. A regular expression, that, if
supplied and it matches a column name it will be the only column name
retained.}
}
\value{
named vector containing 'calendar', 'financial' and 'both'. Gives
the names of columns that contain year. If only calendar years are found the
column name is given under calendar, if only financial years are found the
name is given under financial, and if a mix is found the name is given under
mixed. If more than one column is found for any one of 'calendar',
'financial', or 'both', the one with the most matches is returned along with
a warning.
}
\description{
Get the names of columns that either match a supplied regular
expression or if this is not given, '(?i)year' and also contain valid years.
}
\examples{
\dontrun{
dat <- data.frame(
    financial_year = c("2012-2013", "2013-14 note", NA),
    mixed_year = c("2012-2013", "2013", "2014"),
    calendar_year = 2025,
    year_notes = c("", "something about 2021", ""),
    other = 2021
    )
get_year_column_names(dat, "year")
}
}
