% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fy_start.R
\name{get_year_from_column}
\alias{get_year_from_column}
\title{Extract the first year and save in a new column called year.}
\usage{
get_year_from_column(dat, pattern)
}
\arguments{
\item{dat}{dataframe.}

\item{pattern}{string. Regular expression that matches only the column name
of the column with strings containing year information. In pub sec this
variable is specified with column_pattern_containing_year.}
}
\value{
dataframe dat with a new column added called 'year'. If 'year' is
already a column in dat an error is raised.
}
\description{
Find the column containing year using a regular expression,
get the year from that column and put it in a new column called 'year'.
}
\examples{
\dontrun{
dat_mixed <- data.frame(
     `date span` = c("Jan to Mar 2021", "2020-21")
     )
dat_w_year <- data.frame(
     year = c("Jan to Mar 2021", "2020-21")
     )
dat_calendar <- data.frame(`date span` =
                             c("Jan to Mar 2021", "Apr 2020 to Mar 2021"))
dat_financial <- data.frame(`date span` = c("Q1 to Q4 2020-21", "2021"))

get_year_from_column(dat_mixed, "(?i)date*.span")
get_year_from_column(dat_w_year, "(?i)year")
get_year_from_column(dat_calendar, "(?i)date*.span")
get_year_from_column(dat_financial, "(?i)date*.span")
}
}
