% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fy_start.R
\name{get_year_info_from_fy_end}
\alias{get_year_info_from_fy_end}
\title{Get year from column containing the financial year end}
\usage{
get_year_info_from_fy_end(
  dat,
  fy_end_pattern,
  fy_start_from_fy_end = FALSE,
  fy_from_fy_end = FALSE
)
}
\arguments{
\item{dat}{dataframe containing a column name that matches fy_end_pattern}

\item{fy_end_pattern}{character string. Regex pattern for the name of the
column containing financial year end information in the format YYYY}

\item{fy_start_from_fy_end}{bool. Default is FALSE. If TRUE, a column called
fy_start is added to dat and filled using the format YYYY. If fy_start is
already a column in dat, it is NOT overwritten. In pub sec the setting
is also called fy_start_from_fy_end.}

\item{fy_from_fy_end}{bool. Default is FALSE. If TRUE, a column called 'year'
is added to dat and populated using the format YYYY-YY. If year is
already a column in dat, it is NOT overwritten. In pub sec the setting
is also called fy_start_from_fy_end.}
}
\value{
dataframe dat with new columns.
}
\description{
If the only year column given is the financial year end, this
function can be used to get the financial year (YYYY-YY) or the financial
year start (YYYY), or both. These are added as new columns. No columns are
removed.
}
\examples{
\dontrun{
dat <- data.frame("fy_end" = c(2021, 2022))
get_year_info_from_fy_end(dat, "end",
                          fy_start_from_fy_end=TRUE,
                          fy_from_fy_end=TRUE)
}
}
