% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_year_column.R
\name{get_year_names_by_type}
\alias{get_year_names_by_type}
\title{get the column names and year types of columns containing years}
\usage{
get_year_names_by_type(dat, columns)
}
\arguments{
\item{dat}{dataframe}

\item{columns}{character string vector of column names.}
}
\value{
named vector in which the name is the type of year and the value is
the name of the column. Types of year are 'financial', 'calendar', and
'mixed'. If more than one of the given columns contains the same type of year
the one with the highest frequency of valid years is given for that type.
}
\description{
Create a named vector that states a maximum of one column name
for each year type, where year types are 'financial' (variations on YYYY-YY,
YYYY-YYY, YYYY_YY, YYYY to YYYY etc), 'calendar' (YYYY), or 'mixed' (a mix
of calendar and financial year types).
}
\examples{
\dontrun{
dat <- data.frame(
    calendar_years = 2021,
    alternative_financial_years = c("2022-23", NA),
    financial_years = "2022-23",
    random = c("2021", "2022-23"),
    words = "hello"
    )
get_year_names_by_type(dat, names(dat))

}
}
