% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_handlers.R
\name{get_year_type}
\alias{get_year_type}
\title{Get the year type for each row of a dataframe, or for a single year}
\usage{
get_year_type(dat, column = NA)
}
\arguments{
\item{dat}{character string, integer, or dataframe. If character string or
integer, one calendar or financial year. If dataframe, a dataframe with a
column named the same as \code{column}.}

\item{column}{character string or NA. defaults to NA as this arg is only used
when dat is a dataframe or tibble.}
}
\value{
same type as dat, stating the year type. If dat is a dataframe a new
column is added called 'year_type' containing 'financial', 'calendar' or NA.
}
\description{
Get an unambiguous year type. If more than one year type is
found NA is returned.
}
\examples{
\dontrun{
dat <- data.frame(year = c("2021", "2021-22", "2021 2021-22"))
get_year_type(dat, "year")

get_year_type("2021-22")
get_year_type("2021")
get_year_type("202123")

}
}
