% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_absent_packages.R
\name{identify_absent_packages}
\alias{identify_absent_packages}
\title{identify_absent_packages}
\usage{
identify_absent_packages(packages)
}
\arguments{
\item{packages.}{A vector of packages you want to know if they are installed}
}
\value{
absent_packages A vector of packages not installed
}
\description{
identify packages that are not already installed
}
\details{
value
identify packages that are not already installed

Takes a vector of package names that you need and returns a list of packages 
that are not already installed
}
\examples{
absent_packages <- identify_absent_packages(c("noSuchPackager", "tidyr"))

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}
}
