% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_information.R
\name{identify_columns_to_remove}
\alias{identify_columns_to_remove}
\title{Identify header cells to remove from xlsx_cells data.}
\usage{
identify_columns_to_remove(dat, patterns)
}
\arguments{
\item{dat}{dataframe imported using xlsx_cells.}

\item{patterns}{vector of character strings that are regular expressions,
each one matching the name of a column in the raw data.}
}
\value{
vector of character strings giving the column numbers identified.
If no match is found in the data for a pattern a warning is given.
}
\description{
Use regular expressions to identify which column of the data
they appear in. This function could be improved by adding a mzx_row parameter
that states the rows to look in (which would be known from header row +
number of header rows)

If a pattern is matched in more than one column none of those matches are
returned as columns to remove. This does not affect the use of any other
patterns provided.
}
\examples{
\dontrun{
dat <- data.frame(
  "row" = c(1, 1, 1, 2, 2, 2),
  "col" = c(1, 2, 3, 1, 2, 3),
  "address" = c("A1", "B1", "C1", "A2", "B2", "C2"),
  "numeric" = c(NA, NA, NA, NA, 1, 100),
  "character" = c("name", "id", "value", "first row", NA, NA),
  "data_type" = c(rep("character", 4), rep("numeric", 2))
)
# view as it would be in excel:
rectify(dat)

identify_columns_to_remove(dat, "id")
}
}
