% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_duplicates.R
\name{identify_rows_for_renaming}
\alias{identify_rows_for_renaming}
\title{Identify which of the repeated rows needs to be renamed}
\usage{
identify_rows_for_renaming(
  dat,
  cols_to_group_by,
  column,
  pattern,
  index,
  expected_freq
)
}
\arguments{
\item{dat}{dataframe. Must include a column with the name given in column.}

\item{cols_to_group_by}{vector of character strings giving the names of
columns that will have been created during un-pivotting. The entries in these
columns would uniquely identify each row if there were no duplicates.}

\item{column}{character string. The name of the column the duplicate is in in
dat. Vector of length 1. In pub sec this variable is specified by
rename_duplicate_in_col.}

\item{pattern}{character string giving a regular expression that identifies
the duplicate. Vector of length 1. In pub sec this variable is specified by
rename_duplicate_pattern.}

\item{index}{integer. Which to rename e.g. '2' would rename the 2nd. Can be a
vector of length > 1, but if so the number of items must match that in
prefix. In pub sec this variable is specified with rename_duplicate_index.}

\item{expected_freq}{integer. The expected number of repeats of the
column name in the raw data. Vector of length 1. In pub sec this variable is
specified by rename_duplicate_expected_freq.}
}
\value{
The original dataframe with the following columns added:
\itemize{
\item rename_required: boolean. Identifies the rows containing entries to rename.
\item location: integer. Identifies whether it is the first second, third of the
instance.
}
}
\description{
Flag rows fit the following criteria:
\itemize{
\item Entry in the specified column matches pattern.
\item The entry is repeated in at least 2 groups when data are grouped by
cols_to_group_by.
\item The repeated entry is in the position given by index e.g.
if this is set to 1, only the entries that relate to the first of the
relevant raw data columns are flagged.
\item The number of times that name is used in the raw data equals
expected_freq
}
}
\examples{
\dontrun{
dat <- data.frame(
  "year" = rep(c("2022-23", "2023-24"), each = 5),
  "item_detail" = rep(c("A", "B", "Total", "C", "Total"), 2),
  "numeric" = c(2, 3, 5, 0.1, 5.1,
                1, 2, 3, 0.4, 3.4)
)
cols_to_group_by <- c("year", "item_detail")

result <- identify_rows_for_renaming(
    dat, cols_to_group_by, "item_detail", "(?i)total", 1, 2
    )
}
}
