% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_duplicates.R
\name{identify_target_rows}
\alias{identify_target_rows}
\title{Identify rows where the entry may require a prefix}
\usage{
identify_target_rows(dat, cols_to_group_by, column, pattern, index)
}
\arguments{
\item{dat}{dataframe. Must include a column with the name given in column.}

\item{cols_to_group_by}{vector of character strings giving the names of
columns that will have been created during un-pivotting. The entries in these
columns would uniquely identify each row if there were no duplicates.}

\item{column}{character string. The name of the column the duplicate is in in
dat. Vector of length 1. In pub sec this variable is specified by
rename_duplicate_in_col.}

\item{pattern}{character string giving a regular expression that identifies
the duplicate. Vector of length 1. In pub sec this variable is specified by
rename_duplicate_pattern.}

\item{index}{integer. Which to rename e.g. '2' would rename the 2nd. Can be a
vector of length > 1, but if so the number of items must match that in
prefix. In pub sec this variable is specified with rename_duplicate_index.}
}
\value{
The original dataframe with the following columns added:
\itemize{
\item rename_required: boolean. Identifies the rows containing entries to rename.
\item matches_regex: boolean. Identifies entries that match pattern.
\item location: integer. Identifies whether it is the first second, third of the
instance.
A warning is raised if repeated entries have been found but they do not match
the pattern.
}
}
\description{
Flag rows that fit the following criteria:
\itemize{
\item Entry in the specified column matches the given pattern.
\item The entry is in a position given by index e.g.
if index is set to 1, only the entries that relate to the first of the
relevant raw data columns are flagged.
No checks are done at this point on whether the entry is duplicated.
}
}
\examples{
\dontrun{
dat <- data.frame(
  "year" = rep(c("2022-23", "2023-24"), each = 5),
  "item_detail" = rep(c("A", "B", "Total", "C", "Total"), 2),
  "numeric" = c(2, 3, 5, 0.1, 5.1,
                1, 2, 3, 0.4, 3.4)
)
cols_to_group_by <- c("year", "item_detail")
column <- "item_detail"
pattern <- "(?i)total"
index <- 1

result <- identify_target_rows(dat, cols_to_group_by, column, pattern, index)
}
}
