% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_absent_packages.R
\name{install_absent_packages}
\alias{install_absent_packages}
\title{install_absent_packages}
\usage{
install_absent_packages(packages)
}
\arguments{
\item{packages.}{A vector of packages you want to install if they are not already}
}
\description{
install packages that are not already installed
}
\details{
value
Takes a vector of package names that you need and installs
any packages that are not already installed. Returns a message to say all are
already installed, or installs any absent packages, with the usual 
install.packages messages.
}
\examples{
install_absent_packages(c("dpyr", "tidyr"))

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}
}
