% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_needed_packages.R
\name{install_needed_packages}
\alias{install_needed_packages}
\title{install_needed_packages}
\usage{
install_needed_packages()
}
\description{
Un-pivot data with one row of any number of columns
}
\examples{
start_debug_logfile()
}
\author{
Mark London \email{mark.london@ext.ons.gov.uk}
}
