% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_cols_data.R
\name{join_cols_data}
\alias{join_cols_data}
\title{join_cols_data}
\usage{
join_cols_data(dat, columns_to_concat, col_join_name = "description")
}
\arguments{
\item{dat}{dataframe}

\item{columns_to_concat}{vector of strings of column names to concatenate.}
}
\value{
tibble. Tidy dataframe.
}
\description{
checks if any reserved words are used, halts if they are
}
\details{
Check for reserved words in sheet_structure settings
}
\examples{
my_out <- join_cols_data(dat = new_simple, columns_to_concat = my_columns, col_join_name = 'description'))
}
\author{
Mark London \email{mark.london@ext.ons.gov.uk}
}
