% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot_data.R
\name{join_left_and_right_unpivotted_data}
\alias{join_left_and_right_unpivotted_data}
\title{Join the main value data to the left block of row descriptor columns}
\usage{
join_left_and_right_unpivotted_data(
  left,
  right,
  column_names,
  first_header_row
)
}
\arguments{
\item{left}{dataframe. The beheaded left (row names) block of data.}

\item{right}{dataframe. The beheaded right (numeric) block of data.}

\item{column_names}{character string vector. The names of columns to join on
(all shared column names).}

\item{first_header_row}{integer. The row number of the first header.}
}
\value{
dataframe that will look like the Excel data but with a tidy layout
(one numeric value per row).
}
\description{
The left block contains descriptor columns, the right block
contains both the numeric value column and the descriptors whose entries come
from the Excel column names of the numeric columns. When they are joined all
rows from both blocks are retained.
}
\examples{
\dontrun{
# taken from results of examples in behead_left_block and behead_right_block
# so you can follow it though.
left <- data.frame(
  address = c("B1", "C1", "B2", "C2"),
  row = c(1, 1, 2, 2),
  col = c( 2, 3, 2, 3),
  data_type = c(rep("character", 2), rep("numeric", 2)),
  numeric = c(NA, NA, 100, 22),
  character = c("primary", "secondary", NA, NA),
  Year = c("Year", "Year", "2021", "2021")
)

right <- data.frame(
  address = c("B2", "C2"),
  row = 2,
  col = c(2, 3),
  data_type = "numeric",
  numeric = c(100, 22),
  character = NA,
  description_1 = c("primary", "secondary")
)

column_names <- c("address", "row", "col", "data_type", "numeric", "character")
join_left_and_right_unpivotted_data(left, right, column_names, 1)
}
}
