% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_handlers.R
\name{make_calendar_q1_month_and_quarter_patterns}
\alias{make_calendar_q1_month_and_quarter_patterns}
\title{Make regex patterns for months and quarters in Q1 of a calendar year}
\usage{
make_calendar_q1_month_and_quarter_patterns(
  period = "all",
  q1_jan_to_mar = TRUE
)
}
\arguments{
\item{period}{Must be 'all', 'quarter', or 'month'. Default is 'all'.}

\item{q1_jan_to_mar}{bool, default is TRUE. If FALSE, it is assumed that
Q4 is jan to mar, and Q1 is the start of the financial year.}
}
\value{
a regular expression pattern for the selected period(s)
}
\description{
Create regular expressions that can be used to identify the
months and quarters of a calendar year whose financial year start year is the
calendar year minus 1.
}
\examples{
\dontrun{
make_calendar_q1_month_and_quarter_patterns(q1_jan_to_mar=FALSE)
}
}
