% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_handlers.R
\name{make_quarter_patterns}
\alias{make_quarter_patterns}
\title{Build regex patterns for quarters 1 through 4}
\usage{
make_quarter_patterns()
}
\value{
named vector. Currentluy only one pattern is created: 'single' which
can be used to identify single quarters. No pattern is yet returned for
quarter ranges such as Q1-Q3, but this could be a future addition (hence the
return of a named vector rather than a single value).
}
\description{
Create regular expression patterns that match quarters. NOTE:
The pattern has been designed for use where no other individual numbers
between 1 and 4 are expected e.g. the returned pattern would correctly
identify quarters in a vector containing year and quarter (e.g. 2024 Q1), but
not in a descriptive paragraph e.g. ('This is the 1st release of Q2 data'
would return both 1 and 2).
}
\examples{
\dontrun{
dat <- c("Q1", "1", ".1", "quarter 1", "q1",
           "1-3", "1 - 3", "q1-Q3", "q1 - Q3", "Q1to3", "Q1   to Q3", "13",
          "12-3", "Q1 note 2")
patterns <- make_quarter_patterns()['single']
str_extract_all(dat, patterns)
}
}
