% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_handlers.R
\name{make_year_patterns}
\alias{make_year_patterns}
\title{Build regex patterns for years between 1900 and 2099}
\usage{
make_year_patterns()
}
\value{
named vector with calendar and financial patterns Different year
types can be accessed by subsetting for 'calendar', 'financial',
'long_form_financial', 'short_form_financial', and 'fy_separators'.
}
\description{
Create regular expression patterns that match calendar and
financial years from the 20th and 21st centuries. Will not match years that
are either immediately preceded or followed by a dash. Acceptable financial
year formats are:
YYYY-YY, YYYY/YY, YYYY_YY, YYYY to YY, YYYYtoYY,
YYYY-YYYY, YYY/YYYY, YYYY_YYYY, YYYY to YYYY, YYYYtoYYYY.
Acceptable calendar year format is: YYYY.
}
\examples{
\dontrun{
patterns <- make_year_patterns()
}
}
