% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{match_sheet_to_regex}
\alias{match_sheet_to_regex}
\title{Find values in a vector that match a regular expression}
\usage{
match_sheet_to_regex(sheet_names, pattern)
}
\arguments{
\item{sheet_names}{Vector of character strings.}

\item{pattern}{Regular expression to be detected in sheet_names.}
}
\value{
List of character strings. Sheet names that fitted regex pattern
}
\description{
Find values in a vector that match a regular expression
}
\examples{
\dontrun{
sheet_names <- c("Front Page", "LA drop-down", "RA data", "RA other")
pattern <- "RA|Data"
match_sheet_to_regex(sheet_names, pattern)
}
}
