% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_from_input.R
\name{refine_blanks}
\alias{refine_blanks}
\title{Treat cells with spaces but nothing else as blank cells.}
\usage{
refine_blanks(dat)
}
\arguments{
\item{dat}{dataframe that has been imported using xlsx_cells.}
}
\value{
dataframe in xlsx_cells format with space only cells shown as blank.
}
\description{
If a cell has only spaces in, xlsx_cells reads it as a
character string. Change this so that they are seen as blank cells.
}
\details{
Without this step, if there is a space in a header row like this:\tabular{llll}{
   final \tab  \tab \emph{space} \tab budget \cr
   \strong{2021} \tab \strong{2022} \tab \strong{2023} \tab \strong{2024} \cr
   10 \tab 20 \tab 30 \tab 40 \cr
}


Because of that space we get a blank vintage for 2023.\tabular{lll}{
   year \tab vintage \tab value \cr
   2021 \tab final \tab 10 \cr
   2022 \tab final \tab 20 \cr
   2023 \tab  \tab 30 \cr
   2024 \tab budget \tab 40 \cr
}


By making the cell with the space be seen as blank, the blank vintage entry
for 2023 will be correctly filled in with 'final'.
}
\examples{
\dontrun{
dat <- data.frame(
    data_type = c("character", "character", "numeric", "blank"),
    is_blank = c(FALSE, FALSE, FALSE, TRUE),
    character = c("Final", "  ", NA, NA),
    numeric = c(NA, NA, 1, NA),
    address = c("A1", "A2", "A3", "A4"),
    row = 1,
    col = 1:4
    )
refine_blanks(dat)
}
}
