% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_handlers.R
\name{refine_metadata_year}
\alias{refine_metadata_year}
\title{Refine year taken from sheet or table metadata}
\usage{
refine_metadata_year(years, preferred)
}
\arguments{
\item{preferred}{characters string. Financial or calendar.}

\item{year}{vector of character strings, each of which is a year in either
calendar or financial year format.}
}
\value{
character string vector of years or NA (see details). Years are
returned in chronological order.
}
\description{
Get the year(s) that match(es) the preferred year type, standardise
them, and sort by chronological order.
}
\details{
Used by get_year_for_use_in_data.

The preferred year type is the type of year given in the filename. If a year
is given in the file name, and the year type of the year provided does not
match it, then NA is returned

Note: If there is no year in the filename, the financial year is preferred,
but the calendar year is still returned if not financial year is available
(this is handled in get_year).
}
\examples{
\donotrun{
refine_metadata_year(c("2022-23", "2021-22", "2022"), "financial")
refine_metadata_year(c("2022-23", "2021-22", "2022"), NA)
refine_metadata_year(c("2022-23", "2021-22", "2022"), "calendar")
}

}
