% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_duplicates.R
\name{refine_target_rows}
\alias{refine_target_rows}
\title{Refine which of the repeated rows needs to be renamed}
\usage{
refine_target_rows(dat, column)
}
\arguments{
\item{dat}{dataframe. Must include a column with the name given in column.}

\item{column}{character string. The name of the column the duplicate is in in
dat. Vector of length 1. In pub sec this variable is specified by
rename_duplicate_in_col.}
}
\value{
dat with 'rename_required' updated
}
\description{
Refine the flag for which rows contain entries that require a
prefix to exclude cases that are unique when grouped by cols_to_group_by
E.g. if the entries that are usually repeated in the raw data
are called 'total' the pattern might be '(?i)total'. In
subsequent releases if the data has two entries called 'total A' and one
called 'total B', 'total B' may be flagged to be renamed because it matches
pattern, however it is clearly not a target for renaming, because it is
unique.

Warn the user if there is more than one entry that matches the pattern
and mark these as not to be renamed.
E.g. if the entries that are usually repeated in the raw data
are called 'total' the pattern might be '(?i)total'. In
subsequent releases if the data has two columns called 'total A' and two
called 'total B', we don't want to do the rename, because we don't know which
is the right one.

Warn the user if there are no duplicated entries that match pattern.
}
\examples{
\dontrun{
dat <- data.frame(
  "item_detail" = c("Total A", "Total A", "Total B"),
  "numeric" = c(2, 3, 5),
  "matches_regex" = rep(TRUE, 3),
  "location" = c(1, 2, 1),
  "rename_required" = c(TRUE, FALSE, TRUE)
)
column <- "item_detail"
result <- refine_target_rows(dat, column)
}
}
