% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_multi_table_data.R
\name{reformat_multi_table_data}
\alias{reformat_multi_table_data}
\title{reformat_multi_table_data}
\usage{
reformat_multi_table_data(
  dat,
  table_split_points,
  table_heading,
  subtable_heading,
  group_row,
  nested_row_1,
  group_col
)
}
\arguments{
\item{dat}{tidyxl::xlsx_cells table. Must include info above the first header row}

\item{table_split_points}{vector of regular expressions. The first pattern
will be used to make the initial split, the second, the second split etc}

\item{table_heading}{string. Taken from the data dictionary.
Number of elements must equal the number of parts labelled 'Part ...' in  source data}

\item{group_row}{string. New column name for the grouping variables.
Number of elements must equal the number of parts labelled 'Part ...' in  source data}

\item{nested_row_1}{string. New column name for the nested variable.
Number of elements must equal the number of parts labelled 'Part ...' in  source data}

\item{group_col}{string. Taken from the data dictionary}
}
\value{
tibble. Tidy dataframe containing data from all tables.
}
\description{
Reformat data with multiple tables in one sheet
}
\details{
Reformat data with multiple tables in one sheet

Used for cases where data have multiple tables in one sheet, with tables
stacked vertically. Each table only has one header row. Each table is split
by row headers (e.g. sub-services are listed under a row containing the
service name in the sub-service column, and no data in the other columns)

The data column names from the original table are carried over to the output.

This is used for e.g. 2022 Scottish Gov POBE data.
}
\examples{
EW To do

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
