% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_no_row_headers_data.R
\name{reformat_no_row_headers_data}
\alias{reformat_no_row_headers_data}
\title{reformat_no_row_headers_data}
\usage{
reformat_no_row_headers_data(main_table, group_row, nested_row_1, group_col)
}
\arguments{
\item{main_table}{A dataframe containing the main data table to be processed.}

\item{group_row}{string. New column name for the groups the nested_row_1 
entries belong to.}

\item{nested_row_1}{string. New column name for the entries in the first 
non-empty column.}

\item{group_col}{string. New column name for the header row 
(all columns except the first non-empty column.}
}
\value{
A data frame containing the unpivotted data.
}
\description{
Reformat no_row_headers data
}
\details{
Reformat no_row_headers data

Used for cases where Excel data have only one header row, with one column
of strings in the first non-empty column. 'Total group' items are included in 
the first non-blank column - these are the groups under which the nested_row_1
is nested.

Reformat these data so that there is only one number per row, and 
the group of 'total group' is associated with all the relevant entries.
}
\examples{
sample_data <- data.frame(
  address = c('B2', 'C2', 'A3', 'B3', 'C3', 'A4', 'B4', 'C4', 'A5', 'B5', 'C5'),
  row = c(2, 2, rep(3:5, each = 3)),
  col = c(2, 3, rep(1:3, 3)),
  is_blank = rep(FALSE, 11),
  data_type = c(rep('character', 3), 
              rep(c(rep('numeric', 2), 'character'), 2), rep('numeric', 2)),
  numeric = c(rep(NA, 3), 1, 2, NA, 3, 4, NA, 5, 6),
  character = c('measure 1', 'measure 2', 'primary', rep(NA, 2),
                'secondary', rep(NA, 2), 'Total education', rep(NA, 2))
)

# look at the data as it would be in Excel
rectify(sample_data)

reformat_no_row_headers_data(sample_data, 'service', 'subservice', 'transaction')

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
