% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot_data.R
\name{reformat_tidy_data}
\alias{reformat_tidy_data}
\title{Reformat tidy data}
\usage{
reformat_tidy_data(dat, columns_to_create = NA, tidy_notes_name = NA)
}
\arguments{
\item{dat}{xlsx_cells data}

\item{columns_to_create}{character string. Optional. Used as the name of a
column that will contain the original name of the value column.}

\item{tidy_notes_name}{character string. Only required if there is a
character column to the right of tidy data, in which case it is used as a
name for that column.}
}
\value{
dat as a tibble with the same layout as the raw data.
}
\description{
Where the raw data are already in a tidy format with one numeric value per
row, a single row of column headers, and a single data type in each column,
this function is used to take the data out of xlsx_cells format.
}
\details{
Note that if there are ANY character strings in the 'value' column, that
column will be treated as a character column not numeric. It is assumed
that if data really are in a tidy format, this will not occur. However a fix
may need to be coded in if this ever comes up.
}
\examples{
\dontrun{
dat <- data.frame(
    row = c(1, 1, 2, 2),
    col = c(1, 2, 1, 2),
    data_type = c(rep("character", 3), "numeric"),
    character = c("description", "value", "A", NA),
    numeric = c(NA, NA, NA, 100),
    is_blank = FALSE,
    sheet = "Sheet1"
    )
 reformat_tidy_data(dat)
}
}
