% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_two_header_data.R
\name{reformat_two_header_data}
\alias{reformat_two_header_data}
\title{reformat_two_header_data}
\usage{
reformat_two_header_data(
  dat,
  group_col,
  nested_column_1,
  left_headers,
  tolerance_list = c(tolerance_numeric = 0.4, tolerance_char = 0.7, tolerance_blank =
    0.9)
)
}
\arguments{
\item{dat}{A dataframe containing the main data table to be processed.}

\item{group_col}{string. New column name for the first header row.}

\item{nested_column_1}{string. New column name for the second header row.}

\item{tolerance_list}{an optional list of tolerances which must be in the format:
tolerance_list <- c(tolerance_numeric = 0.4, tolerance_char = 0.7, tolerance_blank = 0.9)}
}
\value{
A data frame containing the processed and rearranged data after unpivoting.
}
\description{
Reformat data with one header row on the left and two on the right
}
\details{
Reformat data with one header row on the left and two on the right

Used for cases where Excel data have only one header row in the left block,
but two on the right. the left headers are on the same row as the second
row of headers on the right.

The two blocks of headers are treated differently:

The column names of block of data on the left (in the original table) are
carried over to the output.
If a column in the left block doesn't have a name it is called 'column_<x>'.

The block of data on the right is changed from wide format to long, so the
original column names become entries in 2 new columns (group_col, and
nested_column_1)

This is used for e.g. dluhc revenue expenditure final data for RO2 in 2021-22.
}
\examples{
sample_data <- data.frame(
   address = c('A1', 'B1', 'C1', 'D1', 'A2', 'B2', 'C2', 'D2', 'A3', 'B3', 'C3', 'D3'),
   row = c(1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3),
   col = c(1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4),
   is_blank = c(TRUE, TRUE, FALSE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE),
   data_type = c('blank', 'blank', 'character', 'blank', 'character',
   'character', 'character', 'character', 'character', 'character', 'numeric', 'numeric'),
   numeric = c(NA, NA, NA, NA,NA, NA, NA, NA, NA, NA, 10, 20),
   date = c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA),
   character = c(NA, NA, 'Education', NA, 'LA', 'Region', 'primary', 'secondary', 'Cornwall', 'SW', NA, NA)
 )
reformat_two_header_data(sample_data, 'service', 'subservice')
}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
