% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot_data.R
\name{remove_duplicated_header_rows}
\alias{remove_duplicated_header_rows}
\title{Only keep the first of multiple rows of matching headers}
\usage{
remove_duplicated_header_rows(dat, first_header_row, header_row_count)
}
\arguments{
\item{dat}{dataframe. Data imported using tidyxl::xlsx_cells.}

\item{first_header_row}{integer. The row number of the first header.}

\item{header_row_count}{integer. The number of header rows (not including
repeated sets).}
}
\value{
dat with repeated headers removed.
}
\description{
This function exists because in the DLUHC CPR files (e.g.
in 2024-25), the year_&_quarter_series table contains two tables stacked
vertically that have exactly the same headers because the top table is a
filtered version of the table below. This function removes repeated rows of
headers, keeping only the first.

If there are multiple header rows for any one table (i.e. if header_row_count
is greater than 1, duplicated header rows will only be removed if ALL header
rows are duplicated. In the example below, the headers of the 2nd table would
be removed in set 1, but no headers would be removed from set 2 (a warning is
given):

Set 1:
> Education
Year  > primary
------ > ----------
2020 > 100,000

\if{html}{\out{<div class="sourceCode">}}\preformatted{   > Education
}\if{html}{\out{</div>}}

Year  > primary
------ > ----------
2021  >  90,000

Set 2:
(note that year is replaced with month in the second table)
> Education
Year  > primary
------ > ----------
2020 > 100,000

\if{html}{\out{<div class="sourceCode">}}\preformatted{   > Education
}\if{html}{\out{</div>}}

Month > primary
------ > ----------
Jan  >  8,000
}
\examples{
\dontrun{
#  Year > Education
# ----- > ----------
#  2020 >    22
#
#  Year > Education
# ----- > ----------
#  2020 >    22

dat <- data.frame(
  address = c("A1", "B1", "A2", "B2", "A3", "B3", "A4", "B4"),
  row = rep(1:4, each = 2),
  col = rep(1:2, times = 4),
  data_type = rep(c(rep("character", 2), rep("numeric", 2)), times = 2),
  numeric = rep(c(rep(NA, 2), 2020, 22), times = 2),
  character = rep(c("Year", "Education", NA, NA), times = 2)
  )
remove_duplicated_header_rows(dat, 1, 1)

# case where there are partially duplicated headers (2 header rows, but only
# one is duplicated)
dat <- data.frame(
  address = c("A1", "B1", "A2", "B2", "A3", "B3", "A4", "B4"),
  row = rep(1:4, each = 2),
  col = rep(1:2, times = 4),
  data_type = c(
      rep("character", 3), "blank", rep("numeric", 2), "character", "blank"
      ),
  numeric = c(rep(NA, 4), 1, 2, NA, NA),
  character = c("a1", "b1", "a2", NA, NA, NA, "a2", NA)
  )
remove_duplicated_header_rows(dat, 1, 2)
}
}
