% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_xlsx_cells_table_data.R
\name{remove_empty_lines}
\alias{remove_empty_lines}
\title{remove empty columns or empty rows from xlsx_cells data}
\usage{
remove_empty_lines(dat, direction, formula_zero_as_blank = TRUE)
}
\arguments{
\item{dat}{dataframe imported using tidyxl::xlsx_cells}

\item{direction}{character string. Either "row" or "col". If "row", rows
relating to cells in an empty row are removed. If "col", rows relating
to cells in an empty column are removed.}

\item{formula_zero_as_blank}{boolean. Defaults to TRUE. When TRUE if a cell
contains a formula the result of which is 0, this is counted as a blank.}
}
\value{
dat with rows removed. A warning is given stating which rows or
columns have been removed.
}
\description{
Remove rows from a dataframe imported using tidyxl::xlsx_cells
if they belongs to a row or column in the Excel data that is empty.

If a cell contains a formula the result of which is 0, this is counted as a
blank row by default. If this is not the desired behaviour set
formula_zero_as_blank to FALSE.
}
\examples{
\dontrun{
dat <- data.frame(
    address = c("A1", "A2", "B1", "B2"),
    row = rep(1:2, times = 2),
    col = rep(1:2, each = 2),
    is_blank = c(TRUE, TRUE, FALSE, TRUE),
    character = c(NA, NA, "hello", NA),
    data_type = c("blank", "blank", "character", "blank"),
    formula = NA,
    numeric = NA
    )
 remove_empty_lines(dat, "col")
 remove_empty_lines(dat, "row")
}
}
