% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_from_output.R
\name{remove_from_output}
\alias{remove_from_output}
\title{Remove unwanted rows and columns from the output dataframe}
\usage{
remove_from_output(
  dat,
  col_patterns_with_values_to_drop,
  value_patterns_to_drop,
  xlsx_cells_names
)
}
\arguments{
\item{dat}{dataframe.}

\item{col_patterns_with_values_to_drop}{character vector. Regular expressions
to match the names of the columns containing the values that match
value_patterns_to_drop.}

\item{value_patterns_to_drop}{character vector. Regular expressions to match
the values whose rows are to be dropped.}

\item{xlsx_cells_names}{character string vector. Column names of the source
data which will have been imported using xlsx_cells.}
}
\value{
dataframe with specified rows, blank rows, repeat rows, and
unwanted columns from xlsx_cells removed. A warning is raised if repeat
rows are found.
}
\description{
Remove blank rows, duplicated rows, and unwanted columns created
by xlsx_cells. These are columns such as 'row', 'col', 'local_format_id',
some of which will have been used as part of tidying the data, but others
are not used by this pipeline.
}
\details{
For more information and examples see documentation for remove_is_blank_rows,
drop_rows_with_values, deduplicate_data, and remove_unwanted_cols.
}
