% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_from_input.R
\name{remove_hidden_character_strings}
\alias{remove_hidden_character_strings}
\title{Remove hidden character strings in the first header row.}
\usage{
remove_hidden_character_strings(dat, filepath, remove_cell = NA)
}
\arguments{
\item{dat}{dataframe imported using tidyxl::xlsx_cells()}

\item{filepath}{string. Filepath for input data to detect cell formats.}

\item{remove_cell}{bool. TRUE, FALSE or NA. Default is NA. FALSE leads to the
same behaviour as NA. In pub sec this variable is specified by
remove_hidden_strings_bool.}
}
\value{
dataframe. dat with rows removed for the identified cells. A warning
is raised if rows are removed.
}
\description{
Remove cells whose text is the same colour as the background.
This function was created because in pub sec scot gov HRA 2021-22 there are
hidden character strings in the
first header row that are hidden by being the same colour as the background.
These need to be removed so that un-pivotting works correctly.
}
\examples{
\dontrun{
dat <- data.frame(address = c("A1"),
                  local_format_id = c(1))
filepath <- D:/.../interim/test_dat.xlsx
remove_hidden_strings_bool <- "TRUE"

remove_hidden_character_strings(dat, filepath, remove_hidden_strings_bool)
}
}
