% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_handlers.r
\name{remove_invalid_args}
\alias{remove_invalid_args}
\title{Remove rows for invalid variables from the dictionary of settings.}
\usage{
remove_invalid_args(dict, variables)
}
\arguments{
\item{dict}{dataframe with columns 'key' and 'item'. 'key' gives the name of
the variable to be, 'item' the value of that variable.}

\item{variables}{character string vector. The name of the variables that are
to be created in R.}
}
\value{
dataframe. dict with invalid rows removed. A warning is given for
each row that is removed.
}
\description{
In pub sec all possible R variable names are passed to R from
Python as a list (r_vars). If there is a variable in the dict that is not in
r_vars it cannot be created as an R variable. Remove it from the dict with a
warning.
}
\examples{
\dontrun{
dict <- data.frame(key = c("do_thing", "sing"), item = c("jump", "A"))
variables <- c("do_thing", "ignore_thing")
remove_invalid_args(dict, variables)
}

}
