% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_from_output.R
\name{remove_is_blank_rows}
\alias{remove_is_blank_rows}
\title{Remove rows marked as blank}
\usage{
remove_is_blank_rows(dat)
}
\arguments{
\item{dat}{dataframe}
}
\value{
dat with is_blank == TRUE rows removed
}
\description{
Once data have been beheaded (unpivotted) any rows marked as
blank that are leftover from xlsx_cells data are no longer required. This
function removes those rows.
}
\examples{
\dontrun{
dat <- data.frame(is_blank = c(TRUE, FALSE), numeric = c(NA, 1))
remove_is_blank_rows(dat)
}
}
