% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_wording.R
\name{remove_line_breaks}
\alias{remove_line_breaks}
\title{Remove line breaks from character strings.}
\usage{
remove_line_breaks(dat)
}
\arguments{
\item{dat}{dataframe.}
}
\value{
dataframe with line breaks from all character strings.
}
\description{
Remove line breaks from character strings.
}
\examples{
\dontrun{
dat <- tibble::tibble(
    "Service"  = c(
       "Education\nServices","Education Services",
       "Education\n\n Services",
       "\nEducation\n Services\n"
       ),
   "Transaction" = c(
       "Expenditure\r",
       "\rExpenditure",
       "\nExpenditure\n",
       "Expenditure"
       ),
   "Units" = c("thousands","thousands","hundreds", "millions")
   )

result <- remove_line_breaks(dat)
}
}
