% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_xlsx_cells_table_data.R
\name{remove_metadata_cells}
\alias{remove_metadata_cells}
\title{Remove first cell from a dataset matching each regular expression}
\usage{
remove_metadata_cells(dat, patterns = NA, n_row = 5)
}
\arguments{
\item{dat}{A data frame imported using tidyxl::xlsx_cells.}

\item{patterns}{vector of character strings. Each must be a regular
expression used to identify the unwanted cells. More than one pattern may be
given, but there should only be one pattern for each cell to be removed:
Only the first match to each pattern will be removed. If two patterns are
given and the first match for each is the same cell, only that one cell will
be removed. In pub sec this variable is specified by
metadata_cells_to_remove_patterns.}

\item{n_row}{integer (optional, defaults to 5) An integer specifying the
number of rows to check for matches to the metadata_cells_to_remove_patterns.
This is based on the rows of dat, not original row number (stated in the
column 'row') i.e. if this function is called after blank rows have been
removed, it will only check that many rows. In pub sec this variable is
specified by populated_rows_to_check_for_metadata_to_remove.}
}
\value{
A data frame with the specified metadata cells removed. If multiple
matches for the pattern are found, only the first instance is removed
and a warning is raised. If the pattern is not found, a warning is raised
and no rows are removed.
}
\description{
Remove rows from an xlsx_cells imported dataframe if the
character value is the first match for the supplied regular expression.

It should only be used when absolutely necessary as it carries the risk of
removing a cell that is actually wanted. To mitigate this risk only a given
number of non-blank rows are checked for the pattern using n_row.
}
\details{
This function is not included in remove_from_input because that is only
called for the whole sheet, whereas this function is better called at the
table level.

This function was written for cases where metadata (title or units) are given
in the same row as the headers, and therefore interfere with beheading.
If the cells to remove are metadata cells this function must be called after
the required information has been extracted from those cells.

Non-metadata cells could theoretically be removed using this function, but
first check if there is a more appropriate function.
}
\examples{
\dontrun{
dat <- data.frame(
 character = c("Title", "Units", "Data1", "Data2", "Data3"),
 address = c("A1", "A2", "B1", "B2", "B3")
 )
remove_metadata_cells(dat, "(?i)units")
}
}
