% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_no_data_rows.R
\name{remove_no_data_rows}
\alias{remove_no_data_rows}
\title{remove_no_data_rows}
\usage{
remove_no_data_rows(df, col_names)
}
\arguments{
\item{df}{Dataframe}

\item{col_names}{character vector between length 1 and 3. The quoted names of
columns used to identify rows that need to be removed}
}
\value{
Character vector or NA.
}
\description{
TBD
}
\details{
Remove rows without data

Return the original dataframe minus the rows that only have NAs in the
columns specified in col_names

Rows are only removed if all col_names have NAs in the same row.
Error raised if input is not a dataframe.
Warning is given if any rows are removed
}
\examples{
test_df <- data.frame("A" = c(1, 2, NA), "B" = c("a", "b", NA))
new_df <- remove_no_data_rows(test_df, c("A", "B"))
print(new_df)

}
