% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot_data.R
\name{remove_non_numeric_rows}
\alias{remove_non_numeric_rows}
\title{Remove rows without a numeric value.}
\usage{
remove_non_numeric_rows(dat, columns_to_create, header_split_to)
}
\arguments{
\item{dat}{dataframe containing the unpivotted data}

\item{columns_to_create}{character string vector of column names.}

\item{header_split_to}{character string vector of column names. Often NA.}
}
\value{
dataframe dat with unwanted rows removed.
}
\description{
When unpivoting data some rows may remain that have been used to
create a column (see explanation below), but are not part of the required
information for that column. This function identifies which they are and
removes them.

The rows to remove are identified as those for which there are NAs in all the
columns that have been created AND the numeric column (we don't want to
accidentally delete data).

Explanation of how the rows that this function removes are created:
Sometimes rows in the xlsx_cells data refer to a column in the raw data
that holds useful information that is not part of the main table, and we want
to include that info in our output. e.g. if a column to the right of the data
contains notes information (as in DLUHC CPR), it should be beheaded in such a
way that it gets associated with a numeric row in the xlsx_cells dataframe:
In the pub sec settings we would use column_to_right_of_data_name_pattern to
associate those character strings to the correct values. This function tidies
up after that association is made, by removing the rows related to that
column from the unpiovotted data.
}
\examples{
\dontrun{
columns_to_create <- "all_headers"
header_split_to <- c("description_1", "description_2")
dat <- data.frame(
  "numeric" = c(1, 2, NA, NA),
  "notes" = c("note 1", "note 2", "note 1", "note 1"),
  "description_1" = c(NA, "this", NA, NA),
  "description_2" = c("this", "that", NA, NA),
  "character" = c(NA, NA, "note 1", "note 2")
)
}
}
