% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot_data.R
\name{remove_remnants_from_behead}
\alias{remove_remnants_from_behead}
\title{Remove info from above the top row of data}
\usage{
remove_remnants_from_behead(dat, first_data_row)
}
\arguments{
\item{dat}{dataframe of unpivotted xlsx_cells data.}

\item{first_data_row}{integer stating the first row of numeric data.}
}
\value{
dataframe. dat with unwanted rows removed.
}
\description{
Because of how the data are imported with every cell in the
Excel data having it's own row in our input, we are left with rows in the
unpivotted data that are no longer needed. These rows describe cells that are
part of the headers, and the info we needed from them has already been
captured. This function removes those cells.
}
\examples{
\dontrun{
dat <- data.frame(character = c("header_1", "a"),
                  row = 1:2,
                  description_1 = c(NA, "a"))
remove_remnants_from_behead(dat, 2)
}
}
