% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_to_multiple_columns.R
\name{remove_repeated_pattern_from_remaining}
\alias{remove_repeated_pattern_from_remaining}
\title{Prevent splits happening twice in the same place}
\usage{
remove_repeated_pattern_from_remaining(
  dat,
  split_point,
  matches_next_pattern,
  whitespace
)
}
\arguments{
\item{dat}{Dataframe output containing columns called 'tmp_remaining'
(contains the string that remains for further splitting), and columns with
the names given by tmp_matches_next_pattern and tmp_whitespace.}

\item{split_point}{character string giving the regex pattern that was last
used to create a split.}

\item{matches_next_pattern}{bool. If FALSE, whitespace will be trimmed from
the start of tmp_remaining but no other changes to tmp_remaining will be done.}

\item{whitespace}{bool. If FALSE (and matches_next_pattern is TRUE), the
whole split_point pattern will be removed from tmp_remaining.}
}
\value{
dat with 'tmp_remaining' column updated
}
\description{
In some cases (described below), the characters that match the
split_point pattern need to be removed from the 'tmp_remaining' column if the
next split_point pattern is the same.
For example if we take the value 'split \emph{newline} this \emph{newline} here'
and the split_patterns are c('newline', 'nelwine'), after the first split
string_1 would be 'split' and 'tmp_remaining' would be
'\emph{newline} this \emph{newline} here'. Without the actions of this function the
second split would result in string_2 being '' and tmp_remaining being
'\emph{newline} this \emph{newline} here',
because it would have split at the first newline.

Use the columns prefixed with 'tmp_whitespace_' (which states whether the
split_pattern in question starts with either whitespace or newline) and
'tmp_matches_next_pattern_' to identify what action needs to be taken.
If tmp_matches_next_pattern is TRUE, AND tmp_whitespace is FALSE the whole
of the split pattern is removed from the start of tmp_remaining. In all
other cases, white space is simply trimmed from the start tmp_remaining.

N.B. A simpler way to do this would just be to remove one or two characters,
from the start of remaining in all cases but we don't do this because doing
so could cause the next pattern to match the beginning of tmp_remaining
rather than the intended match.
}
\examples{
\dontrun{
split_patterns <- c("\n", "\n", "Note", "Note")
dat <- data.frame(
         col_to_split = c("split \n this \n here",
                          "split \n here Note this",
                          "Note this Note that"),
         tmp_use_split_point_1 = c(TRUE, TRUE, FALSE),
         tmp_use_split_point_2 = c(TRUE, FALSE, FALSE),
         tmp_use_split_point_3 = c(FALSE, TRUE, TRUE),
         tmp_use_split_point_4 = c(FALSE, FALSE, TRUE),
         tmp_use_split_point_0 = TRUE,
         tmp_split_1 = c(1, 1, 3),
         tmp_split_2 = c(2, 3, 3),
         tmp_split_3 = c(NA, 3, 3),
         tmp_split_4 = c(NA, NA, 4),
         tmp_string_1 = c("split", "split", "Note this"),
         tmp_remaining = c("\n this \n here", "\n here Note this", "Note that"),
         tmp_matches_next_pattern_1 = c(TRUE, FALSE, TRUE),
         tmp_matches_next_pattern_3 = c(FALSE, FALSE, TRUE),
         tmp_whitespace_1 = c(TRUE, TRUE, FALSE),
         tmp_whitespace_3 = c(FALSE, FALSE, FALSE)
         )

dat2 <- dat \%>\%
    mutate(tmp_remaining = c("\n this \n here", "\n here Note this", "Note that"),
           tmp_string_2 = c("this", "here", "to"),
           tmp_string_3 = c("here", "here", "to"))

# where i is 1
remove_repeated_pattern_from_remaining(dat, "\n", "tmp_matches_next_pattern_1",
                                       "tmp_whitespace_1")
# where i is 3
remove_repeated_pattern_from_remaining(
dat2, "Note", "tmp_matches_next_pattern_3","tmp_whitespace_3"
)
}
}
