% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot_data.R
\name{remove_underscore_lines}
\alias{remove_underscore_lines}
\title{Remove long lines of underscores from headers.}
\usage{
remove_underscore_lines(dat, columns)
}
\arguments{
\item{dat}{dataframe.}

\item{columns}{character string vector. The names of the columns from which
to remove the underscore lines.}
}
\value{
dat dataframe with underscore lines removed.
}
\description{
Long lines of underscores are used in e.g. some Northern Ireland
data to denote a line in one of the headings. This function removes these
underscore lines.
}
\examples{
\dontrun{
dat <- data.frame( description_1 = c("__A___"))
remove_underscore_lines(dat, "description_1")
}
}
