% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_information.R
\name{remove_unwanted_cols}
\alias{remove_unwanted_cols}
\title{Remove unwanted columns from xlsx_cells imported data}
\usage{
remove_unwanted_cols(dat)
}
\arguments{
\item{dat}{dataframe imported using tidyxl::xlsx_cells.}
}
\value{
dat dataframe without columns removed.
}
\description{
xlsx_cells data always has the same column names. Many of these
are not required post processing, so this function removes them if they
exist.
}
\examples{
\dontrun{
dat <- tibble::tibble(
    Year = c(2023, 2023),
    Value = c(1, 2),
    row = c(1, 1),
    local_format_id = NA
)
remove_unwanted_cols(dat)
}
}
