% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_to_multiple_columns.R
\name{remove_unwanted_split_columns}
\alias{remove_unwanted_split_columns}
\title{remove all unwanted columns}
\usage{
remove_unwanted_split_columns(dat)
}
\arguments{
\item{dat}{Dataframe containing columns with the prefixes:
"use_split_point_", "split_", "matches_next_pattern_", or "whitespace_"}
}
\description{
remove all unwanted columns generated by
split_to_mutliple_columns.
}
