% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tables.R
\name{remove_used_tables}
\alias{remove_used_tables}
\title{Clean up list of tables so that each cell is only represented once}
\usage{
remove_used_tables(dat, used_tables, keep = NA)
}
\arguments{
\item{dat}{list of dataframes.}

\item{used_tables}{iteger vector. The indices of the tables to remove from
dat}

\item{keep}{integer vector. The IDs of the tables to keep. Default is NA, in
which case all tables are kept.}
}
\value{
list of dataframes.
}
\description{
Remove tables from the list of tables that are no longer
required. For example, table 1 will always be removed because table 1 is
used to create tables 2 and 3.
}
\examples{
\dontrun{
dat_1 <- data.frame(a = 1)
dat_2 <- data.frame(b = 2)
dat_3 <- data.frame(c = 3)
dat <- list(dat_1, dat_2, dat_3)
remove_used_tables(dat, 1)
}
}
