% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_names.R
\name{rename_columns}
\alias{rename_columns}
\title{rename columns using regular expressions to identify them}
\usage{
rename_columns(dat, exclude_names, patterns, new_names)
}
\arguments{
\item{dat}{dataframe}

\item{exclude_names}{character string. Any column names that should not be
included when looking for matches to the patterns. e.g. If data were
originally imported using xlsx_cells, you may want to exclude the column
names that are automatically generated by importing data with xlsx_cells.}

\item{patterns}{character string vector. One regular expression for each
column to be renamed. Must be given in the corresponding order to new_names.
In pub sec this variable is specified by columns_to_rename_patterns.}

\item{new_names}{character string vector. One name for each column to be
renamed. Must be given in the corresponding order to patterns. In pub sec
this variable is specified by columns_to_rename_names.}
}
\value{
dat with columns renamed
}
\description{
Use regular expressions to identify which columns to rename,
and rename them.
}
\examples{
\dontrun{
dat <- data.frame(
  data_type = "character",
  numeric = 1,
  Type = "A",
  Set = "B"
  )
 rename_columns(
     dat, exclude_names = c("data_type", "numeric"),
     patterns = c("(?i)type", "Set"),
     new_names = c("description_1", "description_2")
     )
}
}
