% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_duplicates.R
\name{rename_identified_duplicates}
\alias{rename_identified_duplicates}
\title{Add specified prefix to entries based on location (1st, 2nd etc)}
\usage{
rename_identified_duplicates(
  dat,
  cols_to_group_by,
  column,
  index,
  prefix,
  expected_freq
)
}
\arguments{
\item{dat}{dataframe. Must include a column with the name given in column.}

\item{cols_to_group_by}{vector of character strings giving the names of
columns that will have been created during un-pivotting. The entries in these
columns would uniquely identify each row if there were no duplicates.}

\item{column}{character string. The name of the column the duplicate is in in
dat. Vector of length 1. In pub sec this variable is specified by
rename_duplicate_in_col.}

\item{index}{integer. Which to rename e.g. '2' would rename the 2nd. Can be a
vector of length > 1, but if so the number of items must match that in
prefix. In pub sec this variable is specified with rename_duplicate_index.}

\item{prefix}{character string. The prefix that is to be pasted onto the
original name. Can be a vector of length > 1, but if so the number of
items must match that in index. In pub sec this variable is specified by
rename_duplicate_prefix.}

\item{expected_freq}{integer. The expected number of repeats of the
column name in the raw data. Vector of length 1. In pub sec this variable is
specified by rename_duplicate_expected_freq.}
}
\value{
dat with duplicates renamed.
}
\description{
Add the specified prefix to the start of the duplicated entries
that are at the specified index location.
}
\examples{
\dontrun{
dat <- data.frame(
  item_detail = rep(c("A", "A", "Total", "C", "Total", "Total"), 2),
  year = rep(c("2021-22", "2022-23"), each = 6),
  numeric = rep(c(2, 3, 5, 0.1, 5.1, 10.1), 2),
  location = rep(c(1, 2, 1, 1, 2, 3), 2),
  rename_required = rep(c(FALSE, FALSE, TRUE), 4)
)

cols_to_group_by <- c("year", "item_detail")

result <- rename_identified_duplicates(
    dat, cols_to_group_by, #
    "item_detail", c(1, 3), c("first", "third"), 3 )
}
}
