% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_names.R
\name{rename_reserved_colnames}
\alias{rename_reserved_colnames}
\title{Rename columns with reserved names}
\usage{
rename_reserved_colnames(dat, reserved_names)
}
\arguments{
\item{dat}{dataframe.}

\item{reserved_names}{vector of character strings. Column names that we
don't want to exist yet in dat (e.g. if they are reserved for some other
future column).}
}
\value{
dat with offending columns given a numeric postscript to the name.
}
\description{
Column names that match a string listed in reserved_names are
given a postscript. This guards against overwriting existing columns with
information we have pulled from outside of the main table, so if there is
a column with the same name as one of the reserved names an _1 postscript is
appended to it.
}
\examples{
\dontrun{
dat <- data.frame("group" = c("Fox", "Bear"),
                  "value" = c("red", "black"),
                  "numeric" = 1:2)
newdat <- rename_reserved_colnames(dat, c("value", "source"))
}
}
