% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_to_multiple_columns.R
\name{rename_split_columns}
\alias{rename_split_columns}
\title{assign new names to columns with the prefix "string_".}
\usage{
rename_split_columns(dat, to)
}
\arguments{
\item{dat}{Dataframe containing the same number of columns starting 'tmp_string_'
as there are elements in \code{to}}

\item{to}{character array of new column names. Must be of the same length as
the number of columns starting 'string_'. The first element of the vector will
be the new name for string_1, the second string_2 etc.}
}
\description{
assign new names to columns with the prefix "string_".
}
\examples{
\dontrun{
dat <- data.frame(tmp_string_2 = "10", tmp_string_1 = "B")
to <- c("letter", "number")
rename_split_columns(dat, to)
}
}
