% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_names.R
\name{rename_value_columns}
\alias{rename_value_columns}
\title{Rename the value columns}
\usage{
rename_value_columns(dat)
}
\arguments{
\item{dat}{(tibble)}
}
\value{
dat with numeric column named 'value' and character column named
non_numeric_value if those columns exist
}
\description{
Standardise column names so that the numeric column gets renamed 'value' and
the character column gets renamed 'non_numeric_value'.
}
\details{
The non-numeric_value column will only be populated if there are
character values in a column that should be numeric in the raw data. For
example if 's' is used to indicate suppressed values in numeric columns,
it will end up in the non_numeric_value column.
}
\examples{
\dontrun{
dat <- tibble(
  item = c("ball", "ball"),
  numeric = c(1, NA),
  character = c(NA, "Small sample size")
)
rename_value_columns(dat)
}
}
