% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_wording.R
\name{replace_blanks}
\alias{replace_blanks}
\title{Fill blanks in one column with values from another}
\usage{
replace_blanks(dat, to_columns = NA, from_columns = NA)
}
\arguments{
\item{dat}{dataframe}

\item{to_columns}{character vector. Regular expression to match the name of
the column in which you want to replace the blanks. Must be the same length
as from_column. In pub sec this variable is specified with
col_pattern_with_blanks_to_replace.}

\item{from_columns}{character vector. Regular expression to match the name of
the column whose values you want to use to replace the blanks in to_column.
Must be the same length as to_column. In pub sec this variable is specified
with col_pattern_to_replace_blanks_with.}
}
\value{
dataframe with blanks replaced
}
\description{
Where there is an NA in a specified column, fill it with the
value from another.
}
\examples{
\dontrun{
dat <- data.frame(
    Code = c(NA, "E08000001"),
    Region_code = c("E92000001", "E12000002"),
    LA = c(NA, "Bolton"),
    Region = c("England", "North West")
    )
replace_blanks(
    dat, c("(?i)^code$", "(?i)la"), c("(?i)region.code", "(?i)region$")
    )
}
}
