% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_wording.R
\name{replace_strings}
\alias{replace_strings}
\title{Replace text in a specified column with different text}
\usage{
replace_strings(dat, column = NA, from_regex = NA, to = NA, keep = FALSE)
}
\arguments{
\item{dat}{DataFrame containing a column with the name given by 'column'}

\item{column}{character string giving name of column that contains the text
to replace. Must be exact. In pub sec this variable is specified by
replace_string_col}

\item{from_regex}{vector of character strings giving the regular expressions
used to identify the locations where replacements should be made. Must be
in the same order as \code{to}. The text identified by the first element in
'from_regex' will be replaced with the text given in the first element of
'to'. In pub sec this variable is specified by
replace_string_from_col_patterns.}

\item{to}{vector of character strings containing the new text.
In pub sec this variable is specified by replace_string_to.}

\item{keep}{boolean. If TRUE make a copy of column and rename it with
'_original' as a postscript. In pub sec this variable is specified by
replace_string_keep_original.}
}
\value{
dat with replacements made.
}
\description{
In very particular cases we may want to change text in a
specified column. In most cases we should do this with user interaction that
ensures the correct replacement is taking place (inb pub_sec this is done in
Python in the pub_sec standardisation module). However, in columns that are
not set to be standardised with user interaction, such as vintage in pub_sec,
we can make text replacements using replace_strings.

Note that the entire text of an identified cell is replaced, not just the bit
identified by the \code{from_regex}.
}
\examples{
\dontrun{
dat <- data.frame("vintage" = c("Near actuals", "Estimates"),
                  "numeric" = c(1, 2))
replace_strings(dat, "vintage", c("(?i)actual", "(?i)est"), c("final", "budget"))
}
}
