% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_handlers.r
\name{select_relevant_args}
\alias{select_relevant_args}
\title{Where args have been split by separator, retain the relevant column}
\usage{
select_relevant_args(dict, part)
}
\arguments{
\item{dict}{dataframe with a column for key and one or more item columns.}

\item{part}{integer. The number of the item column to keep.}
}
\value{
dict with a column for key and a single column for item.
}
\description{
When args are split by separator, they are separated into
columns. This function retains only the relevant column for the table in
question.
}
\details{
See split_args_by_separator for more information.
}
\examples{
\dontrun{
dict <- tibble(
    key = c("var1", "var2"),
    item = list(NA),
    item_1 = c(list(c("a", "b")), list("1")),
    item_2 = c(list("c"), list(NA))
    )
result <- select_relevant_args(dict, 1)
result <- select_relevant_args(dict, 2)
}
}
