% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_information.R
\name{select_units}
\alias{select_units}
\title{Select a single units value}
\usage{
select_units(sheet_units, table_units)
}
\arguments{
\item{sheet_units}{character string. Units taken from the top of the sheet.}

\item{table_units}{character string. Units taken from above the subtable.}
}
\value{
character string. Units from above the subtable unless they are NA,
in which case units from the top of the sheet.
}
\description{
Choose between units taken from the top of the sheet, and units
taken from above the subtable. Units taken from above the subtable take
precendence. If neither are available, return NA.
}
\examples{
\dontrun{
select_units("millions", NA)
select_units("millions", "thousands")
select_units(NA, "thousands")
}
}
