% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_handlers.R
\name{select_year_from_sheet_or_table}
\alias{select_year_from_sheet_or_table}
\title{Select the year from either the sheet or table metadata}
\usage{
select_year_from_sheet_or_table(
  sheet_year,
  table_year,
  prefer_sheet_year = FALSE
)
}
\arguments{
\item{sheet_year}{character vector of years taken from the top of the sheet.}

\item{table_year}{character vector of years taken from above the table.}

\item{prefer_sheet_year}{boolean default is FALSE. If TRUE, and there is a
year given in the information above the sheet, instead of returning the year
from above the table, the year from the top of the sheet will be used.}
}
\value{
single character vector of years.
}
\description{
Select between two provided year vectors. One will be taken from the metadata
at the top of the sheet, the other from the metadata directly above the
table.

If they are the same, just return one. If one is NA, return the other. If
they are different return the one stated by the prefer_sheet_year setting.
}
\details{
Used by get_year_for_use_in_data. Usually only one of sheet_year and
table_year will exist, unless there are multiple tables in a sheet, and each
has the year the data relates to in the table title. In such as case, this
function will by default return the table title year rather that the year at
the top of the sheet. In pub sec, if there is no year column in the source
data, the year from get_year_for_use_in_data will be put in the output year
column.

If the year given in the information at the top of the sheet relates to all
tables, but some other year is mentioned above an individual table, the
prefer_sheet_year setting allows users to specify that the year at the top
of the sheet should be used instead.
}
\examples{
\dontrun{
select_year_from_sheet_or_table(c("2022-23", "2026-27"), "2024-25")
}

}
