% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_global_variables.R
\name{set_global_variables}
\alias{set_global_variables}
\title{set_global_variables}
\usage{
set_global_variables()
}
\description{
Un-pivot data with one row of any number of columns
}
\examples{
start_debug_logfile()
}
\author{
Mark London \email{mark.london@ext.ons.gov.uk}
}
