% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheet_matcher.R
\name{sheet_matcher}
\alias{sheet_matcher}
\title{sheet_matcher}
\usage{
sheet_matcher(matching_sheets)
}
\arguments{
\item{matching_sheets:}{list of strings.}
}
\value{
string
}
\description{
Select a sheet from a list of matching sheets based on a regex pattern.
}
\details{
Select a sheet from a list of matching sheets based on a regex pattern.
}
\examples{
Example 1 case of multiple sheets matching: 
matching_sheets <- list("Sheet1", "Sheet2")

Example 2 case of no sheets matching: 
matching_sheets <- list()

Example 3 case of 1 matching sheet: 
matching_sheets <- list("Scotland")

sheet_matcher(matching_sheets)

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}
}
