% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_helper_functions.R
\name{source_helper_functions}
\alias{source_helper_functions}
\title{source_helper_functions}
\usage{
source_helper_functions()
}
\description{
Un-pivot data with one row of any number of columns
}
\examples{
source_helper_functions()
source_helper_functions:
}
\author{
Mark London \email{mark.london@ext.ons.gov.uk}
}
