% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_handlers.r
\name{split_args_by_separator}
\alias{split_args_by_separator}
\title{Split dictionary of arguments by subtable}
\usage{
split_args_by_separator(dict, separator)
}
\arguments{
\item{dict}{dataframe with columns 'key' and 'item'. 'key' gives the name of
the variable to be, 'item' the value of that variable.}

\item{separator}{character string giving the string that separates args.}
}
\value{
dataframe. dict with arguments split by the separator into new
columns prefixed with 'item_'. The number of 'item_' columns is equal to the
maximum number of identified separators for a single row + 1 (i.e. the number
of tables). If the separator is not present in a row, the same value is
returned in every item_ column.
}
\description{
The dictionary of arguments has a key column that gives the
variable name and an item column that gives the values for each variable.
Some variables can contain values for multiple tables. In such cases, the
arguments for one table will be separated from the arguments for the next
table by a separator string. e.g. if the separator string is 'next_table',
and an arg for table 1 is TRUE, but for table 2 it is FALSE, that variable
will have the item value c(TRUE, 'next_table', FALSE). This function splits
the args for each table into separate columns in the dictionary, based on
the separator string. Where the same value is required for all tables the
separator need not be included - the same value will be given for each table.

So this:
key                    >  item
multitable_arg_separator   > 'next_table'
header_identifier > c('p1', 'next', 'p2')
year_column_name           > "^year$"

is turned into:
key                    >  item                 >  item_1  > item_2
multitable_arg_separator   > 'next'                >  'next'  > 'next'
header_identifier > c('p1', 'next', 'p2') >   'p1'   > 'p2'
year_column_name           > '^year$'              > '^year$' > '^year$'
}
\examples{
\dontrun{
dict <- tibble(
    key = c("var1", "var2", "var 3"),
    item = list(c("a1", "a2", "next", "b", "next", "c"),
                c("d", "next", "e", "next", "f"), ("1"))
    )

result <- split_args_by_separator(dict, separator = "next")
}
}
