% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot_data.R
\name{split_by_consecutives}
\alias{split_by_consecutives}
\title{split a numeric vector into sets of consecutive numbers}
\usage{
split_by_consecutives(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
list where each element of the list is a vector of consecutive
integers.
}
\description{
split a numeric vector into sets of consecutive numbers.
}
\examples{
\dontrun{
x <- c(1, 1, 3, 5, 6, 7, 10, 23, 24)
split_by_consecutives(x)
}
}
