% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_column_by_pattern.R
\name{split_column_by_pattern}
\alias{split_column_by_pattern}
\title{split_column_by_pattern}
\usage{
split_column_by_pattern(dat, original_col, new_col, pattern)
}
\arguments{
\item{dat}{data.frame : A dataset where we split the string column into two columns}

\item{original_col}{(string): It will contain the original string up to the start of the pattern}

\item{new-col}{(string): It will contain the string following the pattern}
}
\value{
split_col_dat: The dataset with a string column split into two columns 
 @examples
 Example where data has two character columns with atleast one pattern
 dataset <- tibble::tibble("info" = c("chair", "chair \r\n (falling)")
                    
  split_column_by_pattern(dat = dataset, original_col = "info", 
  new_col = "direction", pattern = "\r\n\\s*\\(")
}
\description{
split_column_by_pattern- Split a column string into two columns
}
\details{
split_column_by_pattern- Split a column string into two columns
The original_col will contain the original string up to the start of the pattern. 
new_col will contain the string following the pattern. 
If the name of the new column exists, 
then it returns the original data to avoid overwritting column names
The pattern itself will be removed from the output.
If the pattern is not present, new col will be blank
}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
