% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{split_data_from_metadata}
\alias{split_data_from_metadata}
\title{Split metadata from above the data.}
\usage{
split_data_from_metadata(dat, pattern, instance = NA, offset_by = NA)
}
\arguments{
\item{dat}{dataframe imported using xlsx_cells.}

\item{pattern}{character string. A regular expression to search for in the
data to identify the first header row. In pub sec this variable is specified
with header_identifier.}

\item{instance}{An integer specifying which instance of the header identifier
pattern to use. Defaults to 1 (the first instance). In pub sec this variable
is specified with header_identifier_instance.}

\item{offset_by}{An integer specifying the number of rows to offset
the header row. Defaults to 0. In pub sec this variable is specified with
header_row_offset. If the row the pattern is found on is the row after the
first header row, offset_by will be 1. If the pattern is on the row before
the first header row, offset_by will be -1.}
}
\value{
named list: 'data' and 'metadata' are dataframes, 'split_row' is
an integer and is the source data row number of the first header row.
}
\description{
Many Excel datasets have metadata above the data e.g. at the top of the
sheet, or above each table in a multi-table sheet. This function separates
out the metadata from the data.
}
\details{
Metadata may contain information such as the title, units, dates, and notes.
These may be required for output data. However, even if the metadata is not
needed, it must be removed before the data can be un-pivotted using behead.
}
\examples{
\dontrun{
dat <- data.frame(
    address = c("A1", "A2", "A3", "A4"),
    row = 1:4,
    col = 1,
    character = c("Some data", NA, "value", NA),
    numeric = c(NA, NA, NA, 10)
    )

split_data_from_metadata(dat, "(?i)value", NA, NA)
}
}
