% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sheet.R
\name{split_data_from_sheet_info}
\alias{split_data_from_sheet_info}
\title{Split metadata at the top of the sheet from the rest of the data.}
\usage{
split_data_from_sheet_info(dat, pattern, instance, offset_by)
}
\arguments{
\item{dat}{dataframe imported using xlsx_cells.}

\item{pattern}{character string. A regular expression to search for in the
data to identify the first header row. In pub sec this variable is specified
with header_identifier.}

\item{instance}{An integer specifying which instance of the header identifier
pattern to use. Defaults to 1 (the first instance). In pub sec this variable
is specified with header_identifier_instance.}

\item{offset_by}{An integer specifying the number of rows to offset
the header row. Defaults to 0. In pub sec this variable is specified with
header_row_offset. If the row the pattern is found on is the row after the
first header row, offset_by will be 1. If the pattern is on the row before
the first header row, offset_by will be -1.}
}
\value{
named list of dataframes. Tables are called 'data' and 'metadata'.
}
\description{
Many Excel datasets have metadata at the top of each sheet giving, for
example, the title, units, and notes. This function separates out the sheet
metadata from the rest of the sheet.
}
\details{
Where sheets contain multiple tables, each of the sub tables may have their
own blocks of metadata. This function \emph{does not} remove the metadata from
subtables.
}
