% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_info.R
\name{split_date_to_columns}
\alias{split_date_to_columns}
\title{Split a date formatted column into columns for day, month, and year}
\usage{
split_date_to_columns(dat, column)
}
\arguments{
\item{dat}{dataframe that contains the column name given by column.}

\item{column}{character string. The exact name of the column containing
the date. In pub_sec this variable is specified by POSIX_column.}
}
\value{
datafram. dat with columns day, month, and year added, and
column removed.
}
\description{
Where a column can be unambiguously converted to a POSIXt
object, split the column out into day, month, and year columns. If it cannot
be converted, an error is raised.
}
\details{
The information in neonscience.org/resources/learning-hub/tutorials/
dc-convert-date-time-posix-r was used to create code that extracts year and
day.
}
\examples{
\dontrun{

 dat <- data.frame(
   period = c("2017-01-08", "2017-08-14"),
   value = c(1:2)
 )

 split_date_to_columns(dat, "period")
}
}
