% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_table.R
\name{split_table}
\alias{split_table}
\title{split_table}
\usage{
split_table(dat, table_identifier)
}
\arguments{
\item{dat}{A tidyxl dataframe containing the main data table to be processed.}

\item{table_identifier}{string. Regex pattern}
}
\value{
data frame (tibble) containing the split data tables. These tables 
are held in the output dataframe as separate tables within a column called
'cells'
}
\description{
Split a table into component tables
}
\details{
Split a table into component tables

Used for cases where there are several tables within one tab of a 
spreadsheet. Splits by row based on a regex pattern (table_identifier).

Reformat these data so that there is only one number per row.
}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
