% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tables.R
\name{split_tables}
\alias{split_tables}
\title{split an xlsx_cells dataframe into a list of dataframes.}
\usage{
split_tables(dat, pattern, instance, offset_by, direction, table_id)
}
\arguments{
\item{dat}{list of dataframes originally created from a single dataframe
imported using tidyxl::xlsx_cells().}

\item{pattern}{character string. Regular expressions that first
matches a string in the first column of the second table.}

\item{instance}{integer or integer as character string. The instance of the
pattern match (1st 2nd, 3rd etc).}

\item{offset_by}{integer or integer as character string. The number of rows
or columns to subtract from the row/column on which the pattern match was
found in order to get to the row/column of the split.}

\item{direction}{character string. Either 'col' if the table is to be split
by column, or 'row' if it is to be split by row.}

\item{table_id}{integer. The full table is id 1. Table 1 will be
split into table 2 and 3. If split direction is col, table 2 is to the left,
table 3 to the right. If split direction is row, table 2 is at the top, table
3 below.}
}
\value{
A list of dataframes, each containing a subset of dat corresponding
to a table.
}
\description{
Split one of the dataframes in a list of dataframes in two. The
split is done either by row or by column (specified by direction), where
the first match to pattern is found in the first row (if direction is row) or
column of the 2nd of the new dataframes.

e.g. where this is the dataframe:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ Col 1 > col 2 > col 3 > col 4 > col 5
}\if{html}{\out{</div>}}

row 1      >   A   >       >       >  B
row 2   C1 >   1   >       >  D1   >  3
row 3   C2 >   2   >       >  D2   >  4

Split would be by column because the tables are side-by-side not one above
the other.
pattern could be "D", because this is first found in the first column of the
second table (the table to the right).
}
\examples{
\dontrun{
dat <- data.frame(
  "row" = c(1, 1, 1, 2, 2, 2),
  "col" = c(1, 2, 3, 1, 2, 3),
  "address" = c("A1", "B1", "C1", "A2", "B2", "C2"),
  "character"= c(NA,NA,NA,"table 1","NA","table 2")
)
split_tables(list(dat), "table 2", "col", 1)
}
}
