% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_to_multiple_columns.R
\name{split_to_multiple_columns}
\alias{split_to_multiple_columns}
\title{Split the contents of one column over multiple columns}
\usage{
split_to_multiple_columns(dat, from, to, split_point_descriptions)
}
\arguments{
\item{dat}{Dataframe containing a column to split.}

\item{from}{character string. The name of the column to split.}

\item{to}{vector of strings. The new names to put the split in.}

\item{split_point_descriptions}{vector of strings that code for regular
expressions. Each one will only be matched once, so any repeats must be
explicit.}
}
\description{
Sometimes we may want to split a single character column into
multiple columns.
}
\details{
The layout of a dataset may change so that it has only
one row of headers that contains all the information that was once in
multiple headers. Use this function to create the output as it would
have been when the input had multiple headers. This may be required as a
quick fix if the output has to remain stable. However, it is recommended to
not be used as a long term fix - the datset with fewer columns is to be
preferred as it easier and less error prone to join columns than to split
them.

Not every split_point has to be present in every string to be split - if
a pattern does not exist in a given string, the previous value is used (see
example below). In other words, if the split is not consistent across all
data, split points are given for the element that requires the most splits.

split_patterns are regular expressions used to determine where each split
should occur. The split_point pattern is not removed from the result, but
any separators (new lines, :, and - ) are.

The order split_patterns are given in must match the order the patterns occur
in.

If no splits patterns are found for a given item, the whole of the from
column will be given in every to column.

Note: The patterns stated MUST appear in all strings in the order they are
given so should be as specific as possible. For example, given the string
'A: B- C' if the split points are given as '-' then ':', the string will be
split only once, at the '-' giving 'A: B' and 'C'.
}
\examples{
\dontrun{
If split_points are '\n1.','\n2.', '\n3.', missing parts are handled thus:

| col_to_split | description_1 | description_2 | description_3 | units    |
|--------------|---------------|---------------|---------------|----------|
| split \n     |               |               |               |          |
| 1.this \n    |  split        | this          | one           | thousand |
| 2.one \n     |               |               |               |          |
| 3.thousand   |               |               |               |          |
|--------------|---------------|---------------|---------------|----------|
| split \n     |               |               |               |          |
| 2.one \n     |  split        | split         | one           | thousand |
| 3.thousand   |               |               |               |          |
|--------------|---------------|---------------|---------------|----------|
| split \n     |               |               |               |          |
| 1.this \n    |  split        | this          | this          | thousand |
| 3.thousand   |               |               |               |          |
|--------------|---------------|---------------|---------------|----------|
| split \n     |               |               |               |          |
| 1.this \n    |  split        | this          | one           | one      |
| 2.one        |               |               |               |          |
|--------------|---------------|---------------|---------------|----------|
| split \n     |  split        | split         | split         | thousand |
| 3.thousand   |               |               |               |          |
|--------------|---------------|---------------|---------------|----------|
| split \n     |               |               |               |          |
| 2.one \n     |  split        | split         | one           | one      |
|--------------|---------------|---------------|---------------|----------|
| 1. this \n   |  this         | this          | this          | thousand |
| 3. thousand  |               |               |               |          |
|--------------|---------------|---------------|---------------|----------|
|  other       |  other        |  other        |  other        | other    |
|--------------|---------------|---------------|---------------|----------|
| wrong \n     |  wrong \n     | 1. matches    | 1. matches    |1. matches|
| 2. order \n  |  2. order     |               |               |          |
| 1. matches   |               |               |               |          |
|--------------|---------------|---------------|---------------|----------|


from <- "col_to_split"
to <- c("description_1", "description_2", "description_3", "units")
split_point_descriptions <- c(
"ALT_newline_whitespace_=1", "ALT_newline_whitespace_=2",
"ALT_newline_whitespace_=3"
)

dat <- data.frame(
  "col_to_split" = c("split:\n 1.this \n 2.one \n 3.thousand",
                     "split-\n 2.one \n 3.thousand",
                     "split\n 1.this \n 3.thousand",
                     "split\n 1.this \n 2.one",
                     "split\n 3.thousand",
                     "split\n 2.one",
                     "1. this \n 3. thousand",
                     "other",
                     "wrong\n 2. order \n 1. matches"
                     ),
  "value" = c(1:9)
)

split_to_multiple_columns(dat,
                         from = "col_to_split",
                         to = c("description_1", "description_2",
                         "description_3", "units"),
                         split_point_descriptions = split_point_descriptions)
#----------------------------------------------------------------------------
# Example 2

split_point_descriptions <- c(
    "ALT_newline", "ALT_newline", "ALT_=Note", "ALT_=Note"
    )
#split_patterns <- c("\n", "\n", "Note", "Note")
dat <- data.frame(
         col_to_split = c("split \n this \n here",
                          "split \n here Note this",
                          "to Note this Note that",
                          "Note this Note that")
         )
split_to_multiple_columns(
dat, "col_to_split", c("A", "B", "C", "D", "E"), split_point_descriptions
)
}
}
