% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_info.R
\name{split_year_and_vintage}
\alias{split_year_and_vintage}
\title{Separate year and vintage found in the same column into new columns}
\usage{
split_year_and_vintage(dat, pattern)
}
\arguments{
\item{dat}{Dataframe.}

\item{pattern}{character string. A single regular expression to match the
column containing the year and the vintage. Year can be financial or
calendar. Accepted vintages are budget|provisional|forecast|final (not case
sensitive).}
}
\value{
dat dataframe with new columns 'year' and 'vintage'. If either of
these columns already existed, the originals are renamed to '_year' and
'_vintage'
}
\description{
In the rare case that vintage and year are given in the same
column, separate them out into separate columns called 'year' and 'vintage'
If those columns already exist, rename the originals so they don't get
overwitten.
}
\examples{
\dontrun{
dat <- data.frame(
     year_and_vintage = c("2023 final", "2024 provisional", "2025 budget"),
     year = c("2023", NA, NA),
     vintage = "final"
     )
split_year_and_vintage(dat, "year.*vintage")
}
}
