% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_wording.R
\name{standardise_mentioned_years}
\alias{standardise_mentioned_years}
\title{standardise the wording of years in specified columns}
\usage{
standardise_mentioned_years(dat, columns = NA, year)
}
\arguments{
\item{dat}{dataframe}

\item{columns}{character string vector. Each element must match the name of
a column in dat. These are the columns in which you want the replacement to
be made. In pub sec this variable is specified by
descriptors_to_standardise_year_in.}

\item{year}{character string. A single year, either financial or calendar.}
}
\value{
dataframe dat with replacements made. NOTE: in the unlikely case
of an entry with both a calendar year to replace and a financial year that
includes that calendar year. e.g. '2022 and 2022-23' will change to 'current
year and current year-23'. However '2022 and 2023-24' would be fine. As it
is currently written, this function will also only replace financial years
that are in the same format as the supplied year i.e. '2024-2025' would not
get replaced if year was supplied as '2025-26'.
}
\description{
Where the given year is mentioned in a specified character column, replace
it with standardised wording. This wording differs depending on whether the
year is on a financial or calendar year basis.
If the year itself is found it is replaced replaced with "current year"/
"current financial year". If the prior year is found it is replaced with
"previous year"/ "previous financial year". If the subsequent year is found
it is replaced with "next year"/ "next financial year".
}
\examples{
\dontrun{
dat <- data.frame(
    desc_1 = c("Total 2022-23", "Total 2023-24", "Total 2024-25", "Total"),
    desc_2 = c("2022-23", "2023-24", "2024-25", "Total"),
    original_1 = c("Total 2022-23", "Total 2023-24", "Total 2024-25", "Total")
    )
standardise_mentioned_years(dat, c("desc_1", "desc_2"), "2023-24")
}
}
