% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_year_column.R
\name{standardise_year}
\alias{standardise_year}
\title{Standardise year column so that financial years are yyyy-yy}
\usage{
standardise_year(dat, column = NA)
}
\arguments{
\item{dat}{dataframe with a column of the name given in \code{column}.}

\item{column}{character string. The name of the column holding years to
convert.}
}
\value{
dataframe dat with a column called year containing the same as
\code{column} but with financial years standardised.
}
\description{
Where financial years are not in the standard format of
YYYY-YY, convert them to this format. Accepted input formats are:
YYYY_YY or YYYY_YYYY and any number of spaces either side of the underscore,
YYYY-YY or YYYY-YYYY and any number of spaces either side of the dash,
YYYYtoYY, and YYYY_to_YY.

Calendar years (YYYY) remain as they are.
}
\examples{
\dontrun{
dat <- data.frame(fy = c("2021_22", "2020", "now 2022-2023", "other", "2021-22 to 2022-23"))
standardise_year(dat, "fy")
}
}
