% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_debug_logfile.R
\name{start_debug_logfile}
\alias{start_debug_logfile}
\title{start_debug_logfile}
\usage{
start_debug_logfile()
}
\description{
Un-pivot data with one row of any number of columns
}
\examples{
start_debug_logfile()
}
\author{
Mark London \email{mark.london@ext.ons.gov.uk}
}
