% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sheet.R
\name{tidy_sheet}
\alias{tidy_sheet}
\title{Runner for tidying messy Excel data.}
\usage{
tidy_sheet(
  input_filepath,
  tab_regex,
  output_filepath,
  settings,
  file_part = 1,
  to_csv = TRUE
)
}
\arguments{
\item{input_filepath}{character string. Path to the xlsx/xlsm source data.}

\item{tab_regex}{character string. Regular expression to match the sheet name.}

\item{output_filepath}{character string. Path to save the output csv to.}

\item{settings}{flattened dictionary e.g. from json.dumps (see example).
Must contain only the relevant variables (settings) for the specific
dataset. The key is the name of the variable, and the value is the value.
Where a setting does not have a value in the settings it is assigned NA.}

\item{file_part}{integer. Defaults to 1. If a dataset is in two files, each
with the same layout, the one that comes second will have file_part = 2.}

\item{to_csv}{boolean. Default is TRUE. If FALSE the output is returned
rather than being saved to csv.}
}
\value{
saved csv file. File is saved to the specified output_filepath,
and contains tidy data. Data has only one numeric value per row, with all
other columns holding descriptive data for that value. Metadata are added to
the output as columns including year, year_type, supplier, source, dataset,
title, units, and vintage.
}
\description{
Runner script for pre-processing xlsx data so that it is in a
tidy layout, with years formatted to be YYYY or YYYY-YY, vintage added as
budget, provisinal, or final, metadata added as columns, and options to add
or remove columns, and edit specified text.
}
\details{
See the Wiki at https://github.com/ONSdigital/tidysheet for settings details.

filepath must only contain a single year. Where the data are for
a single year/financial year, the year in filepath should be the year the
data refer to (i.e. not the publication date). The year given in the filepath
must be the correct format (financial or calendar year).
}
\examples{
\dontrun{
tidy_sheet(
    "D:/some_file-2023_24.xlsx", "(?i)sheet1", "D:/some_file-2023_24.csv",
    "{header_identifier: (?i)proportion, columns_to_create: desc1, desc2}",
    "1", FALSE
    )
}
}
