% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sheet.R
\name{tidy_sheet_via_subprocess}
\alias{tidy_sheet_via_subprocess}
\title{Runner for tidy_sheet via python}
\usage{
tidy_sheet_via_subprocess(arg_values, to_csv = TRUE)
}
\arguments{
\item{arg_values}{character string vector. The values for the following
(order is important): "--args" (optional), input_filepath, sheet_name,
output_filepath, settings, file_part. See example and the param descriptions
in tidy_sheet for details.}

\item{to_csv}{boolean. Default is TRUE. If FALSE the output is returned
rather than being saved to csv.}
}
\value{
saved csv file. File is saved to the specified output_filepath,
and contains tidy data. Data has only one numeric value per row, with all
other columns holding descriptive data for that value. Metadata are added to
the output as columns including year, year_type, supplier, source, dataset,
title, units, and vintage.
}
\description{
Wrapper allowing tidy_sheet to be run in Python via
subprocess.run()
}
\examples{
\dontrun{
# If picking up args from the command line:
arg_values <- commandArgs(trailingOnly = TRUE)

# Without using the command line, this would look like e.g.:
arg_values <- c(
    "D:/some_file-2023_24.xlsx", "(?i)sheet1", "D:/some_file-2023_24.csv",
    "{header_identifier: (?i)proportion, columns_to_create: desc1, desc2}",
    "1"
    )
serializeJSON(arg_values)
tidy_sheet(arg_values, FALSE)
}
}
