% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysheet-package.R
\docType{package}
\name{tidysheet-package}
\alias{tidysheet}
\alias{tidysheet-package}
\title{tidysheet: tidysheet}
\description{
Turn messy excel data into tidy data ready for processing.
}
\keyword{internal}
