% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_to_multiple_columns.R
\name{trim_separators}
\alias{trim_separators}
\title{clean the start and end of strings}
\usage{
trim_separators(string, separators)
}
\arguments{
\item{string}{character string.}

\item{separators}{vector of character strings, each of which is a regular
expression.}
}
\value{
vector of strings. Original vector with trailing separators removed
}
\description{
Remove hyphens, colons etc, and spaces from the start and end of
strings
}
\examples{
\dontrun{
example_vector <- c("these - ", "hyphens-", " and colons :", ":now   gone- ")
trim_separators(example_vector, "-|:")
}
}
