% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_to_multiple_columns.R
\name{trim_separators_from_new_cols}
\alias{trim_separators_from_new_cols}
\title{Remove colons and dashes from the 'tmp_string_' columns.}
\usage{
trim_separators_from_new_cols(dat)
}
\arguments{
\item{dat}{Dataframe containing columns with prefix "tmp_string_"}
}
\description{
Remove colons and dashes from the 'tmp_string_' columns.
}
\examples{
\dontrun{
dat <- data.frame(tmp_string_1 = "-Remove dash -",
                  tmp_string_2 = ": Remove_colon:",
                  tmp_string_3 = "- remove both :",
                  other = ":remove none-")

trim_separators_from_new_cols(dat)
}
}
