% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot_data.R
\name{unpivot_right_block}
\alias{unpivot_right_block}
\title{Behead the main block of data (not including row names).}
\usage{
unpivot_right_block(dat, columns_to_create)
}
\arguments{
\item{dat}{dataframe. Data imported using tidyxl::xlsx_cells. Data should not
include the rows containing row name information (character columns to the
left of numeric columns in Excel).}

\item{columns_to_create}{character string vector. The names of the columns
that information in the headers will be unpivotted to.}
}
\value{
dataframe that is an unpivotted version of the numeric columns in
the Excel data.
}
\description{
xlsx_cells imports data so that each row describes a single cell
in an xlsx table. This function re-associates cells with their column headers
in a tidy format with one numeric value per row. Each row of Excel column
headers are put into columns whose names are specified by columns_to_create.

This function assumes that column headers will have the following
type of relationship to the data: In the final header row, the header sits
directly above the data, and every data column has a cell directly above it
with header info in it. All header rows above that are either directly above
it or staggered to the left. These can be merged cells.

e.g.

A1 >    >
--- > ---> ---
b1 >    > b2
--- > ---> ---
c1 > c2 > c3
--- > ---> ---
10 > 20 > 30

becomes:

numeric > description_1 > description_2 > description_3
------- > ------------- > ------------- > -------------
10    >     A1        >    b1         >   c1
20    >     A1        >    b1         >   c2
30    >     A1        >    b2         >   c3
}
\examples{
\dontrun{
dat <- data.frame(
  address = c("B1", "C1", "B2", "C2"),
  row = c(1, 1, 2, 2),
  col = c(2, 3, 2, 3),
  data_type = c(rep("character", 2), rep("numeric", 2)),
  numeric = c(rep(NA, 2), 100, 22),
  character = c("primary", "secondary", NA, NA))

unpivot_right_block(dat, "description_1")
}
}
