% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_period_columns.R
\name{update_quarter}
\alias{update_quarter}
\title{Standardise quarter for data on a financial year basis}
\usage{
update_quarter(
  dat,
  financial_year = FALSE,
  q1_is_jan_to_mar = FALSE,
  quarter_col_name = NA
)
}
\arguments{
\item{dat}{dataframe}

\item{financial_year}{boolean. If NA, defaults to FALSE. If TRUE, and
q1_is_jan_to_mar is TRUE, update the quarters so that Q1 is apr-jun. In pub
sec this is set using calendar_year_to_fy_start.}

\item{quarter_col_name}{character string. The name of the quarter column.}
}
\value{
data with apr-jun as quarter 1 if q1_is_jan_to_mar is TRUE and
financial year is TRUE. Otherwise dat is returned unchanged.
}
\description{
If Q1 is jan to march, update the quarters accordingly.
}
\details{
As standard Q1 should be Jan to march for data on a calendar year
basis, but Apr - Jun for data on a financial year basis. At least one source
in the public sector project has Q1 as jan to Mar in data used for
financial year analysis. This function allows for correction of the quarter
number in such situations.
}
\examples{
\dontrun{
dat <- data.frame(
    months = c("jan-mar", "apr-jun", "jul-sep", "oct-dec"),
    quarter = c("1", "2", "3", "4"),
    year = "2024",
    fy_start = c("2023", "2024", "2024", "2024")
    )
update_quarter(dat, TRUE, TRUE, "quarter")
}
}
