% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_to_multiple_columns.R
\name{update_remaining}
\alias{update_remaining}
\title{Remove information that has already been captured from 'tmp_remaining'}
\usage{
update_remaining(dat, split_count, current_string_col, index)
}
\arguments{
\item{dat}{Dataframe output containing a column called 'tmp_remaining' that
contains the string that remains for further splitting, and
a column with the name given by current_string_col.}

\item{split_count}{integer. The number of split_patterns patterns.}

\item{current_string_col}{character. Name of column  with prefix 'string_'
This column contains at least part of what is contained in 'tmp_remaining'.}

\item{index}{integer - iterator from the calling funciton.}
}
\value{
dat with current_string_col info removed from the string in
'tmp_remaining' unless that would leave 'tmp_remaining' blank, in which case
it is left as it is.
}
\description{
Once a 'tmp_string_' column has been created, the information it
contains is no longer needed for future 'tmp_string_' columns.
This function removes the character string in the current tmp_string column
from the 'tmp_remaining' column. However, we don't want 'tmp_remaining'
to ever be blank because we want all tmp_string_ columns filled even when not
all split_point patterns have a match. Therefore if there is no string left
for further splitting, string_ is carried forward into tmp_remaining.
}
\examples{
\dontrun{
split_patterns <- c( "(\n)\\\\s*1\\\\.", "(\n)\\\\s*2\\\\.", "(\n)\\\\s*3\\\\.")

dat <- data.frame(
  tmp_remaining = c("split\n 1.this \n 2.one \n 3.thousand", "other"),
  tmp_use_split_point_0 = TRUE,
  tmp_split_1 = c(1, NA),
  tmp_matches_next_pattern_1 = c(FALSE, FALSE),
  tmp_whitespace_1 = c(TRUE, FALSE),
  value = c(100, 300),
  tmp_string_1 = c("split", "other")
)

update_remaining(dat, 3, "tmp_string_1", 1)
}
}
