% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_to_multiple_columns.R
\name{update_start_of_string}
\alias{update_start_of_string}
\title{Flag strings where the pattern is found at the start of the string.}
\usage{
update_start_of_string(dat, split_patterns, from)
}
\arguments{
\item{dat}{dataframe containing the column specifed by 'from' and columns
prefixed with tmp_use_split_point and tmp_split.}

\item{split_patterns}{character vector of regular expressions.}

\item{from}{character. The name of the column whose entries are to be split.
Must be the name of a character column.}
}
\value{
dat, with an additional boolean column (tmp_pattern_at_start).
If tmp_pattern_at_start is true, the string in 'from' is modified to begin
with an underscore (or if the split_pattern starts with an underscore, a
hyphen).
}
\description{
If the first split pattern that is found in the 'from' column is
right at the start of the string, we need to act like that pattern is not
there so that we don't end up with a blank in the first column.

This function checks if the first matching pattern is at the start of the
'from' string. If it is at the start, the pattern to be ignored is flagged
as such in a new column called 'tmp_pattern_at_start'.  An underscore is also
added to the 'from' string where tmp_pattern_at_start is TRUE (if it already
starts with an underscore a hyphen is added in case of patterns beginnging
'any number of underscores'). This addition of a hyphen or underscore is to
prevent errors when calling update_remaining(): Without there being something
to the left of the pattern match in the \code{from} string, we would get
'empty search patterns are not supported' errors and the 'tmp_string_'
columns would remain empty.

How this info is used in later functions:
Instead of using the pattern that matches the start of the string, the next
matching pattern is used for that split. The relevant tmp_split_ (which gives
the pattern index) and tmp_use_split_ columns are updated accordingly.

N.B. Known issues
\itemize{
\item This does not cover the potential situation where there
is a pattern match at the start of the string, and the next split_patterns
pattern is a copy of that pattern. However this is very unlikely to happen
and would increase the complexity of the code even more, so we will leave it
until it is an issue.
\item It also only anchors the first pattern if a pattern contains an OR (|), so
if one of the patterns after the | is found anywhere in the string an
underscore will be added to the start of the string. This is not a major
issue though.
}
}
\examples{
\dontrun{
dat <- data.frame(
    col_to_split = c("1 A 2 B 3 C 4", "B 3 C 4", "C 4"),
    tmp_use_split_point_1 = c(TRUE, FALSE, FALSE),
    tmp_use_split_point_2 = c(TRUE, TRUE, FALSE),
    tmp_use_split_point_3 = c(TRUE, TRUE, TRUE),
    tmp_split_1 = c(1, 2, 3),
    tmp_split_2 = c(2, 2, 3),
    tmp_split_3 = c(3, 3, 3)
    )
split_patterns <- c("A", "B", "C")

update_start_of_string(dat, split_patterns, "col_to_split")
}
}
