% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_duplicates.R
\name{void_wrong_frequency_targets}
\alias{void_wrong_frequency_targets}
\title{Avoid adding a prefix when there is an unexpected number of repeats.}
\usage{
void_wrong_frequency_targets(dat, column, expected_freq)
}
\arguments{
\item{dat}{dataframe. Must include a column with the name given in column.}

\item{column}{character string. The name of the column the duplicate is in in
dat. Vector of length 1. In pub sec this variable is specified by
rename_duplicate_in_col.}

\item{expected_freq}{integer. The expected number of repeats of the
column name in the raw data. Vector of length 1. In pub sec this variable is
specified by rename_duplicate_expected_freq.}
}
\value{
dat with 'rename_required' updated
}
\description{
Update rename_required to be FALSE if the number of expected
repeats does not matche the number of repeats found.

Warn the user if there are either too few or too many repeats and flag these
as not to be renamed (void them).

The user needs to be warned if there are more columns with a repeated name
than usual, because the ones that get renamed are specified based on there
being the expected number of duplicates E.g. Take a case where there are
usually two columns called 'total', and the user wants to rename the second
(index = 2). Imagine that one year there are three columns called total,
with a new total column falling before the one we usually rename. In this
case, we would otherwise end up renaming the 'new' total column, not the one
we intended, which is now the third.
}
\examples{
\dontrun{
dat <- data.frame(
  "item_detail" = c("Total", "Total", "Total", "Other"),
  "numeric" = c(2, 3, 5, 10),
  "matches_regex" = c(rep(TRUE, 3), FALSE),
  "location" = c(1, 2, 3, 1),
  "rename_required" = c(TRUE, FALSE, FALSE, FALSE)
)
void_wrong_frequency_targets(dat, "item_detail", 2)
}
}
