% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_period_columns.R
\name{add_time_period_columns}
\alias{add_time_period_columns}
\title{Add quarter, fy_start, and year columns.}
\usage{
add_time_period_columns(
  dat,
  quarter_from_col_pattern,
  quarter_col_name,
  year_from_pattern,
  fy_from_fy_end,
  fy_start_from_fy_end,
  fy_end_pattern,
  calendar_year_to_fy_start,
  q1_is_jan_to_mar,
  month_col_pattern,
  year_col_pattern,
  single_year_of_data,
  year_for_column,
  single_year_overrides_all,
  multi_year_range_is_not_valid
)
}
\arguments{
\item{dat}{dataframe.}

\item{quarter_from_col_pattern}{character string. A regular expression that
only matches the column containing quarter information (e.g. '2025 Q1',
'quarter 1', or '1').}

\item{quarter_col_name}{character string. The name of the column that quarter
will be put in.}

\item{year_from_pattern}{character string. Regular expression that matches
the name of the column containing year. This column may contain extra
information alongside year, such as quarter. Required unless fy_from_fy_end
is provided.}

\item{fy_from_fy_end}{boolean. Default is FALSE. TRUE if the financial
year is to be calculated from the financial year end.fy_end_pattern must be
supplied if TRUE.}

\item{fy_start_from_fy_end}{boolean. Default is FALSE. TRUE if the financial
year start is to be calculated from the financial year end.}

\item{fy_end_pattern}{characters string. Regular expression that matches the
name of the column containing the financial year end. Default is NA}

\item{calendar_year_to_fy_start}{boolean. TRUE if calendar year is to be
converted to financial year start. Default is FALSE.}

\item{q1_is_jan_to_mar}{boolean. TRUE if Jan to Mar is Q1. FALSE if Apr to
Jun is Q1. Default is FALSE.}

\item{month_col_pattern}{character string. Regular expression used to
identify the column that contains only the month.}

\item{year_col_pattern}{character string. The name of the year column
provided in settings. If this is NA, the year column is identified
automatically by get_year_column_names().}

\item{single_year_of_data}{boolean or NA. If TRUE create a column called year
and populate it with single_year. Unless single_overrides_all is TRUE, this
will only be used as a last option.}

\item{year_for_column}{list of length 2: 'year' is a character string
containing a valid year; 'warn' is boolean. TRUE if a warning needs to be
raised because of inconsistent years from different locations. This is only
used if use_single_year is TRUE. If so, the 'year' column is populated with
this year value. Must be specified if use_single_year is TRUE. In pub sec
this variable is taken either from the file name (which gets its year or from
information above the data.}

\item{single_year_overrides_all}{boolean or NA. If TRUE, and
single_year_of_data is TRUE, single_year overrides any columns found with
year in the name. If FALSE or NA, other options for filling the year are used
first and single year is only used as a final option.  N.B If
single_overrides_all is TRUE we would not expect year_col_pattern to have
been specified (unless it is also used in other functions).}

\item{multi_year_range_is_not_valid}{bool. If NA or NULL defaults to FALSE
If TRUE, and the year is a range of years e.g '2021-22 to 2023-24' or
'2023, 2024', the year column will show NA. Whether it is TRUE or FALSE,
year_type will be NA for invalid year entries.}
}
\value{
dataframe with quarter, fy_start, and year columns.
}
\description{
Get quarter from a column containing more than just quarter, get fy_start
from the available information on year, and select the correct information
for the year column (see documentation for process_year_column).
}
\details{
Some functions that deal with time periods (split_date_to_columns,
and split_year_and_vintage) are done prior to this in fill_missing_info.
For more details see the documentation for add_quarter_column,
get_fy_start, process_year_column, and fill_missing_fy_start.
}
