% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dropdown.R
\name{check_dropdown}
\alias{check_dropdown}
\title{check_dropdown}
\usage{
check_dropdown(dropdown_regex, info_above_table)
}
\arguments{
\item{dropdown_regex}{string. A regex pattern that should match only the 
dropdown cell.}

\item{info_above_table}{tidyxl::xlsx_cells dataframe that contains the dropdown#'}
}
\value{
message or warning
    
info_above_table <- data.frame(
  row = c(1:3),
  col = rep(1,3),
  character = c("Title", "some other info", "England")
) 

check_dropdown("Eng", info_above_table) # found
check_dropdown("Wales", info_above_table) # not found
}
\description{
Send check message about string is present in the character column
}
\details{
Send check message about  string is present in the character column

Data imported with tidyxl::xlsx_cells have one row for each cell. As such
the usual ways of finding duplicated rows in the raw data (e.g `unique` or 
`distinct`) do not work. 

Given a row number, find any rows that are duplicates of that row.
}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
