% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_reserved_words.R
\name{check_reserved_words}
\alias{check_reserved_words}
\title{check_reserved_words}
\usage{
check_reserved_words(check_nesting_dict)
}
\arguments{
\item{nesting_dict}{A dictionary containing the sheet structure settings.}
}
\value{
A logical value indicating if reserved words are used in the sheet_structure settings.
}
\description{
checks if any reserved words are used, halts if they are
}
\details{
Check for reserved words in sheet_structure settings

This function checks if any reserved words are used in the sheet_structure settings.
Reserved words can cause issues and should be avoided.
These keys need to avoid the reserved words, however, the function is written
to look at the items only, so new keys can be added.
group_col, nested_column_1 , nested_column_2, group_row, nested_row_1, 
nested_row_2, left_headers
Returns true if not, else halts.
}
\examples{
check_reserved_words(nesting_dict_without_reserved)

}
\author{
Mark London \email{mark.london@ext.ons.gov.uk}
}
