% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_information.R
\name{clean_xlsx_cells_data}
\alias{clean_xlsx_cells_data}
\title{Remove unwanted cells from xlsx_cells data}
\usage{
clean_xlsx_cells_data(
  dat,
  cells_to_remove = NA,
  input_filepath = NA,
  hidden_strings = NA
)
}
\arguments{
\item{dat}{data that has been imported using xlsx_cells.}

\item{cells_to_remove}{character string.}

\item{input_filepath}{character string.}

\item{hidden_strings}{bool. TRUE, FALSE or NA. Default is NA. FALSE leads to
the same behaviour as NA. In pub sec this variable is specified by
remove_hidden_strings_bool.}
}
\value{
dataframe (dat) with unwanted information removed
}
\description{
Remove cells using specified cell addresses and cells containing
strings where the font is the same colour as the background.
}
\details{
The cells removed by this function are usually specified as needing removal
because not doing so results in errors during beheading (un-pivotting).

For more information and for examples, see the documentation for
remove_unwanted_cells and remove_hidden_character_strings.
}
