% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_information.R
\name{combine_columns}
\alias{combine_columns}
\title{concatenate columns}
\usage{
combine_columns(
  dat,
  columns_to_combine_patterns,
  combined_names,
  group_counts = NA
)
}
\arguments{
\item{dat}{dataframe with the columns specified by columns_to_combine_patterns.}

\item{columns_to_combine_patterns}{vector of character strings. Each
string matches the name of a column in dat.}

\item{combined_names}{character string vector giving the names of the columns
that will hold the newly combined strings. Must be the same length as
group_counts, unless group_counts is NA, in which case it must contain a
single value. In pub sec this variable is specified by
columns_to_combine_combined_names.}

\item{group_counts}{integer vector. Default is NA. The first element of
group_count is used to supply the number of elements of columns_to_combine_patterns
that should be used to create the first of combined_names, the second to
create the second of combined_names etc.}
}
\value{
dat with columns concatenated into a single column. If
columns_to_combine_patterns is Na, dat is returned unmodified
}
\description{
combine the information in the specified columns to create a
single column. information from the different columns are separated with a
hyphen. If more than one grouping is required, group_count can be used to
show where the split is.
}
\examples{
\dontrun{
dat <- data.frame(
    "group" = c("Fox", "Fox"),
    "common_name" = c("Red", "Arctic"),
    "age" = c("cub", "adult"),
    "cats" = c("lion", "lion"),
    "species" = c("African", "Asiatic")
    )
columns_to_combine_patterns <- c("common_name", "group", "age", "species", "cats")
combined_names <- c("Non-cats", "Cats")
group_count <- c(3, 2)
newdat <- combine_columns(
    dat, columns_to_combine_patterns, combined_names, group_count)
}
}
