% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutive_years_to_financial.R
\name{consecutive_years_to_financial}
\alias{consecutive_years_to_financial}
\title{consecutive_years_to_financial}
\usage{
consecutive_years_to_financial(year)
}
\arguments{
\item{year}{the year wanted to convert}
}
\value{
year the newly converted year
}
\description{
TBD
}
\details{
TBD
}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
