% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_date_to_char.R
\name{convert_date_to_char}
\alias{convert_date_to_char}
\title{convert_date_to_char}
\usage{
convert_date_to_char(source_df)
}
\arguments{
\item{source_df}{dataframe}
}
\value{
transformed_date_df dataframe
}
\description{
convert dates to character strings
if the data_type is 'date' and the date column is not NA
Update the data_type and date columns to be consistent.
}
\examples{
convert_date_to_char(source_df)
}
