% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_list.R
\name{create_table_list}
\alias{create_table_list}
\title{create_table_list}
\usage{
create_table_list(header_rows_combined, table_start_cols, table_count)
}
\arguments{
\item{header_rows_combined}{A dataframe containing column positions and other relevant data.}

\item{table_start_cols}{A numeric vector specifying the starting column indices for each table.}

\item{table_count}{An integer representing the number of tables to extract.}
}
\value{
A list of dataframes, each containing a subset of `header_rows_combined` corresponding to a table.
}
\description{
Create a list of tables by segmenting header rows based on column positions.
}
\details{
Create a list of tables by segmenting header rows based on column positions.

This function loops through the starting column positions of tables and extracts
segments from `header_rows_combined`, storing them as individual tables in a list.
}
\examples{
header_rows_combined <- data.frame(
  "row" = c(1, 1, 1, 2, 2, 2),
  "col" = c(1, 2, 3, 1, 2, 3),
  "address" = c("A1", "B1", "C1", "A2", "B2", "C2"),
  "character"= c(NA,NA,NA,"Education Services","NA","NA")
)
table_start_cols <- 1
table_count <- 1
table_list <- create_table_list(header_rows_combined, table_start_cols, table_count)
}
\author{
Barbie Mwangi \email{Barbie.Mwangi@ons.gov.uk}
}
