% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_xlsx_cells_table_data.R
\name{extend_row_value}
\alias{extend_row_value}
\title{Add information from above or below a row into another row.}
\usage{
extend_row_value(dat, pattern = NA, placement = NA, direction = NA)
}
\arguments{
\item{dat}{dataframe imported using tidyxl::xlsx_cells (one row for
each cell).}

\item{pattern}{character string. Regular expressions used to identify the
rows that need added information from the  row above or below. In pub sec
this variable is specified by extend_row_pattern.}

\item{placement}{character string. Either 'forward' or 'reverse'. Used
to determine if the added information is placed before or after original row
value. In pub sec this variable is specified with extend_row_order.}

\item{direction}{character string. Either 'above' or 'below'. Used to
determine if the added information is from above or below the row that
matches pattern. In pub sec this variable is specified with extend_row_with.}
}
\value{
dataframe in xlsx_cells format. If all vars are soecified, returns
dat with modified values in the character column. If pattern,
placement and direction are NA returns dat with no modification.
If the pattern from the data dict doesn't match any character in the
dataframe.
}
\description{
Use a specified pattern to find values in the character column
of an xlsx_cells dataframe and extend the found value with information from
either the rows above or below.
}
\details{
This function is used for datasets where the descriptors for a
value are not unique because they relate to information above or (less
likely) below them. For example the descriptor '...of which HRA' may be
repeated multiple times to indicate that the values in that row are a
subset of the value above. However, taken on their own they are fairly
meaningless.

NOTE: If the pattern has matches in multiple columns, the change is only made
to the column where it occurs most. If this is not the desired behaviour, the
function will need to be updated - suggest making the change on data after it
has been unpivotted rather than on xlsx_cells data, as then the column name
to  which the changes are desired can be specified (e.g. using column name
pattern matching).

This was written to address an issue in pub sec in the LA Dropdown tab of the
DLUHC (MHCLG) capital payments receipts data (see example).

TODO: this function should be re-written to be called \emph{after} data have
been unpivotted. this will allow users to specify which column the changes
should be made in.
}
\examples{
\dontrun{
# see Wiki for example with illustration

dat <- data.frame(
 id = 1:13,
 address = c("A9", "A10", "A11", "A12", "A13", "A14", "A15", "A16",
             "A17", "A18","A19","A20","A21"),
 character = c("Acquisition of land & existing buildings", "...of which HPA",
               "Another character", "...of which expenditure", "More characters",
               "...of which grants","...of which HRA", "Even more characters",
               "...of which receipts", "Some text", "...of which loans",
               "...of which payment","total","payment")
)

result <- extend_row_value(
    dat, "^\\\\s*\\\\.\\\\.\\\\.\\\\s*of\\\\s*which", "reverse", "below"
    )
}
}
