% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_row_value.R
\name{extend_row_value}
\alias{extend_row_value}
\title{extend_row_value}
\usage{
extend_row_value(
  source_data,
  extend_row_regex,
  extend_row_order,
  extend_row_with
)
}
\arguments{
\item{source_data}{Tidyxl format dataframe (Data imported with tidyxl::xlsx_cells,
which have one row for each cell).}

\item{extend_row_regex}{(str).  regex to identify the rows that need added information from the row above or below}

\item{extend_row_order}{(str). Either forward or reverse. Used to determine if the added information is placed before or after original row value}

\item{extend_row_with}{(str). Either above or below. Used to determine if the added information is from above or below the row.}
}
\value{
source_data with modified values in the character column

Pipeline in halted if extend_row_with and extend_row_order are not the expected variables or if
the pattern from the data dict doesnt match any character in the dataframe
}
\description{
extend_row_value adds information from above or below a row into another row.
}
\details{
extend_row_value adds information from above or below a row into another row.

Uses pattern passed from the data dictionary to find values in the character column that match then using
variables passed from the data dict to decide which rows to add that found values to.

This was written to address an issue in the LA Dropdown tab of the DLUHC capital payments receipts data, see example.
}
\examples{
(see confulence for visual example)

data <- data.frame(
 id = 1:13,
 address = c("A9", "A10", "A11", "A12", "A13", "A14", "A15", "A16", "A17", "A18","A19","A20","A21"),
 character = c("Acquisition of land & existing buildings", "...of which HPA",
               "Another character", "...of which expenditure", "More characters", "...of which grants","...of which HRA",
               "Even more characters", "...of which receipts", "Some text", "...of which loans", "...of which payment","total","payment")
)

result <- extend_row_value(data, "^\\\\s*\\\\.\\\\.\\\\.\\\\s*of\\\\s*which", "reverse", "below")


}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}
}
