% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_all_years.R
\name{extract_all_years}
\alias{extract_all_years}
\title{extract_all_years}
\usage{
extract_all_years(dat)
}
\arguments{
\item{dat}{Numeric or string. Also takes vectors, matrices, or dataframes}

\item{year_type}{Character vector. "yyyy" for calendar year, "yyyy-yy" for e.g. 2021-22}
}
\value{
A string, vector of strings, or NA. 
Returns a vector of all patterns that matched the year_type pattern.
If no years are identified in any of the input strings NA is returned.
}
\description{
Identify years from strings
}
\details{
Identify years from strings

Identifies whether a string contains a number that looks like a year from the
20th or 21st century. If no year is identified NA is returned.
If one or more years are identified, a string, or vector of unique strings 
is returned.
}
\examples{
test_dat <- c("a date: 2005", "1995 6", "1993 end", "a1995", "1995a", "a
date:1995, another date: 2012", "not a date: 475", "also not a date: 1234",
"abc 12345",  "34200943", "190932")#' get_all_country_names(test_dat)
extract_all_years(test_dat)

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
