% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_matches.R
\name{extract_matches}
\alias{extract_matches}
\title{extract_matches}
\usage{
extract_matches(dat, pattern)
}
\arguments{
\item{dat}{Numeric or string}

\item{pattern}{String. Must be a Regex pattern.}
}
\value{
List of character strings. See description for detail.
}
\description{
Extract patterns from a vector of strings
}
\details{
Extract patterns from a vector of strings

Looks for a pattern and returns the parts of the strings that matched as a 
list. Numeric input will be converted to strings. Each item of the input data 
will be returned in it's own vector, with each match given as an item
in that vector. The output is therefore a list of vectors. For example if
dat is c( "a2", "2 and 3"), and the search pattern is for 
numbers the list returned will be [[1]][1] "2" [[2]][1] "2" [[2]][2] "3"
}
\examples{
numbers <- extract_matches(c( "a2", "2 and 3"), '[0-9]')

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
