% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_info.R
\name{fill_blanks}
\alias{fill_blanks}
\title{Vertically fill blanks in a descriptor column.}
\usage{
fill_blanks(dat, column_names, fill_dirs = "down")
}
\arguments{
\item{dat}{dataframe.}

\item{column_names}{charater string vector. The names of the columns to
apply the function to. In pub sec this variable is specified using
fill_columns_column_names.}

\item{fill_dirs}{character string vector. Must be the same length as
column_names. Values allowed are "down", "up", "downup" (i.e. first down and
then up) or "updown" (first up and then down). Defaults to 'down'. In pub sec
this variable is specified using fill_columns_fill_dirs.}
}
\value{
dat with blanks filled.
}
\description{
Fill blanks within a column with the first populated entry
either above or below the blank.
}
\details{
This function is useful for data with layouts such as the following:\tabular{lll}{
   \strong{group} \tab \strong{subgroup} \tab \strong{value} \cr
   cat \tab seval \tab 10 \cr
    \tab margay \tab 2 \cr
    \tab ocelot \tab 15 \cr
   dog \tab dhole \tab 1 \cr
    \tab dingo \tab 15 \cr
}


The group column may or may not have contained merged cells but each row in
group will be given a value if this function is used:\tabular{lll}{
   \strong{group} \tab \strong{subgroup} \tab \strong{value} \cr
   cat \tab serval \tab 10 \cr
   cat \tab margay \tab 2 \cr
   cat \tab ocelot \tab 15 \cr
   dog \tab dhole \tab 1 \cr
   dog \tab dingo \tab 15 \cr
}
}
\examples{
\dontrun{
dat <- data.frame(
    var1 = c("colour", NA, NA, "wavelength"),
    var2 = c("red", "green", "blue", "short")
    )
fill_blanks(dat, "var1", "down")
}

}
