% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_info.R
\name{fill_missing_fy_start}
\alias{fill_missing_fy_start}
\title{Where year_type is 'financial' fill in missing fy_start info.}
\usage{
fill_missing_fy_start(dat)
}
\arguments{
\item{dat}{dataframe containing columns called 'year' and 'year_type'.}
}
\value{
dataframe. dat with fy_start filled in for years where year_type is
'financial'.
}
\description{
Where year_type is 'financial' and there are NAs for fy_start
(or fy_start does not exist), use the first four digits of year for fy_start.
}
\examples{
\dontrun{
dat <- data.frame(
    year = c("2020", "2020-21", "2021-22"),
    year_type = c("calendar", "financial", "financial"),
    fy_start = c(NA, NA, "2021")
    )

fill_missing_fy_start(dat)

}

}
