% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dict_item.R
\name{get_dict_item}
\alias{get_dict_item}
\title{get_dict_item}
\usage{
get_dict_item(key, dict)
}
\arguments{
\item{key}{character string that is present in dict}

\item{dict}{dataframe with key and item columns.}
}
\value{
The item from dict for the key in question
}
\description{
Get value for a key in a python dict (as a data-frame)
}
\details{
Get value for a key in a python dict (as a data-frame)

Used on a python dictionary that has been turned into a data-frame with a 
column for key and a column for item
}
\examples{
layout_dict <- data.frame("key" = c("header", "group"),
                           "item" = c("NA", "mice"))
group <- get_dict_item("header", layout_dict)
print(group)

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}
}
