% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_duplicated_rows.R
\name{get_duplicated_rows}
\alias{get_duplicated_rows}
\title{get_duplicated_rows}
\usage{
get_duplicated_rows(dat, row_num)
}
\arguments{
\item{dat}{tidyxl::xlsx_cells dataframe}

\item{row_num}{integer. The row you want to find duplicates of}
}
\value{
integer vector. The row numbers of duplicated rows
         (but not the original row: row_num).
}
\description{
Get the row numbers of duplicated rows
}
\details{
Get the row numbers of duplicated rows

Data imported with tidyxl::xlsx_cells have one row for each cell. As such
the usual ways of finding duplicated rows in the raw data (e.g `unique` or
`distinct`) do not work.

Given a row number, find any rows that are duplicates of that row.
}
\examples{
example_dat <- data.frame(
  row = rep(1:4, each = 2),
  col = rep(1:2, by = 4),
  character = rep(c("year", "quarter", rep(NA, 2)), 2),
  numeric = c(NA, NA, 1, 2, NA, NA, 3, 4)
)

get_duplicated_rows(example_dat, 1)
}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
