% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_instance.R
\name{get_first_instance}
\alias{get_first_instance}
\title{get_first_instance}
\usage{
get_first_instance(dat, pattern, column = "row")
}
\arguments{
\item{dat}{An xlsx_cells dataframe}

\item{pattern}{string}

\item{column}{"row" or "col". Default is row}
}
\value{
numeric. The first row/col from the xlsx on which known_string occurs.
        Defaults to 1 if dat is not an xlsx_cells df.
}
\description{
Identify the first row or column on which a string occurs
}
\details{
Identify the first row or column on which a string occurs

Takes an xlsx_cells dataframe and a string known to be
in the row or column you wish to identify but not in any previous rows/cols.
Returns the row/col number from the original excel sheet.
This function is not case sensitive - it will for example identify
pattern = "Blue" as a match with "blue" in the dataframe.
}
\examples{
dat <- data.frame("character" = c("title", "million", NA,
                                 "Education services", "Highways services",
                                 NA),
                  "row" = c(1:3, 4, 4, 5),
                  "numeric" = c(rep(NA, times = 5), 50))
get_first_instance(dat, "services")

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
