% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_of_consecutives.R
\name{get_first_of_consecutives}
\alias{get_first_of_consecutives}
\title{get_first_of_consecutives}
\usage{
get_first_of_consecutives(x)
}
\arguments{
\item{x}{vector of integers}
}
\value{
integer. If there are consecutive values, this will be the first of 
the consecutive values. If there are no consecutive values a warning will
be given and the first value returned.
}
\description{
Get the first integer where the following number is consecutive
}
\details{
Get the first integer where the following number is consecutive

This function returns the first number in a vector where the next number is 
the consecutive number.

For example, if you pass it a vector of numbers that are column positions
where the datatype is numeric, it will return the position of the first 
column where both it and the following column are numeric.
}
\examples{
x <- c(1, 1, 3, 5, 6, 8)

get_first_of_consecutives(x)

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
