% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_above_table.R
\name{get_info_above_table}
\alias{get_info_above_table}
\title{get_info_above_table}
\usage{
get_info_above_table(main_table, header_row)
}
\arguments{
\item{main_table}{The source data table.}

\item{header_row}{The row number of the first header row.}
}
\value{
info_above_table the data above the table
}
\description{
Get the first header row index based on specified identifiers

Returns data from above the header row

This function removes empty rows from the source data.
}
\details{
Returns data from above the header row
}
\examples{
main_table <- remove_empty_rows(main_table, first_header_row)
}
\author{
Mark London \email{mark.london@ext.ons.gov.uk}

Your Name
}
