% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matching_colnames.R
\name{get_matching_colnames}
\alias{get_matching_colnames}
\title{get_matching_colnames}
\usage{
get_matching_colnames(dat, pattern)
}
\arguments{
\item{dat}{dataframe}

\item{pattern}{string. Regular expression you want to match}
}
\value{
vector of strings. names of the col names the pattern is found in
}
\description{
TBD
}
\details{
Get column names that contain a pattern
}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
