% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sheet.R
\name{get_metadata}
\alias{get_metadata}
\title{Extract metadata from a dataframe and from other supplied information.}
\usage{
get_metadata(
  sheet_dat,
  dropdown_pattern,
  single_vintage = NA,
  release_number = NA,
  table_dat = NA,
  sheet_title = NA,
  sheet_units = NA,
  sheet_year = NA,
  filename_year = NA,
  use_year_from_filename = NA,
  no_warning = NA
)
}
\arguments{
\item{dat}{dataframe in xlsx_cells format, containing only metadata i.e.
information above the table or tables.}
}
\description{
Extract metadata from a dataframe and from other supplied information.
}
\details{
In some DLUHC datasets we have to use the LA_dropdown tab. In these sheets
there is a dropdown at the top that controls what data are shown on the
sheet, and therefore what data are imported.
}
