% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_information.R
\name{get_units}
\alias{get_units}
\title{Extract information about units}
\usage{
get_units(dat)
}
\arguments{
\item{dat}{dataframe imported using tidyxl::xlsx_cells() containing the cells
that are likely to include information about units.}
}
\value{
A string containing extracted unique unit labels separated by
commas, or an empty string if no units are found.
}
\description{
Extract unit labels such as 'thousand', 'million', '000s', or
'£ Millions' from the character column of data imported using
tidyxl::xlsx_cells().
}
\examples{
\dontrun{
info_at_top_of_sheet <- data.frame(
    character = c(
        "Revenue in £ millions (millions)",
        "All figures are in thousands of pounds (£ thousand)."
        )
    )

units <- get_units(info_at_top_of_sheet)
}
}
