% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vintage_from.R
\name{get_vintage_from}
\alias{get_vintage_from}
\title{get_vintage_from}
\usage{
get_vintage_from(vintage_check_item, sheet_vintage = NA)
}
\arguments{
\item{vintage_check_item}{a cell or filepath that could contain the vintage information}

\item{sheet_vintage}{default NA, but it set if found.}

\item{vintage_patterns}{a list of allowable vintage strings}
}
\value{
vintage_result the matched result or 'NA' if none found.
}
\description{
Gets a vintage as a string match
}
\details{
Additional details...

get_vintage_from checks items against a string to find a match 
for the vintage string possibilities
}
\examples{
example code sets the sheet vintage to 'forecast', finding it in the patterns
  sheet_vintage <- "NA"
  vintage_check_item <- "2023/24 Forecast"
  expected_result <- "forecast"
  sheet_vintage <- get_vintage_from(vintage_check_item, vintage_patterns, sheet_vintage)

Caes of multiple vintages:
sheet_vintage <- NA
vintage_check_item <- "budget and provisional data"
expected_result <- "budget, provisional"
actual_result <- suppressWarnings(get_vintage_from(vintage_check_item, sheet_vintage))


}
\author{
Mark London \email{mark.london@ext.ons.gov.uk}
}
