% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_handlers.R
\name{get_year_for_use_in_data}
\alias{get_year_for_use_in_data}
\title{Get and compare year from above the data to a set year}
\usage{
get_year_for_use_in_data(
  sheet_year,
  table_year,
  filename_year = NA,
  use_filename_year = FALSE,
  prefer_sheet_year = FALSE,
  suppress_warning = FALSE
)
}
\arguments{
\item{filename_year}{character string vector.}

\item{use_filename_year}{boolean. Default is FALSE. If FALSE and there is a
mismatch, the year found above the table will be returned.If TRUE and no year
that is found in the information above the table matches the filename year,
the filename year will be returned.  In pub sec this variable is specified
by use_year_from_filename_over_year_above_data.}

\item{prefer_sheet_year}{boolean. Default is FALSE. Only used when there is a
mismatch between the year(s) given at the top of the sheet and the years
found above a table. If either are NA, this variable is not used. If FALSE,
the year from above the table is used. If TRUE, the year from the top of the
sheet is used. In pub sec this variable is specified by
use_sheet_year_over_table_year.}

\item{suppress_warning}{boolean. Default is FALSE. If TRUE, warnings about
mismatching years between the filename and the information above the data
will be turned off. In pub sec this variable is specified by
suppress_year_above_table_warning. It is required because there are some
cases where we expect a mismatch and giving a warning would be confusing to
users.}

\item{title}{character string.}
}
\value{
list of 'year' and 'warn': 'year' is a character string, 'warn' is
boolean. If TRUE, a mismatch was found between the filename year and the year
found above the data. This is given as a boolean output rather than raising
a warning, because the warning is only relevant if the year output is added
to the data as a column. If 'warn' is TRUE a warning is raised if and when
'year' is added as a column to the data.
}
\description{
For data that are for a single year where there is no year
column. Year may be given in information above the table. In public sector
a year is also given in the file name. This function finds years in the
info above the data (dat) that are of the same type (calendar/financial)
as the year in the file name. It compares them to the year in the file name,
and if they are different, it selects one based on \code{use_filename_year}.

If multiple are found and one of them matches the file name, the year from
the file name is used.
If multiple are found and none of them match the file name, and
use_filename_year is FALSE, the first is used and a warning is given.
If none are found, the file name year is used as default.
}
\examples{
\dontrun{
filename <- "D:/data-2024_25.xlsx"
get_year_for_use_in_data(
    "2024-25", "2025-26", use_filename_year = FALSE
    )
get_year_for_use_in_data(
    "Title 2026", "2024_25", use_filename_year = TRUE
    )
}
}
