% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_handlers.R
\name{get_year_for_use_in_data}
\alias{get_year_for_use_in_data}
\title{Get and compare year from above the data to a set year}
\usage{
get_year_for_use_in_data(
  sheet_year,
  table_year,
  filename_year = NA,
  use_filename_year = FALSE,
  prefer_sheet_year = FALSE,
  suppress_warning = FALSE
)
}
\arguments{
\item{sheet_year}{character string vector. Years found at the top of the
sheet above the first header row.}

\item{table_year}{character string vector. Years found above an individual
table before the table's first header row.}

\item{filename_year}{character string.}

\item{use_filename_year}{boolean. Default is FALSE. If FALSE and there is a
mismatch, the year found above the table will be returned.If TRUE and no year
that is found in the information above the table matches the filename year,
the filename year will be returned.  In pub sec this variable is specified
by use_year_from_filename_over_year_above_data.}

\item{prefer_sheet_year}{boolean. Default is FALSE. Only used when there is a
mismatch between the year(s) given at the top of the sheet and the years
found above a table. If either are NA, this variable is not used. If FALSE,
the year from above the table is used. If TRUE, the year from the top of the
sheet is used. In pub sec this variable is specified by
use_sheet_year_over_table_year.}

\item{suppress_warning}{boolean. Default is FALSE. If TRUE, warnings about
mismatching years between the filename and the information above the data
will be turned off. In pub sec this variable is specified by
suppress_year_above_table_warning. It is required because there are some
cases where we expect a mismatch and giving a warning would be confusing to
users.}
}
\value{
list of 'year' and 'warn': 'year' is a character string, 'warn' is
boolean. If TRUE, a mismatch was found between the filename year and the year
found above the data. This is given as a boolean output rather than raising
a warning, because the warning is only relevant if the year output is added
to the data as a column. If 'warn' is TRUE a warning is raised if and when
'year' is added as a column to the data.
}
\description{
This function is required for data that relate to a single year.
Year is unlikely to be given as a column in such datasets so we need to get
the year from the information above the data instead (either at the top of
sheet or directly above the table if there are multiple tables in the sheet).
In public sector we can also get year from the file name. This function gets
the year from above the data, compares in to the year in the file name, and
returns the preferred year. If a mismatch is found, this is flagged.
}
\details{
This function finds years in the info above the data (dat) that are of the
same type (calendar/financial) as the year in the file name. It compares them
to the year in the file name, and if they are different, it selects one based
on \code{use_filename_year}.

If multiple unique years are found and one of them matches the file name, the
year from the file name is returned.

If multiple unique years are found, none of them match the file name, and
use_filename_year is FALSE, the first of either the year at the top of the
sheet or the first of the year above the table is used and the mismatch is
flagged. By default the year above the table is preferred to the year at the
top of the sheet if both exist. However this can be changed by setting
prefer_sheet_year to TRUE.

If no years are found above the data, or if use_filename_year is TRUE, the
filename year is used.

Mismatches are flagged rather than raising a warning because the user only
needs to be given the warning if the variable is actually used. For example,
if year is already a column in the data the returned variable will not be
used and it does not matter if multiple years are mentioned in the metadata.
}
\examples{
\dontrun{
filename <- "D:/data-2024_25.xlsx"
get_year_for_use_in_data(
    "2024-25", "2025-26", use_filename_year = FALSE
    )
get_year_for_use_in_data(
    "Title 2026", "2024_25", use_filename_year = TRUE
    )
}
}
