% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_information.R
\name{identify_columns_to_remove}
\alias{identify_columns_to_remove}
\title{Identify header cells to remove from xlsx_cells data.}
\usage{
identify_columns_to_remove(dat, patterns, offset, target_rows)
}
\arguments{
\item{dat}{dataframe imported using xlsx_cells.}

\item{patterns}{vector of character strings that are regular expressions,
each one matching the name of a column in the raw data.}

\item{offset}{vector of integers. Must be either NA or the same length as
patterns. If the column to remove is not named but it's location is known in
relation to a named column, specify the pattern for the named column and use
offset to move x columns to the left (negative) or right (positive).}

\item{target_rows}{vector of integers. The numbers of the rows in which
column headings are given}
}
\value{
vector of character strings giving the column numbers identified.
If no match is found in the data for a pattern a warning is given.
}
\description{
Use regular expressions to identify which column of the data
they appear in. Only header rows are checked for matches.

If a pattern is matched in more than one column none of those matches are
returned as columns to remove. This does not affect the use of any other
patterns provided.
}
\examples{
\dontrun{
dat <- data.frame(
  "row" = c(1, 1, 1, 2, 2, 2),
  "col" = c(1, 2, 3, 1, 2, 3),
  "address" = c("A1", "B1", "C1", "A2", "B2", "C2"),
  "numeric" = c(NA, NA, NA, NA, 1, 100),
  "character" = c("name", "id", "value", "first row", NA, NA),
  "data_type" = c(rep("character", 4), rep("numeric", 2))
)
# view as it would be in excel:
rectify(dat)

identify_columns_to_remove(dat, "id", NA, 1)
}
}
