% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_year_patterns.R
\name{make_year_patterns}
\alias{make_year_patterns}
\title{make_year_patterns}
\usage{
make_year_patterns()
}
\value{
Dataframe containing one row per column. Columns are different year types
}
\description{
Build regex patterns for years between 1900 and 2099
}
\details{
Build regex patterns for years between 1900 and 2099

Creates regex patterns that look like a year or a financial year from the
20th or 21st century.
}
\examples{
patterns <- make_year_patterns()

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
