% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_year_column.R
\name{process_year_column}
\alias{process_year_column}
\title{Process the 'year' column in a data frame}
\usage{
process_year_column(
  dat,
  pattern,
  use_single_year,
  single_year,
  single_overrides_all,
  multi_year_entries_as_na
)
}
\arguments{
\item{dat}{The data frame containing the data to be processed.}

\item{pattern}{character string. The name of the year column as given by a
user. If this is NA, the year column is identified automatically by
get_year_column_names(). In pub sec this variable is specified by
year_col_pattern.}

\item{use_single_year}{boolean or NA. If TRUE create a column called year and
populate it with single_year, in pub sec this variable is specified by
single_year_of_data. Unless single_overrides_all is TRUE, this will only be
used as a last option.}

\item{single_year}{list of length 2: 'year' is a character string containing
a valid year; 'warn' is boolean. TRUE if a warning needs to be raised.
single_year is only used if use_single_year is TRUE. IF so, the 'year'
column is populated with this year value. Must be specified if
use_single_year is TRUE. In pub sec this variable is taken either from the
file name (which gets its year from the user config selections) or from the
information above the data. 'warn' is TRUE when the file name year and the
year from above the data both exist but do not match.}

\item{single_overrides_all}{boolean or NA. If TRUE, and use_single_year is
TRUE, single_year overrides any columns found with year in the name. If FALSE
or NA, other options for filling the year are used first and single year is
only used as a final option.  N.B If single_overrides_all is TRUE we would
not expect pattern to have been specified (unless it is also used in other
functions).}

\item{multi_year_entries_as_na}{bool. If NA or NULL defaults to FALSE
If TRUE, and the year is a range of years e.g '2021-22 to 2023-24' or
'2023, 2024', the year column will show NA. Whether it is TRUE or FALSE,
year_type will be NA for invalid year entries. In pub sec this variable is
specified with the same name (multi_year_entries_as_na)}
}
\value{
The modified data frame after processing the 'year' column. If they
did not already exist 'year' and 'year_type' are added as new columns,
otherwise they are overwritten with standardised text.
}
\description{
Potential year columns are identified either by their matching a provided
pattern OR by matching '(?i)year' if pattern is not provided. Where there is
no year column, it will be generated from a provided single year if
use_single_year is true.
}
\details{
If more than one year column is available, financial is preferred over
calendar, which is preferred over a mix of the two, which is preferred over
a given single year, unless single_overrides all is true. Note that in pub
sec, single year is taken from the filename or the information above the
data.

If the user has supplied a pattern and it matches the name of a column/
columns containing valid years, this/these columns will be the only
possible_cols. If pattern is not supplied, any column with 'year' in the
name that contains valid years will be included (one column per year type).
However, if single_overrides_all is TRUE and use_single_year is TRUE.

Year type is ascertained for each row (financial or calendar). If it is
financial year it is standardised to the format YYYY-YY (rather than e.g.
YYYY/YY or YYYY-YYYY)

If the original year column is called 'year' it will be overwritten.
}
\examples{
\dontrun{
dat <- data.frame(year_col = c("2021", "2022_23 to 2025_26", "2021, 2022"))
process_year_column(dat, "year_col", NA, NA, NA, NA)

dat <- data.frame(
    financial_year = c("2021-22 to 2022-23", "2022_2023", "words"),
    calendar_year = c(2021, 2022, NA)
    )
process_year_column(dat, NA, NA, NA, NA, TRUE)
process_year_column(dat, NA, NA, NA, NA, FALSE)

# case where there is no year column, but there is a column that looks like
# it might be:
dat <- data.frame(
    description = c("A", "B", "C"),
    implementation_year = c(1980, 1992, 2003),
    value = 1:3
    )
single_year = list("year" = "2021_22", "warn" = TRUE)
process_year_column(
    dat, pattern = NA, use_single_year = TRUE, single_year = single_year,
    single_overrides_all = TRUE, multi_year_entries_as_na = TRUE
    )
}
}
