% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_year_column.R
\name{process_year_column}
\alias{process_year_column}
\title{process_year_column}
\usage{
process_year_column(dat, year_pattern, year_for_layout)
}
\arguments{
\item{dat}{The data frame containing the data to be processed.}

\item{year_for_layout}{Year value to be used for creating a new 'year' column if needed.}

\item{year_value_patterns}{Regular expression pattern to match year values.}
}
\value{
The modified data frame after processing the 'year' column.
}
\description{
Process the 'year' column in a data frame
}
\details{
Process the 'year' column in a data frame

Process the 'year' column in a data frame according to specific conditions.
Format year so that it follows the structure yyyy-yy.
}
\examples{
# Sample usage:
output <- data.frame(year_col = c("2021", "2022-23"))
year_for_layout <- "2023"
processed_output <- process_year_column(output, year_value_patterns, year_for_layout)
}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
