% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_xlsx_cells_table_data.R
\name{remove_columns}
\alias{remove_columns}
\title{Remove lines that were unwanted columns in the Excel data}
\usage{
remove_columns(dat, patterns, offset, first_row, header_count)
}
\arguments{
\item{dat}{dataframe imported using tidyxl::xlsx_cells.}

\item{patterns}{vector of character strings that are regular expressions,
each one matching the name of a column in the raw data to be removed. In
pub sec this variable is specified by columns_to_remove_patterns.}

\item{offset}{vector of integers. Must be either NA or the same length as
patterns. If the column to remove is not named but it's location is known in
relation to a named column, specify the pattern for the named column and use
offset to move x columns to the left (negative) or right (positive). In pub
sec this variable is specified by columns_to_remove_offset.}

\item{first_row}{integer. The first row in which column headings are found.}

\item{header_count}{integer. The number of header rows.}
}
\value{
dataframe. dat with rows relating to the matched columns removed.
If a pattern matches character strings in more than one column it is not
removed and a warning is given. This does not affect the use of any other
patterns provided in the patterns vector. If no match is found in the
data for a patterns pattern a warning is given.
}
\description{
Remove all rows relating to a column in data imported using
tidyxl::xlsx_cells if a pattern is matched by a character in that column.

More than one column can be removed by providing more than one pattern.

If the name of the column to be removed is blank, or if it is unstable (i.e.
writing a pattern to match it is hard/impossible), a pattern that matches a
nearby column can be used along with offset. This requires that you are
confident about the order of the columns.
}
\examples{
\dontrun{
dat <- data.frame(
  "row" = c(1, 1, 1, 2, 2, 2),
  "col" = c(1, 2, 3, 1, 2, 3),
  "address" = c("A1", "B1", "C1", "A2", "B2", "C2"),
  "numeric" = c(NA, NA, NA, NA, 1, 100),
  "character" = c("name", "id", "value", "first row", NA, NA),
  "data_type" = c(rep("character", 4), rep("numeric", 2))
)
# view as it would be in excel:
rectify(dat)

output <- remove_columns(dat, "id", 0, 1, 1)
rectify(output)
}
}
