% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_columns.R
\name{remove_columns}
\alias{remove_columns}
\title{remove_columns}
\usage{
remove_columns(dat, column_regex)
}
\description{
This function searches for the specified header identifiers and second identifiers (if provided) in the file part.
It then identifies the first row of the header based on the found identifiers and returns its index.
Remove lines that were unwanted columns in the raw Excel data

In some spreadsheets there may be a column or columns that we don't want to be
there because they mess up the way the data are un-pivoted, and are not
needed by the users. For example the Scottish Government POBE capital 2024-25
download has a column of reference numbers in column D. This causes problems
in reformat_multi_layout_data.

This function uses a regular expression given in the data dictionary to
identify the rows that need to be removed from the xlsx_cells data. The
identified rows are then removed.

More than one column can be removed by providing more than one column_regex
(this will most likely come from cols_to_remove_regex in the data dict).
However, if a column_regex pattern has matches for more than one col it is
not removed and a warning is given. This does not affect the use of
any other patterns provided in the column_regex vector.

If no match is found in the data for a column_regex pattern a warning is given.

 @param dat Dataframe. Imported using xlsx_cells
 @param column_regex vector of strings that are regular expressions, each one
 matching the nae of a column in the raw data that should be removed.

 @example
dat <- data.frame(
  "row" = c(1, 1, 1, 2, 2, 2),
  "col" = c(1, 2, 3, 1, 2, 3),
  "address" = c("A1", "B1", "C1", "A2", "B2", "C2"),
  "numeric" = c(NA, NA, NA, NA, 1, 100),
  "character" = c("name", "id", "value", "first row", NA, NA),
  "data_type" = c(rep("character", 4), rep("numeric", 2))
)
# view as it would be in excel:
rectify(dat)

output <- remove_column(dat, "id")
rectify(output)
}
