% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empty_lines.R
\name{remove_empty_lines}
\alias{remove_empty_lines}
\title{remove_empty_lines}
\usage{
remove_empty_lines(dat, column)
}
\arguments{
\item{dat}{Dataframe. Imported using xlsx_cells}

\item{column}{str. Either "row" or "col"}
}
\description{
remove empty columns or empty rows
}
\details{
remove empty columns or empty rows
}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
