% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_line_breaks.R
\name{remove_line_breaks}
\alias{remove_line_breaks}
\title{remove_line_breaks}
\usage{
remove_line_breaks(dataset)
}
\arguments{
\item{dataset}{data.frame : A dataset with line breaks in values}
}
\value{
cleaned_dataset : The dataset with removed line breaks from values
}
\description{
remove_line_breaks - A function to Remove line breaks from character strings in data
}
\details{
remove_line_breaks - A function to Remove line breaks from character strings in data

Some raw (downloaded) datafiles have got newlines included in character strings
}
\examples{
Example data where there is line breaks in data values
dataset <- tibble::tibble(
   "Service"  <- c("Education\nServices","Education Services", "Education\n\n Services","\nEducation\n Services\n"),
   "Transaction" <- c("Expenditure\n", "\nExpenditure", "\nExpenditure\n", "Expenditure"),
   "Units" <- c("thousands","thousands","hundreds", "millions"),)
   
remove_line_breaks(dataset)

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
