% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_information.R
\name{remove_unwanted_cells}
\alias{remove_unwanted_cells}
\title{Remove cells from input data that are not needed.}
\usage{
remove_unwanted_cells(dat, cells = NA)
}
\arguments{
\item{dat}{dataframe imported using tidyxl::xlsx_cells}

\item{cells}{vector of character strings. The Excel addresses of
cells to be removed from data e.g. 'A1'. Each of cells must have a
corresponding min_year and max_year. In pub sec this variable is
specified by cells_to_remove}
}
\description{
Remove rows from an xlsx_cells imported dataframe
using cell addresses to identify which rows to remove. The action is only
carried out if the given year is within the stated year range.
}
\examples{
\dontrun{
dat <- data.frame(address = c("A1", "A2", "A3"), numeric = 1:3)

cells_to_remove <- c("A1", "A2")

remove_unwanted_cells(dat, cells_to_remove)
}
}
