% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_duplicates.R
\name{rename_duplicates}
\alias{rename_duplicates}
\title{Add a prefix to a duplicated entry}
\usage{
rename_duplicates(
  dat,
  cols_to_group_by,
  column = NA,
  pattern = NA,
  index = NA,
  prefix = NA,
  expected_freq = NA
)
}
\arguments{
\item{dat}{dataframe. Must include a column with the name given in column.}

\item{cols_to_group_by}{vector of character strings giving the names of
columns that will have been created during un-pivotting. The entries in these
columns would uniquely identify each row if there were no duplicates.}

\item{column}{character string. The name of the column the duplicate is in in
dat. Vector of length 1. In pub sec this variable is specified by
rename_duplicate_in_col.}

\item{pattern}{character string giving a regular expression that identifies
the duplicate. Vector of length 1. In pub sec this variable is specified by
rename_duplicate_pattern.}

\item{index}{integer. Which to rename e.g. '2' would rename the 2nd. Can be a
vector of length > 1, but if so the number of items must match that in
prefix. In pub sec this variable is specified with rename_duplicate_index.}

\item{prefix}{character string. The prefix that is to be pasted onto the
original name. Can be a vector of length > 1, but if so the number of
items must match that in index. In pub sec this variable is specified by
rename_duplicate_prefix.}

\item{expected_freq}{integer. The expected number of repeats of the
column name in the raw data. Vector of length 1. In pub sec this variable is
specified by rename_duplicate_expected_freq.}
}
\value{
dat dataframe with an additional column 'rename_note' and with the
identified duplicate renamed. A warning is raised if expected freq is not
equal to the number of repeats found.
}
\description{
If there are values that cannot be uniquely identified by their
descriptors, add a prefix to one of character strings in the specified
column.
}
\details{
Any number of repeats can be renamed e.g. if the data has two descriptors
called 'total', the user may choose to rename only one
of them (and can choose which using 'index'). Alternatively,
they could choose to rename both. See params.

Only one \emph{set} of repeats can be handled: If there are two lots of
duplicated names in the data e.g. two columns called 'total' and 3 called
'subtotal', the function currently only allows one of those to be handled -
the column name of interest is identified by pattern.

To get around this you could combine the columns into a single column.

Note: The duplication in the raw data may be repeated column names in the
same row as each other. However, by this stage the data will have been
unpivotted, so they will appear as duplicated entries in a column.
}
\examples{
\dontrun{
dat <- data.frame(
  "item_detail" = rep(c("A", "A", "Total", "C", "Total", "Total"), 2),
  "year" = rep(c("2021-22", "2022-23"), each = 6),
  "numeric" = rep(c(2, 3, 5, 0.1, 5.1, 10.1), 2)
)
cols_to_group_by <- c("year", "item_detail", rep(NA, 3))
column <- "item_detail"
pattern <- "(?i)total"
index <- c(1, 3)
prefix <- c("first", "third")
expected_freq <- 3

result <- rename_duplicates(
    dat, cols_to_group_by, column, pattern, index, prefix, expected_freq)
print(result)
}
}
