% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_reserved_colnames.R
\name{rename_reserved_colnames}
\alias{rename_reserved_colnames}
\title{rename_reserved_colnames}
\usage{
rename_reserved_colnames(dat, reserved_names)
}
\arguments{
\item{dat}{dataframe}

\item{reserved_colnames}{vector of column names (strings) that we don't want 
to exist yet in dat (they are reserved for some other future column).}
}
\value{
dat with offending columns renamed
}
\description{
If a column in the data has a name reserved for something else, rename it.
}
\details{
If a column in the data has a name reserved for something else, rename it.

Column names that match a string listed in reserved_names are given a 
postscript.
}
\examples{
dat <- data.frame("group" = c("Fox", "Bear"),
                  "value" = c("red", "black"),
                  "numeric" = 1:2)
newdat <- rename_reserved_colnames(dat, c("value", "source"))
}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
