% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_consecutives.R
\name{split_by_consecutives}
\alias{split_by_consecutives}
\title{split_by_consecutives}
\usage{
split_by_consecutives(x)
}
\arguments{
\item{x}{vector of integers}
}
\value{
integer
}
\description{
split a numeric vector into sets of consecutive numbers
}
\details{
split a numeric vector into sets of consecutive numbers
}
\examples{
output <- split_by_consecutives(x)
}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
