% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_to_dict.R
\name{string_to_dict}
\alias{string_to_dict}
\title{string_to_dict}
\usage{
string_to_dict(input_string)
}
\arguments{
\item{input_string}{A character string containing key-value pairs separated 
by commas. See example.}
}
\value{
A data frame representing a dictionary with keys corresponding 
to a list of items. If the input string is empty, an empty data frame is returned.If the 
input string contains any empty keys, they are replaced with NA values.
}
\description{
Generate a dictionary-style dataframe from an input string
}
\details{
Generate a dictionary-style dataframe from an input string

The form of a Python dict brought into R as a string is 
"{key: item, key: item}". 

In R, to access items using their key we convert this string into a dataframe 
where the first column contains keys, and the second column contains item 
lists. 
 
This is not written for use with nested dictionaries.
}
\examples{
# simple
input_string <- "{key1: value1, key2: value2, key3: value3}"
result_dict <- string_to_dict(input_string)

# nested
input_string <- "{key1: [value 1, value 1b], key2: [value2, value2b], key3: value3}"
result_dict <- string_to_dict(input_string)

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}
}
