% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sheet.R
\name{tidy_sheet}
\alias{tidy_sheet}
\title{Runner for public sector preprocessing}
\usage{
tidy_sheet(arg_values, to_csv = TRUE)
}
\arguments{
\item{arg_values}{character string vector. The values for the following
(order is important):
"--args" (optional), input filepath, regular expression matching the
sheet name, output filepath, settings (see example), file part (integer,
nearly always 1).}

\item{to_csv}{boolean. Default is TRUE. If FALSE the output is returned
rather than being saved to csv.}
}
\value{
saved csv file. File is saved to the specified output_filepath,
and contains tidy data. Data has only one numeric value per row, with all
other columns holding descriptive data for that value. Metadata are added to
the output as columns including year, year_type, supplier, source, dataset,
title, units, and vintage.
}
\description{
Runner script for pre-processing xlsx data for public sector
input data at the ONS.

settings is a flattened dictionary (a json.dumps string from Python) that
contains only the relevant variables (settings) for the specific
dataset: The key is the name of the variable, and the value is the
value. Where a setting does not have a value in the sheet_structure
it is assigned NA.

filepath must only contain a single year. Where the data are for
a single year/financial year, the year in filepath should be the year the
data refer to (i.e. not the publication date). The year given in the filepath
must be the correct format (financial or calendar year).
}
\examples{
\dontrun{
# If picking up args from the command line:
arg_values <- commandArgs(trailingOnly = TRUE)

# Without using the command line, this would look like e.g.:
arg_values <- c(
    "D:/some_file-2023_24.xlsx", "(?i)sheet1", "D:/some_file-2023_24.csv",
    "{header_identifier: (?i)proportion, columns_to_create: desc1, desc2}",
    "1"
    )
serializeJSON(arg_values)
tidy_sheet(arg_values, FALSE)
}
}
