% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_separators.R
\name{trim_separators}
\alias{trim_separators}
\title{trim_separators}
\usage{
trim_separators(string, separators)
}
\arguments{
\item{string}{character string.}
}
\value{
vector of strings. Original vector with trailing separators removed
}
\description{
Remove hyphens, colons etc, and spaces from the start and end of strings
}
\details{
Remove hyphens, colons etc, and spaces from the start and end of strings
}
\examples{
example_vector <- c("these - ", "hyphens-", " and colons :", ":now   gone- ")
trim_separators(example_vector, "-|:")

}
\author{
Emma Wood \email{emma.wood@ons.gov.uk}

Mark London \email{mark.london@ext.ons.gov.uk}
}
