% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot_data.R
\name{unpivot_data}
\alias{unpivot_data}
\title{take non-tidy data out of xlsx_cells layout}
\usage{
unpivot_data(
  dat,
  columns_to_create,
  first_header_row,
  tolerance = 0.4,
  left_headers,
  minimum_number_of_consecutive_columns,
  right_block_offset,
  header_to_split,
  header_split_to,
  split_points,
  column_to_right_of_data_name_pattern,
  tidy_data,
  tidy_notes_name
)
}
\arguments{
\item{dat}{Dataframe created using tidyxl::xlsx_cells().}

\item{columns_to_create}{character vector. The names of columns that will
be created during behead. The top header in dat will be pivoted to a
column that is given the first name in columns_to_create. If there is a
second row of headers, they will be pivoted to a column that is given the
second name in columns_to_create etc.}

\item{first_header_row}{integer. the number of the first row in which there
are headers.}

\item{tolerance}{numeric between 0 and 1. Defaults to 0.4. If tolerance is
0.4, the column will be considered 'numeric' if 40\% or more of the cells
in that row are numeric. 0.4 works for the vast majority of datasets.
However, if, for example, the data has many header rows but very few numeric
data rows, or has many suppressed values in the first couple of value columns
you may need to decrease the tolerance. If the last descriptive column in the
left block contains a lot of cells with numbers, you may need to set the
tolerance to higher than 0.4.}

\item{left_headers}{character vector or NA. Optional. The names to be given
to the columns containing descriptor data (to the left of the data). If
a column in the left block has previously been removed by
columns_to_remove_patterns, it must not be included in left_headers.}

\item{minimum_number_of_consecutive_columns}{integer. Defaults to 2. If more
than 2 consecutive numeric columns are required to identify the start of the
right block of numeric data, specify the number required. i.e. if only
columns 3 and 4 contain numbers, this counts as 2 consecutive columns and
minimum_number_of_consecutive_columns would not need to be specified. If,
however, columns 3 and 4 contain numbers but are in the left block of
descriptors, and columns 6, 7, and 8 contain numbers and are the first
columns of the right block, minimum_number_of_consecutive_columns would be 3.}

\item{right_block_offset}{integer. The value to add to the identified column
number. If the last of the descriptive columns appear(s) numeric use a
positive offset. If the first of the numeric right block columns contains a
very low proportion of numbers and for some reason you can't change the
tolerance, you could use a negative offset value.}

\item{header_to_split}{character string or NA. Optional. The name of the
column whose information is to be split.}

\item{header_split_to}{vector of character strings or NA. The new names of
the columns that the split data will be saved to. There must be one name for
each chunk of the split string, so the length of header_split_to will be one
greater than the length of split_points.}

\item{split_points}{vector of character strings that code for regular
expressions or NA. See the Wiki or build_regex() for more information on how
these codes are written. Each one will only be matched once, so any repeats
must be explicit.}

\item{column_to_right_of_data_name_pattern}{Character string or NA. Regular
expression that matches the name of a descriptor column to the right of the
numeric data in the excel file. Only one can be given. If more are needed the
function will need to be updated.}

\item{tidy_data}{boolean. Defaults to FALSE. Use TRUE when data only have one
column of numeric values.}

\item{tidy_notes_name}{character string. Only used if tidy_data is true.
Only required if there is a character column to the right of tidy data,
in which case it is used as a name for that column.}
}
\description{
Convert untidy Excel data to tidy data with one numeric value
per row and any number of descriptor columns. dat is tidy because it is
xlsx_cells data but it has one row of data for every cell in a sheet, so is
not useful for subsequent purposes. This function is largely based on
unpivotr.
}
